Gather MCMC chain draws to data.frame on a long format

Useful to convert MCMC chain draws of particular parameters or output from the model object to a long format for further data wrangling

gather_draws(mod, ..., transformed = F)

Arguments

mod
an object of class "plm0","plm","gplm0" or "gplm".

...
any number of character vectors containing valid names of parameters in the model or "rating_curve" and "rating_curve_mean". Also accepts "latent_parameters" and "hyperparameters".

transformed
boolean value determining whether the parameter is to be represented on the transformed scale used for sampling in the MCMC chain or the original scale. Defaults to FALSE.

Value

Data frame with columns chain iter param value

References

Birgir Hrafnkelsson, Helgi Sigurdarson, & Sigurdur M. Gardarsson. (2020). Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling.

See also

Examples

if (FALSE) {
data(V316_river)
f <- Q~W
plm0.fit <- plm0(f,V316_river)
hyp_samples <- spread_draws(plm0.fit,'hyperparameters')
head(hyp_samples)
rating_curve_samples <- spread_draws(plm0.fit,'rating_curve','rating_curve_mean')
head(rating_curve_samples)
}