Report for a discharge rating curve or tournament

Save a pdf file with a report of a discharge rating curve object or tournament.

get_report(x, path = NULL, type = 1, ...)

# S3 method for plm0
get_report(x, path = NULL, type = 1, ...)

# S3 method for plm
get_report(x, path = NULL, type = 1, ...)

# S3 method for gplm0
get_report(x, path = NULL, type = 1, ...)

# S3 method for gplm
get_report(x, path = NULL, type = 1, ...)

# S3 method for tournament
get_report(x, path = NULL, type = 1, ...)

Arguments

x

an object of class "tournament", "plm0", "plm", "gplm0" or "gplm".

path

file path to which the pdf file of the report is saved. If NULL, the current working directory is used.

type

an integer denoting what type of report is to be produced. Defaults to type 1. Only type 1 is permissible for an object of class "plm0", "plm", "gplm0" or "gplm". Possible types are

1 - produces a report displaying the results of the model (winning model if a tournament provided). The first page contains a panel of four plots and a summary of the posterior distributions of the parameters. On the second page a tabular prediction of discharge on an equally spaced grid of stages is displayed. This prediction table can span multiple pages.

2 - produces a ten page report and is only permissible for objects of class "tournament". The first four pages contain a panel of four plots and a summary of the posterior distributions of the parameters for each of the four models in the tournament, the fifth page shows model comparison plots and tables, the sixth page convergence diagnostics plots, and the final four pages shows the histograms of the parameters in each of the four models.

...

further arguments passed to other methods (currently unused).

Value

No return value, called for side effects

Details

This function can only be used in an interactive R session as it asks permission from the user to write to their file system.

Methods (by class)

plm0: Get report for plm0 model object

plm: Get report for plm model object

gplm0: Get report for gplm0 model object

gplm: Get report for gplm

tournament: Get report for discharge rating curve tournament

See also

Examples

# \donttest{
data(krokfors)
set.seed(1)
plm0.fit <- plm0(formula=Q~W,data=krokfors,num_cores=2)
# }
if (FALSE) {
get_report(plm0.fit)
}