Report pages for a discharge rating curve or tournament

Get a list the pages of a report of a discharge rating curve object or tournament

get_report_pages(x, type = 1, ...)

# S3 method for plm0
get_report_pages(x, type = 1, ...)

# S3 method for plm
get_report_pages(x, type = 1, ...)

# S3 method for gplm0
get_report_pages(x, type = 1, ...)

# S3 method for gplm
get_report_pages(x, type = 1, ...)

# S3 method for tournament
get_report_pages(x, type = 1, ...)

Arguments

x
an object of class "tournament", "plm0", "plm", "gplm0" or "gplm".

type
an integer denoting what type of report is to be produced. Defaults to type 1. Possible types are

1 - produces a two page report displaying the results of the model (winning model if a tournament provided). The first page contains a panel of four plots and a summary of the posterior distributions of the parameters. The second one contains a tabular prediction of discharge on an equally spaced grid of stages.

2 - produces a ten page report and is only permissible for objects of class "tournament". The first four pages contain a panel of four plots and a summary of the posterior distributions of the parameters for each of the four models in the tournament, the fifth page shows model comparison plots and tables, the sixth page convergence diagnostics plots, and the final four pages shows the histograms of the parameters in each of the four models.

...
further arguments passed to other methods (currently unused).

Methods (by class)

plm0: Get report pages for plm0 result object

plm: Get report pages for plm result object

gplm0: Get report pages for gplm0 result object

gplm: Get report pages for gplm result object

tournament: Get report pages for discharge rating curve tournament result object

See also

Examples

# \donttest{
data(skogsliden)
plm0.fit <- plm0(Q~W,skogsliden)
plm0_pages <- get_report_pages(plm0.fit)
t_obj <- tournament(Q~W,skogsliden)

#> Running tournament:
#> 25% - gplm finished
#> 50% - gplm0 finished
#> 75% - plm finished
#> 100% - plm0 finished

tournament_pages <- get_report_pages(t_obj,type=2)
# }