Generalized Power Law Model with Constant Variance

gplm0 is used to fit a discharge rating curve for paired measurements of stage and discharge using a Generalized Power Law Model with constant variance as described in Hrafnkelsson et al. See "Details" for a more elaborate description of the model.

gplm0(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  parallel = T,
  forcepoint = rep(FALSE, nrow(data))
)

Arguments

formula
an object of class "formula", with discharge column name as response and stage column name as a covariate, i.e. of the form y~x where y is discharge in \(m^3/s\) and x is stage in \(m\) (it is very important that the data is in the correct units).

data
data.frame containing the variables specified in formula.

c_param
stage for which there is zero discharge. If NULL, it is treated as unknown in the model and inferred from the data.

h_max
maximum stage to which the rating curve should extrapolate to. If NULL, the maximum stage value in data is selected as an upper bound.

parallel
logical value indicating whether to run the MCMC in parallel or not. Defaults to TRUE.

forcepoint
logical vector of the same length as the number of rows in data. If an element at index i is TRUE it indicates that the rating curve should be forced through the i-th measurement. Use with care, as this will strongly influence the resulting rating curve.

Value

gplm0 returns an object of class "gplm0". An object of class "gplm0" is a list containing the following components:

rating_curve

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of the rating curve.

rating_curve_mean

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of the mean of the rating curve.

param_summary

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of latent- and hyperparameters.

beta_summary

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of \(\beta\).

Deviance_summary

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of the deviance.

rating_curve_posterior

a matrix containing the full thinned posterior samples of the posterior distribution of the rating curve (excluding burn-in).

rating_curve_mean_posterior

a matrix containing the full thinned posterior samples of the posterior distribution of the mean of the rating curve (excluding burn-in).

a_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(a\).

b_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(b\).

c_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(c\).

sigma_eps_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\sigma_{\epsilon}\).

sigma_beta_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\sigma_{\beta}\).

phi_beta_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\phi_{\beta}\).

sigma_eta_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\sigma_{\eta}\).

beta_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\beta\).

Deviance_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of the deviance excluding burnin samples.

D_hat

deviance at the median value of the parameters

num_effective_param

number of effective parameters, which is calculated as median(Deviance_posterior) - D_hat

DIC

Deviance Information Criterion for the model, calculated as D_hat + 2*num_effective_parameters

acceptance_rate

proportion of accepted samples in the thinned MCMC chain (excluding burn-in).

formula

object of type "formula" provided by the user.

data

data provided by the user.

run_info

Information about the specific parameters used in the MCMC chain.

Details

The generalized power-law model is of the form $$Q=a(h-c)^{f(h)}$$ where \(Q\) is discharge, \(h\) is stage, \(a\) and \(c\) are unknown constants and \(f\) is a function of \(h\) referred to as the generalized power-law exponent.

The Bayesian generalized power-law model is presented as a Bayesian hierarchical model. The function \(f\) is modelled at the latent level as a fixed constant b plus a continuous stochastic process which is assumed to be twice differentiable. The model is on a logarithmic scale $$log(Q_i) = log(a) + (b + \beta(h_i)) log(h_i - c) + \epsilon, i = 1,...,n$$ where \(\epsilon\) follows a normal distribution with mean zero and variance \(\sigma_\epsilon^2\), independent of stage. The stochastic process \(\beta(h)\) is assumed a priori to be a Gaussian process governed by a Matern covariance function with smoothness parameter \(\nu = 2.5\). An efficient posterior simulation is achieved by sampling from the joint posterior density of the hyperparameters of the model, and then sampling from the density of the latent parameters conditional on the hyperparameters.

Bayesian inference is based on the posterior density and summary statistics such as the posterior mean and 95% posterior intervals are based on the posterior density. Analytical formulas for these summary statistics are intractable in most cases and thus they are computed by generating samples from the posterior density using a Markov chain Monte Carlo simulation.

References

B. Hrafnkelsson, H. Sigurdarson, S.M. Gardarsson, 2020, Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv preprint 2010.04769

See also

Examples

if (FALSE) {
data(V316_river)
set.seed(1)
formula <- Q~W
gplm0.fit <- gplm0(formula,V316_river)
summary(gplm0.fit)
plot(gplm0.fit)
gplm0.fit_known_c <- gplm0(formula,V316_river,c_param=0.75,h_max=2)
summary(gplm0.fit_known_c)
plot(gplm0.fit_known_c)
}