Power Law Model

plm is used to fit a discharge rating curve for paired measurements of stage and discharge using a Power Law Model with variance that may vary with stage as described in Hrafnkelsson et al. See "Details" for a more elaborate description of the model.

plm(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  parallel = T,
  forcepoint = rep(FALSE, nrow(data))
)

Arguments

formula
an object of class "formula", with discharge column name as response and stage column name as a covariate, i.e. of the form y~x where y is discharge in \(m^3/s\) and x is stage in \(m\) (it is very important that the data is in the correct units).

data
data.frame containing the variables specified in formula.

c_param
stage for which there is zero discharge. If NULL, it is treated as unknown in the model and inferred from the data.

h_max
maximum stage to which the rating curve should extrapolate to. If NULL, the maximum stage value in data is selected as an upper bound.

parallel
logical value indicating whether to run the MCMC in parallel or not. Defaults to TRUE.

forcepoint
logical vector of the same length as the number of rows in data. If an element at index i is TRUE it indicates that the rating curve should be forced through the i-th measurement. Use with care, as this will strongly influence the resulting rating curve.

Value

plm returns an object of class "plm". An object of class "plm" is a list containing the following components:

rating_curve

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of the rating curve.

rating_curve_mean

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of the mean of the rating curve.

param_summary

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of latent- and hyperparameters.

sigma_eps_summary

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior of \(\sigma_{\epsilon}\)

Deviance_summary

a data frame with 2.5%, 50% and 97.5% quantiles of the posterior distribution of the deviance.

rating_curve_posterior

a matrix containing the full thinned posterior samples of the posterior distribution of the rating curve (excluding burn-in).

rating_curve_mean_posterior

a matrix containing the full thinned posterior samples of the posterior distribution of the mean of the rating curve (excluding burn-in).

a_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(a\).

b_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(b\).

c_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(c\).

sigma_eps_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\sigma_{\epsilon}\).

eta_1_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\eta_1\).

eta_2_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\eta_2\).

eta_3_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\eta_3\).

eta_4_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\eta_4\).

eta_5_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\eta_5\).

eta_6_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\eta_6\).

Deviance_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of the deviance excluding burnin samples.

D_hat

deviance at the median value of the parameters

num_effective_param

number of effective parameters, which is calculated as median(Deviance_posterior) - D_hat

DIC

Deviance Information Criterion for the model, calculated as D_hat + 2*num_effective_parameters

acceptance_rate

proportion of accepted samples in the thinned MCMC chain (excluding burn-in).

formula

object of type "formula" provided by the user.

data

data provided by the user.

run_info

Information about the specific parameters used in the MCMC chain.

Details

The power-law model, which is commonly used in hydraulic practice, is of the form $$Q=a(h-c)^{b}$$ where \(Q\) is discharge, \(h\) is stage and \(a\), \(b\) and \(c\) are unknown constants. It is presented here as a Bayesian hierarchical model. The model is on a logarithmic scale $$log(Q_i) = log(a) + b log(h_i - c) + \epsilon_i, i = 1,...,n$$ where \(\epsilon_i\) follows a normal distribution with mean zero and variance \(\sigma_\epsilon(h_i)^2\) that can vary with stage. The error variance, \(\sigma_\epsilon^2(h)\), of the log-discharge data is modeled as an exponential of a B-spline curve, that is, a linear combination of six B-spline basis functions that are defined over the range of the stage observations. An efficient posterior simulation is achieved by sampling from the joint posterior density of the hyperparameters of the model, and then sampling from the density of the latent parameters conditional on the hyperparameters.

Bayesian inference is based on the posterior density and summary statistics such as the posterior mean and 95% posterior intervals are based on the posterior density. Analytical formulas for these summary statistics are intractable in most cases and thus they are computed by generating samples from the posterior density using a Markov chain Monte Carlo simulation.

References

B. Hrafnkelsson, H. Sigurdarson, S.M. Gardarsson, 2020, Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv preprint 2010.04769

See also

Examples

if (FALSE) {
data(V316_river)
set.seed(1)
formula <- Q~W
plm.fit <- plm(formula,V316_river)
summary(plm.fit)
plot(plm.fit)
plm.fit_known_c <- plm(formula,V316_river,c_param=0.75,h_max=2)
summary(plm.fit_known_c)
plot(plm.fit_known_c)
}