Power-law model with constant variance

plm0 is used to fit a discharge rating curve for paired measurements of stage and discharge using a power-law model with a constant variance as described in Hrafnkelsson et al. (2020). See "Details" for a more elaborate description of the model.

plm0(
  formula,
  data,
  c_param = NULL,
  h_max = NULL,
  parallel = T,
  forcepoint = rep(FALSE, nrow(data))
)

Arguments

formula
an object of class "formula", with discharge column name as response and stage column name as a covariate, i.e. of the form y~x where y is discharge in m\(^3/\)s and x is stage in m (it is very important that the data is in the correct units).

data
data.frame containing the variables specified in formula.

c_param
stage for which there is zero discharge. If NULL, it is treated as unknown in the model and inferred from the data.

h_max
maximum stage to which the rating curve should extrapolate to. If NULL, the maximum stage value in the data is selected as an upper bound.

parallel
logical value indicating whether to run the MCMC in parallel or not. Defaults to TRUE.

forcepoint
logical vector of the same length as the number of rows in data. If an element at index \(i\) is TRUE it indicates that the rating curve should be forced through the \(i\)-th measurement. Use with care, as this will strongly influence the resulting rating curve.

Value

plm0 returns an object of class "plm0". An object of class "plm0" is a list containing the following components:

rating_curve

a data frame with 2.5%, 50% and 97.5% percentiles of the posterior predictive distribution of the rating curve.

rating_curve_mean

a data frame with 2.5%, 50% and 97.5% percentiles of the posterior distribution of the mean of the rating curve.

param_summary

a data frame with 2.5%, 50% and 97.5% percentiles of the posterior distribution of latent- and hyperparameters. Additionally contains columns with r_hat and the effective number of samples for each parameter as defined in Gelman et al. (2013).

Deviance_summary

a data frame with 2.5%, 50% and 97.5% percentiles of the posterior distribution of the deviance.

rating_curve_posterior

a matrix containing the full thinned posterior samples of the posterior predictive distribution of the rating curve (excluding burn-in).

rating_curve_mean_posterior

a matrix containing the full thinned posterior samples of the posterior distribution of the mean of the rating curve (excluding burn-in).

a_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(a\).

b_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(b\).

c_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(c\).

sigma_eps_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of \(\sigma_{\varepsilon}\).

Deviance_posterior

a numeric vector containing the full thinned posterior samples of the posterior distribution of the deviance excluding burnin samples.

D_hat

deviance at the median value of the parameters

num_effective_param

number of effective parameters, which is calculated as median(Deviance_posterior) minus D_hat.

DIC

Deviance Information Criterion for the model, calculated as D_hat plus 2*num_effective_parameters.

autocorrelation

a data frame with the autocorrelation of each parameter for different lags.

acceptance_rate

proportion of accepted samples in the thinned MCMC chain (excluding burn-in).

formula

object of type "formula" provided by the user.

data

data provided by the user, ordered by stage.

run_info

information about the input arguments and the specific parameters used in the MCMC chain.

Details

The power-law model, which is commonly used in hydraulic practice, is of the form $$Q=a(h-c)^{b}$$ where \(Q\) is discharge, \(h\) is stage and \(a\), \(b\) and \(c\) are unknown constants.

The power-law model is here inferred by using a Bayesian hierarchical model. The model is on a logarithmic scale $$\log(Q_i) = \log(a) + b \log(h_i - c) + \varepsilon, i = 1,...,n$$ where \(\varepsilon\) follows a normal distribution with mean zero and variance \(\sigma_\varepsilon^2\), independent of stage. An efficient posterior simulation is achieved by sampling from the joint posterior density of the hyperparameters of the model, and then sampling from the density of the latent parameters conditional on the hyperparameters.

Bayesian inference is based on the posterior density and summary statistics such as the posterior mean and 95% posterior intervals are based on the posterior density. Analytical formulas for these summary statistics are intractable in most cases and thus they are computed by generating samples from the posterior density using a Markov chain Monte Carlo simulation.

References

Hrafnkelsson, B., Sigurdarson, H., and Gardarsson, S. M. (2020). Generalization of the power-law rating curve using hydrodynamic theory and Bayesian hierarchical modeling. arXiv preprint 2010.04769.

Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., and Rubin, D. B. (2013). Bayesian Data Analysis, Third Edition. Chapman & Hall/CRC Texts in Statistical Science. Taylor & Francis.

See also

Examples

# \donttest{
data(krokfors)
set.seed(1)
formula <- Q~W
plm0.fit <- plm0(formula,krokfors)
summary(plm0.fit)

#> 
#> Formula: 
#>  Q ~ W
#> Latent parameters:
#>   lower-2.5% median-50% upper-97.5%
#> a 1.16       1.64       2.12       
#> b 2.47       2.82       3.21       
#> 
#> Hyperparameters:
#>           lower-2.5% median-50% upper-97.5%
#> c         7.58       7.676      7.750      
#> sigma_eps 0.15       0.191      0.251      
#> 
#> DIC: 23.37559

# }