Predict method for discharge rating curves

Predict the discharge for given stage values based on a discharge rating curve result object.

# S3 method for plm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for plm
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm
predict(object, newdata = NULL, wide = FALSE, ...)

Arguments

object
an object of class "plm0", "plm", "gplm0" or "gplm".

newdata
a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.

wide
a logical value denoting whether to produce a wide prediction output.If TRUE, then the output is a table with median prediction values for an equally spaced grid of stages with 1 cm increments, each row containing predictions in a decimeter range of stages.

...
not used in this function

Value

an object of class "data.frame" with four columns, h (stage), lower (2.5% posterior predictive quantile), median (50% posterior predictive quantile), upper (97.5% posterior predictive quantile). If wide=T, a matrix as described above (see wide parameter) is returned.

Functions

predict.plm0: Predict method for plm0

predict.plm: Predict method for plm

predict.gplm0: Predict method for gplm0

predict.gplm: Predict method for gplm

See also

Examples

# \donttest{
data(krokfors)
plm0.fit <- plm0(Q~W,krokfors,h_max=24)
#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(plm0.fit,newdata=seq(23,24,by=0.01))

#>         h    lower   median     upper
#> 1   23.00 1538.124 3635.183  8693.321
#> 2   23.01 1541.022 3640.602  8707.225
#> 3   23.02 1540.484 3646.480  8710.466
#> 4   23.03 1539.945 3652.359  8713.707
#> 5   23.04 1539.407 3658.237  8716.948
#> 6   23.05 1538.869 3664.115  8720.190
#> 7   23.06 1542.275 3670.172  8723.906
#> 8   23.07 1552.973 3676.556  8728.499
#> 9   23.08 1563.671 3682.941  8733.093
#> 10  23.09 1574.369 3689.325  8737.686
#> 11  23.10 1585.067 3695.710  8742.280
#> 12  23.11 1590.666 3701.687  8759.817
#> 13  23.12 1588.125 3707.012  8798.012
#> 14  23.13 1585.584 3712.338  8836.206
#> 15  23.14 1583.043 3717.664  8874.401
#> 16  23.15 1580.502 3722.989  8912.596
#> 17  23.16 1578.777 3729.866  8942.191
#> 18  23.17 1578.183 3738.889  8959.881
#> 19  23.18 1577.589 3747.913  8977.571
#> 20  23.19 1576.995 3756.937  8995.260
#> 21  23.20 1576.400 3765.961  9012.950
#> 22  23.21 1578.968 3773.868  9026.865
#> 23  23.22 1585.346 3780.427  9036.230
#> 24  23.23 1591.724 3786.987  9045.596
#> 25  23.24 1598.102 3793.547  9054.961
#> 26  23.25 1604.480 3800.107  9064.327
#> 27  23.26 1610.158 3806.965  9070.407
#> 28  23.27 1615.102 3814.136  9073.037
#> 29  23.28 1620.046 3821.307  9075.666
#> 30  23.29 1624.990 3828.478  9078.295
#> 31  23.30 1629.934 3835.649  9080.925
#> 32  23.31 1631.347 3842.288  9101.866
#> 33  23.32 1629.526 3848.441  9139.586
#> 34  23.33 1627.704 3854.593  9177.306
#> 35  23.34 1625.883 3860.746  9215.026
#> 36  23.35 1624.062 3866.898  9252.746
#> 37  23.36 1624.705 3873.975  9285.556
#> 38  23.37 1627.318 3881.790  9314.445
#> 39  23.38 1629.930 3889.605  9343.333
#> 40  23.39 1632.542 3897.420  9372.222
#> 41  23.40 1635.154 3905.235  9401.110
#> 42  23.41 1639.791 3912.918  9418.426
#> 43  23.42 1645.833 3920.511  9427.703
#> 44  23.43 1651.875 3928.103  9436.981
#> 45  23.44 1657.918 3935.696  9446.259
#> 46  23.45 1663.960 3943.288  9455.536
#> 47  23.46 1666.707 3949.067  9466.248
#> 48  23.47 1667.471 3953.755  9477.824
#> 49  23.48 1668.235 3958.442  9489.399
#> 50  23.49 1668.998 3963.129  9500.975
#> 51  23.50 1669.762 3967.817  9512.550
#> 52  23.51 1670.707 3972.968  9545.060
#> 53  23.52 1671.745 3978.361  9588.428
#> 54  23.53 1672.784 3983.753  9631.796
#> 55  23.54 1673.822 3989.145  9675.164
#> 56  23.55 1674.861 3994.537  9718.532
#> 57  23.56 1677.555 4005.073  9729.779
#> 58  23.57 1680.983 4017.892  9726.770
#> 59  23.58 1684.412 4030.711  9723.761
#> 60  23.59 1687.841 4043.530  9720.752
#> 61  23.60 1691.269 4056.349  9717.742
#> 62  23.61 1693.139 4064.138  9724.458
#> 63  23.62 1694.423 4070.036  9734.830
#> 64  23.63 1695.706 4075.934  9745.202
#> 65  23.64 1696.990 4081.832  9755.574
#> 66  23.65 1698.274 4087.729  9765.945
#> 67  23.66 1702.796 4096.635  9762.778
#> 68  23.67 1708.335 4106.486  9755.358
#> 69  23.68 1713.875 4116.337  9747.937
#> 70  23.69 1719.415 4126.188  9740.516
#> 71  23.70 1724.954 4136.040  9733.095
#> 72  23.71 1726.689 4141.671  9784.510
#> 73  23.72 1727.443 4146.216  9851.091
#> 74  23.73 1728.197 4150.760  9917.671
#> 75  23.74 1728.951 4155.304  9984.251
#> 76  23.75 1729.705 4159.848 10050.831
#> 77  23.76 1732.741 4165.626 10064.892
#> 78  23.77 1736.247 4171.658 10068.136
#> 79  23.78 1739.753 4177.689 10071.381
#> 80  23.79 1743.259 4183.721 10074.625
#> 81  23.80 1746.765 4189.752 10077.870
#> 82  23.81 1749.894 4197.940 10106.830
#> 83  23.82 1752.964 4206.470 10139.859
#> 84  23.83 1756.034 4214.999 10172.888
#> 85  23.84 1759.104 4223.529 10205.917
#> 86  23.85 1762.174 4232.058 10238.947
#> 87  23.86 1767.651 4236.261 10246.270
#> 88  23.87 1773.404 4239.971 10250.658
#> 89  23.88 1779.157 4243.680 10255.047
#> 90  23.89 1784.910 4247.389 10259.435
#> 91  23.90 1790.662 4251.099 10263.824
#> 92  23.91 1794.929 4258.404 10288.675
#> 93  23.92 1799.086 4265.972 10315.026
#> 94  23.93 1803.244 4273.540 10341.377
#> 95  23.94 1807.401 4281.108 10367.729
#> 96  23.95 1811.559 4288.677 10394.080
#> 97  23.96 1812.256 4301.891 10407.219
#> 98  23.97 1812.831 4315.305 10419.890
#> 99  23.98 1813.405 4328.720 10432.561
#> 100 23.99 1813.980 4342.134 10445.232
#> 101 24.00 1814.555 4355.548 10457.903

predict(plm0.fit,wide=T)

#>                 0         0.01         0.02         0.03         0.04
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> 7.7  1.651196e-04 2.281684e-04 2.912172e-04 5.643955e-04 9.669563e-04
#> 7.8  5.022077e-03 6.050758e-03 7.079439e-03 8.649296e-03 1.055237e-02
#> 7.9  2.442587e-02 2.772376e-02 3.102165e-02 3.478978e-02 3.891314e-02
#> 8    6.890823e-02 7.477178e-02 8.177026e-02 8.876873e-02 9.588644e-02
#> 8.1  1.471519e-01 1.574937e-01 1.678355e-01 1.781773e-01 1.892799e-01
#> 8.2  2.672652e-01 2.818868e-01 2.970334e-01 3.124558e-01 3.278782e-01
#> 8.3  4.360746e-01 4.559383e-01 4.762988e-01 4.974047e-01 5.195401e-01
#> 8.4  6.618427e-01 6.880376e-01 7.145368e-01 7.418382e-01 7.691396e-01
#> 8.5  9.508137e-01 9.833022e-01 1.016232e+00 1.052046e+00 1.090246e+00
#> 8.6  1.315287e+00 1.355129e+00 1.394954e+00 1.434779e+00 1.478820e+00
#> 8.7  1.751804e+00 1.801105e+00 1.850407e+00 1.899708e+00 1.950779e+00
#> 8.8  2.283615e+00 2.341953e+00 2.400292e+00 2.459160e+00 2.521629e+00
#> 8.9  2.907527e+00 2.973278e+00 3.038738e+00 3.104197e+00 3.177482e+00
#> 9    3.620612e+00 3.692189e+00 3.771562e+00 3.850936e+00 3.930984e+00
#> 9.1  4.445128e+00 4.527466e+00 4.613054e+00 4.719970e+00 4.824800e+00
#> 9.2  5.371267e+00 5.480912e+00 5.589643e+00 5.690150e+00 5.791084e+00
#> 9.3  6.471601e+00 6.578218e+00 6.679123e+00 6.780028e+00 6.892817e+00
#> 9.4  7.621235e+00 7.741588e+00 7.861940e+00 7.982293e+00 8.117156e+00
#> 9.5  8.948855e+00 9.082820e+00 9.218355e+00 9.355685e+00 9.497180e+00
#> 9.6  1.037579e+01 1.052513e+01 1.068325e+01 1.084908e+01 1.101490e+01
#> 9.7  1.199741e+01 1.216801e+01 1.233861e+01 1.250920e+01 1.267832e+01
#> 9.8  1.372588e+01 1.391158e+01 1.409777e+01 1.428483e+01 1.447190e+01
#> 9.9  1.556558e+01 1.578257e+01 1.599955e+01 1.621654e+01 1.643352e+01
#> 10   1.768888e+01 1.791615e+01 1.814341e+01 1.837068e+01 1.859794e+01
#> 10.1 1.991862e+01 2.016034e+01 2.040206e+01 2.064377e+01 2.088549e+01
#> 10.2 2.236417e+01 2.262698e+01 2.288979e+01 2.315260e+01 2.341542e+01
#> 10.3 2.499534e+01 2.524656e+01 2.549778e+01 2.574900e+01 2.600022e+01
#> 10.4 2.779408e+01 2.808345e+01 2.837283e+01 2.866220e+01 2.895158e+01
#> 10.5 3.067105e+01 3.097005e+01 3.126904e+01 3.156803e+01 3.186702e+01
#> 10.6 3.398801e+01 3.428592e+01 3.458382e+01 3.488172e+01 3.517962e+01
#> 10.7 3.708086e+01 3.750573e+01 3.793059e+01 3.835545e+01 3.878031e+01
#> 10.8 4.094423e+01 4.137630e+01 4.180836e+01 4.224043e+01 4.267250e+01
#> 10.9 4.462020e+01 4.499625e+01 4.537229e+01 4.574834e+01 4.612583e+01
#> 11   4.878997e+01 4.915978e+01 4.952958e+01 4.989939e+01 5.027584e+01
#> 11.1 5.295614e+01 5.339960e+01 5.384307e+01 5.428654e+01 5.473049e+01
#> 11.2 5.744957e+01 5.792939e+01 5.840921e+01 5.888903e+01 5.935138e+01
#> 11.3 6.194019e+01 6.248596e+01 6.303173e+01 6.357750e+01 6.411862e+01
#> 11.4 6.724642e+01 6.773787e+01 6.822931e+01 6.872076e+01 6.922287e+01
#> 11.5 7.229788e+01 7.278173e+01 7.326559e+01 7.374944e+01 7.426637e+01
#> 11.6 7.765889e+01 7.823717e+01 7.881546e+01 7.939375e+01 7.999459e+01
#> 11.7 8.361275e+01 8.415500e+01 8.469725e+01 8.523950e+01 8.577464e+01
#> 11.8 8.924242e+01 8.995353e+01 9.066464e+01 9.137575e+01 9.201876e+01
#> 11.9 9.568015e+01 9.630825e+01 9.693636e+01 9.756447e+01 9.826692e+01
#> 12   1.024390e+02 1.030387e+02 1.036383e+02 1.042380e+02 1.048864e+02
#> 12.1 1.089678e+02 1.096879e+02 1.104080e+02 1.111282e+02 1.118823e+02
#> 12.2 1.163522e+02 1.170624e+02 1.177725e+02 1.184827e+02 1.191804e+02
#> 12.3 1.238166e+02 1.247491e+02 1.256815e+02 1.266140e+02 1.273211e+02
#> 12.4 1.314842e+02 1.321639e+02 1.328436e+02 1.335316e+02 1.344110e+02
#> 12.5 1.398117e+02 1.407509e+02 1.416902e+02 1.425948e+02 1.432191e+02
#> 12.6 1.475779e+02 1.484881e+02 1.493983e+02 1.502863e+02 1.510723e+02
#> 12.7 1.559608e+02 1.568249e+02 1.576890e+02 1.585548e+02 1.594260e+02
#> 12.8 1.650941e+02 1.661587e+02 1.672232e+02 1.682126e+02 1.690390e+02
#> 12.9 1.741185e+02 1.749965e+02 1.758744e+02 1.767743e+02 1.777093e+02
#> 13   1.836309e+02 1.846946e+02 1.857584e+02 1.868054e+02 1.878321e+02
#> 13.1 1.939943e+02 1.950218e+02 1.960493e+02 1.970333e+02 1.979771e+02
#> 13.2 2.044560e+02 2.057192e+02 2.069825e+02 2.080156e+02 2.088877e+02
#> 13.3 2.147475e+02 2.158587e+02 2.169699e+02 2.181516e+02 2.193701e+02
#> 13.4 2.258456e+02 2.267536e+02 2.276617e+02 2.286481e+02 2.296644e+02
#> 13.5 2.367133e+02 2.380744e+02 2.394355e+02 2.407490e+02 2.420501e+02
#> 13.6 2.485266e+02 2.493573e+02 2.501880e+02 2.514249e+02 2.527271e+02
#> 13.7 2.601221e+02 2.612799e+02 2.624377e+02 2.639146e+02 2.654157e+02
#> 13.8 2.732462e+02 2.743506e+02 2.754550e+02 2.768618e+02 2.782691e+02
#> 13.9 2.861366e+02 2.873538e+02 2.885753e+02 2.898556e+02 2.911359e+02
#> 14   2.981124e+02 2.991652e+02 3.003381e+02 3.022818e+02 3.042255e+02
#> 14.1 3.127307e+02 3.136783e+02 3.147252e+02 3.161621e+02 3.175990e+02
#> 14.2 3.261006e+02 3.275005e+02 3.290062e+02 3.307955e+02 3.325849e+02
#> 14.3 3.418408e+02 3.431824e+02 3.444690e+02 3.456487e+02 3.468284e+02
#> 14.4 3.554946e+02 3.571450e+02 3.587487e+02 3.602846e+02 3.618206e+02
#> 14.5 3.704911e+02 3.718687e+02 3.733986e+02 3.750958e+02 3.767930e+02
#> 14.6 3.865566e+02 3.881343e+02 3.897443e+02 3.913814e+02 3.930184e+02
#> 14.7 4.019734e+02 4.033681e+02 4.048910e+02 4.064949e+02 4.080987e+02
#> 14.8 4.185639e+02 4.203986e+02 4.219068e+02 4.232623e+02 4.246177e+02
#> 14.9 4.356878e+02 4.378338e+02 4.395427e+02 4.411057e+02 4.426687e+02
#> 15   4.535747e+02 4.555415e+02 4.570151e+02 4.583789e+02 4.597426e+02
#> 15.1 4.704056e+02 4.724133e+02 4.743670e+02 4.763138e+02 4.782605e+02
#> 15.2 4.888753e+02 4.905502e+02 4.923171e+02 4.940884e+02 4.958597e+02
#> 15.3 5.069007e+02 5.087715e+02 5.104651e+02 5.121588e+02 5.138524e+02
#> 15.4 5.249074e+02 5.268182e+02 5.287000e+02 5.305818e+02 5.324635e+02
#> 15.5 5.468512e+02 5.493874e+02 5.514161e+02 5.534448e+02 5.554735e+02
#> 15.6 5.650747e+02 5.667587e+02 5.693525e+02 5.719463e+02 5.745401e+02
#> 15.7 5.853105e+02 5.868499e+02 5.885556e+02 5.902613e+02 5.919671e+02
#> 15.8 6.044738e+02 6.068967e+02 6.096550e+02 6.124134e+02 6.151717e+02
#> 15.9 6.271768e+02 6.291934e+02 6.315922e+02 6.339910e+02 6.363898e+02
#> 16   6.470098e+02 6.490378e+02 6.515377e+02 6.540375e+02 6.565374e+02
#> 16.1 6.723047e+02 6.747228e+02 6.769509e+02 6.791791e+02 6.814072e+02
#> 16.2 6.920332e+02 6.948331e+02 6.982964e+02 7.017596e+02 7.052228e+02
#> 16.3 7.169866e+02 7.189096e+02 7.209947e+02 7.230799e+02 7.251650e+02
#> 16.4 7.409552e+02 7.435820e+02 7.461650e+02 7.487481e+02 7.513311e+02
#> 16.5 7.636370e+02 7.662319e+02 7.689531e+02 7.716742e+02 7.743954e+02
#> 16.6 7.902334e+02 7.923803e+02 7.944809e+02 7.965815e+02 7.986821e+02
#> 16.7 8.132284e+02 8.168200e+02 8.204350e+02 8.240500e+02 8.276650e+02
#> 16.8 8.380124e+02 8.410697e+02 8.441270e+02 8.471842e+02 8.502415e+02
#> 16.9 8.649669e+02 8.684179e+02 8.718688e+02 8.753198e+02 8.787708e+02
#> 17   8.970684e+02 8.993649e+02 9.016613e+02 9.039578e+02 9.062543e+02
#> 17.1 9.250036e+02 9.265393e+02 9.280751e+02 9.296108e+02 9.311466e+02
#> 17.2 9.553656e+02 9.566518e+02 9.579381e+02 9.592244e+02 9.605107e+02
#> 17.3 9.749389e+02 9.791069e+02 9.832748e+02 9.874427e+02 9.916107e+02
#> 17.4 1.010275e+03 1.012418e+03 1.014560e+03 1.016703e+03 1.018846e+03
#> 17.5 1.029117e+03 1.033193e+03 1.037268e+03 1.041344e+03 1.045419e+03
#> 17.6 1.063330e+03 1.065495e+03 1.067660e+03 1.069825e+03 1.071990e+03
#> 17.7 1.095966e+03 1.099167e+03 1.102367e+03 1.105567e+03 1.108767e+03
#> 17.8 1.128916e+03 1.131469e+03 1.134021e+03 1.136574e+03 1.139126e+03
#> 17.9 1.156978e+03 1.160583e+03 1.164188e+03 1.167793e+03 1.171398e+03
#> 18   1.194166e+03 1.196448e+03 1.198729e+03 1.201011e+03 1.203292e+03
#> 18.1 1.230075e+03 1.231506e+03 1.232937e+03 1.234369e+03 1.235800e+03
#> 18.2 1.256197e+03 1.259615e+03 1.263034e+03 1.266453e+03 1.269861e+03
#> 18.3 1.288823e+03 1.292025e+03 1.295228e+03 1.298430e+03 1.301768e+03
#> 18.4 1.329317e+03 1.334381e+03 1.339444e+03 1.344508e+03 1.348805e+03
#> 18.5 1.356999e+03 1.361089e+03 1.365178e+03 1.369268e+03 1.373356e+03
#> 18.6 1.397350e+03 1.401032e+03 1.404714e+03 1.408396e+03 1.412082e+03
#> 18.7 1.434491e+03 1.438359e+03 1.442226e+03 1.446093e+03 1.450023e+03
#> 18.8 1.472339e+03 1.475052e+03 1.477764e+03 1.480476e+03 1.483592e+03
#> 18.9 1.506138e+03 1.510334e+03 1.514530e+03 1.518726e+03 1.522672e+03
#> 19   1.548683e+03 1.554813e+03 1.560942e+03 1.567071e+03 1.571108e+03
#> 19.1 1.587898e+03 1.591369e+03 1.594840e+03 1.598312e+03 1.601470e+03
#> 19.2 1.627353e+03 1.635046e+03 1.642739e+03 1.650432e+03 1.653386e+03
#> 19.3 1.666053e+03 1.670556e+03 1.675058e+03 1.679561e+03 1.683493e+03
#> 19.4 1.708887e+03 1.714155e+03 1.719423e+03 1.724691e+03 1.728062e+03
#> 19.5 1.746175e+03 1.749111e+03 1.752047e+03 1.754984e+03 1.760400e+03
#> 19.6 1.789535e+03 1.792829e+03 1.796123e+03 1.799417e+03 1.804780e+03
#> 19.7 1.835192e+03 1.839661e+03 1.844130e+03 1.848529e+03 1.851875e+03
#> 19.8 1.874069e+03 1.878425e+03 1.882782e+03 1.887096e+03 1.891127e+03
#> 19.9 1.917705e+03 1.922841e+03 1.927976e+03 1.933031e+03 1.937760e+03
#> 20   1.968371e+03 1.974102e+03 1.979833e+03 1.984862e+03 1.987971e+03
#> 20.1 2.007207e+03 2.010569e+03 2.013931e+03 2.018475e+03 2.025354e+03
#> 20.2 2.058198e+03 2.061519e+03 2.064841e+03 2.068507e+03 2.072678e+03
#> 20.3 2.101916e+03 2.107813e+03 2.113711e+03 2.119021e+03 2.123677e+03
#> 20.4 2.142604e+03 2.143666e+03 2.144728e+03 2.149226e+03 2.156643e+03
#> 20.5 2.198529e+03 2.204931e+03 2.211333e+03 2.217088e+03 2.222427e+03
#> 20.6 2.254116e+03 2.259324e+03 2.264531e+03 2.269203e+03 2.273619e+03
#> 20.7 2.298446e+03 2.302246e+03 2.306046e+03 2.311491e+03 2.317496e+03
#> 20.8 2.344490e+03 2.347244e+03 2.349997e+03 2.356070e+03 2.362900e+03
#> 20.9 2.397494e+03 2.402082e+03 2.406669e+03 2.412161e+03 2.417773e+03
#> 21   2.444801e+03 2.448134e+03 2.451467e+03 2.459434e+03 2.467642e+03
#> 21.1 2.504059e+03 2.507969e+03 2.511898e+03 2.516760e+03 2.521622e+03
#> 21.2 2.551823e+03 2.557022e+03 2.562381e+03 2.569416e+03 2.576452e+03
#> 21.3 2.604590e+03 2.607116e+03 2.610894e+03 2.621460e+03 2.632026e+03
#> 21.4 2.665178e+03 2.666262e+03 2.668785e+03 2.676297e+03 2.683808e+03
#> 21.5 2.719611e+03 2.724279e+03 2.729632e+03 2.736646e+03 2.743660e+03
#> 21.6 2.774726e+03 2.778427e+03 2.783962e+03 2.792746e+03 2.801531e+03
#> 21.7 2.834699e+03 2.837728e+03 2.842690e+03 2.850225e+03 2.857760e+03
#> 21.8 2.898574e+03 2.904839e+03 2.909435e+03 2.912344e+03 2.915253e+03
#> 21.9 2.939850e+03 2.944782e+03 2.950070e+03 2.955632e+03 2.961194e+03
#> 22   3.000503e+03 3.007714e+03 3.013535e+03 3.018551e+03 3.023567e+03
#> 22.1 3.065888e+03 3.074236e+03 3.080696e+03 3.086354e+03 3.092012e+03
#> 22.2 3.119533e+03 3.123471e+03 3.129435e+03 3.136003e+03 3.142571e+03
#> 22.3 3.175021e+03 3.179760e+03 3.184857e+03 3.190022e+03 3.195188e+03
#> 22.4 3.243785e+03 3.253496e+03 3.258804e+03 3.263661e+03 3.268518e+03
#> 22.5 3.308738e+03 3.316406e+03 3.322772e+03 3.329105e+03 3.335437e+03
#> 22.6 3.373867e+03 3.380395e+03 3.388824e+03 3.397254e+03 3.405683e+03
#> 22.7 3.432233e+03 3.435733e+03 3.446863e+03 3.457993e+03 3.469123e+03
#> 22.8 3.489993e+03 3.491293e+03 3.498250e+03 3.505206e+03 3.512163e+03
#> 22.9 3.574303e+03 3.584655e+03 3.590640e+03 3.596625e+03 3.602610e+03
#> 23   3.635183e+03 3.640602e+03 3.646480e+03 3.652359e+03 3.658237e+03
#> 23.1 3.695710e+03 3.701687e+03 3.707012e+03 3.712338e+03 3.717664e+03
#> 23.2 3.765961e+03 3.773868e+03 3.780427e+03 3.786987e+03 3.793547e+03
#> 23.3 3.835649e+03 3.842288e+03 3.848441e+03 3.854593e+03 3.860746e+03
#> 23.4 3.905235e+03 3.912918e+03 3.920511e+03 3.928103e+03 3.935696e+03
#> 23.5 3.967817e+03 3.972968e+03 3.978361e+03 3.983753e+03 3.989145e+03
#> 23.6 4.056349e+03 4.064138e+03 4.070036e+03 4.075934e+03 4.081832e+03
#> 23.7 4.136040e+03 4.141671e+03 4.146216e+03 4.150760e+03 4.155304e+03
#> 23.8 4.189752e+03 4.197940e+03 4.206470e+03 4.214999e+03 4.223529e+03
#> 23.9 4.251099e+03 4.258404e+03 4.265972e+03 4.273540e+03 4.281108e+03
#> 24   4.355548e+03           NA           NA           NA           NA
#>              0.05         0.06         0.07         0.08         0.09
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 3.902194e-05 1.020707e-04
#> 7.7  1.369517e-03 1.772078e-03 2.174639e-03 2.964715e-03 3.993396e-03
#> 7.8  1.245544e-02 1.435851e-02 1.626159e-02 1.869382e-02 2.145188e-02
#> 7.9  4.303650e-02 4.771304e-02 5.258465e-02 5.745626e-02 6.314149e-02
#> 8    1.036673e-01 1.114481e-01 1.195236e-01 1.286072e-01 1.376908e-01
#> 8.1  2.008436e-01 2.124073e-01 2.256435e-01 2.389369e-01 2.526437e-01
#> 8.2  3.442938e-01 3.607094e-01 3.783196e-01 3.972949e-01 4.162702e-01
#> 8.3  5.423620e-01 5.644859e-01 5.864352e-01 6.109579e-01 6.356479e-01
#> 8.4  7.978559e-01 8.283017e-01 8.587474e-01 8.893667e-01 9.200902e-01
#> 8.5  1.127216e+00 1.161241e+00 1.195512e+00 1.235437e+00 1.275362e+00
#> 8.6  1.525435e+00 1.572050e+00 1.617740e+00 1.662098e+00 1.706457e+00
#> 8.7  2.001978e+00 2.053178e+00 2.109014e+00 2.167193e+00 2.225373e+00
#> 8.8  2.584098e+00 2.646567e+00 2.709036e+00 2.774379e+00 2.840953e+00
#> 8.9  3.253012e+00 3.328541e+00 3.401970e+00 3.474926e+00 3.547881e+00
#> 9    4.017105e+00 4.103226e+00 4.189129e+00 4.273079e+00 4.357029e+00
#> 9.1  4.907138e+00 4.989475e+00 5.076845e+00 5.171072e+00 5.265298e+00
#> 9.2  5.892757e+00 6.000819e+00 6.112581e+00 6.231709e+00 6.351655e+00
#> 9.3  7.017491e+00 7.142165e+00 7.265039e+00 7.383715e+00 7.502391e+00
#> 9.4  8.253631e+00 8.390107e+00 8.545139e+00 8.704144e+00 8.824178e+00
#> 9.5  9.641450e+00 9.785720e+00 9.929991e+00 1.007713e+01 1.022646e+01
#> 9.6  1.118072e+01 1.134473e+01 1.150692e+01 1.166911e+01 1.183130e+01
#> 9.7  1.284552e+01 1.301271e+01 1.317990e+01 1.335449e+01 1.354019e+01
#> 9.8  1.465896e+01 1.484291e+01 1.502063e+01 1.519835e+01 1.537608e+01
#> 9.9  1.664712e+01 1.685397e+01 1.706082e+01 1.726767e+01 1.747451e+01
#> 10   1.882122e+01 1.903857e+01 1.925592e+01 1.947328e+01 1.969063e+01
#> 10.1 2.112889e+01 2.137418e+01 2.161946e+01 2.186475e+01 2.211004e+01
#> 10.2 2.367928e+01 2.394405e+01 2.420881e+01 2.447358e+01 2.473834e+01
#> 10.3 2.628350e+01 2.658749e+01 2.689148e+01 2.719547e+01 2.749946e+01
#> 10.4 2.923774e+01 2.952237e+01 2.980699e+01 3.009161e+01 3.037623e+01
#> 10.5 3.221514e+01 3.258016e+01 3.294518e+01 3.331020e+01 3.367522e+01
#> 10.6 3.547856e+01 3.577775e+01 3.607693e+01 3.637611e+01 3.667530e+01
#> 10.7 3.913568e+01 3.948164e+01 3.982759e+01 4.017354e+01 4.051950e+01
#> 10.8 4.299189e+01 4.330522e+01 4.361855e+01 4.393188e+01 4.424521e+01
#> 10.9 4.658562e+01 4.704542e+01 4.750521e+01 4.796500e+01 4.842017e+01
#> 11   5.072330e+01 5.117076e+01 5.161822e+01 5.206569e+01 5.251267e+01
#> 11.1 5.517709e+01 5.562370e+01 5.607030e+01 5.651690e+01 5.696975e+01
#> 11.2 5.975258e+01 6.015377e+01 6.055497e+01 6.095616e+01 6.139442e+01
#> 11.3 6.464838e+01 6.517814e+01 6.570790e+01 6.623766e+01 6.675498e+01
#> 11.4 6.974403e+01 7.026519e+01 7.078636e+01 7.130752e+01 7.181402e+01
#> 11.5 7.482764e+01 7.538892e+01 7.595020e+01 7.651148e+01 7.708060e+01
#> 11.6 8.061841e+01 8.124223e+01 8.186605e+01 8.248987e+01 8.307049e+01
#> 11.7 8.630427e+01 8.683391e+01 8.736354e+01 8.789318e+01 8.853131e+01
#> 11.8 9.262212e+01 9.322547e+01 9.382883e+01 9.443219e+01 9.505204e+01
#> 11.9 9.900118e+01 9.973543e+01 1.004697e+02 1.012040e+02 1.018393e+02
#> 12   1.055495e+02 1.062126e+02 1.068757e+02 1.075388e+02 1.082476e+02
#> 12.1 1.126431e+02 1.134038e+02 1.141646e+02 1.149254e+02 1.156420e+02
#> 12.2 1.198768e+02 1.205732e+02 1.212696e+02 1.219660e+02 1.228842e+02
#> 12.3 1.280220e+02 1.287230e+02 1.294240e+02 1.301248e+02 1.308045e+02
#> 12.4 1.352904e+02 1.361698e+02 1.370492e+02 1.379331e+02 1.388724e+02
#> 12.5 1.438433e+02 1.444676e+02 1.450919e+02 1.457574e+02 1.466677e+02
#> 12.6 1.518582e+02 1.526442e+02 1.534301e+02 1.542327e+02 1.550968e+02
#> 12.7 1.602972e+02 1.611684e+02 1.620396e+02 1.629651e+02 1.640296e+02
#> 12.8 1.698655e+02 1.706919e+02 1.715183e+02 1.723627e+02 1.732406e+02
#> 12.9 1.786444e+02 1.795795e+02 1.805146e+02 1.815034e+02 1.825671e+02
#> 13   1.888588e+02 1.898855e+02 1.909122e+02 1.919393e+02 1.929668e+02
#> 13.1 1.989209e+02 1.998648e+02 2.008086e+02 2.019295e+02 2.031927e+02
#> 13.2 2.097598e+02 2.106319e+02 2.115040e+02 2.125250e+02 2.136362e+02
#> 13.3 2.205885e+02 2.218070e+02 2.230255e+02 2.240295e+02 2.249375e+02
#> 13.4 2.306806e+02 2.316968e+02 2.327131e+02 2.339911e+02 2.353522e+02
#> 13.5 2.433512e+02 2.446524e+02 2.459535e+02 2.468653e+02 2.476960e+02
#> 13.6 2.540293e+02 2.553315e+02 2.566337e+02 2.578066e+02 2.589643e+02
#> 13.7 2.669167e+02 2.684177e+02 2.699187e+02 2.710374e+02 2.721418e+02
#> 13.8 2.796764e+02 2.810838e+02 2.824849e+02 2.837022e+02 2.849194e+02
#> 13.9 2.924163e+02 2.936966e+02 2.949540e+02 2.960068e+02 2.970596e+02
#> 14   3.061691e+02 3.081128e+02 3.098881e+02 3.108357e+02 3.117832e+02
#> 14.1 3.190359e+02 3.204728e+02 3.219009e+02 3.233008e+02 3.247007e+02
#> 14.2 3.343743e+02 3.361636e+02 3.378161e+02 3.391577e+02 3.404993e+02
#> 14.3 3.480081e+02 3.491878e+02 3.505435e+02 3.521938e+02 3.538442e+02
#> 14.4 3.633565e+02 3.648925e+02 3.663584e+02 3.677360e+02 3.691136e+02
#> 14.5 3.784902e+02 3.801873e+02 3.818235e+02 3.834012e+02 3.849789e+02
#> 14.6 3.946555e+02 3.962925e+02 3.977893e+02 3.991840e+02 4.005787e+02
#> 14.7 4.097026e+02 4.113064e+02 4.130597e+02 4.148944e+02 4.167292e+02
#> 14.8 4.259732e+02 4.273286e+02 4.292497e+02 4.313957e+02 4.335418e+02
#> 14.9 4.442316e+02 4.457946e+02 4.476741e+02 4.496410e+02 4.516078e+02
#> 15   4.611064e+02 4.624702e+02 4.643826e+02 4.663903e+02 4.683980e+02
#> 15.1 4.802073e+02 4.821541e+02 4.838506e+02 4.855255e+02 4.872004e+02
#> 15.2 4.976309e+02 4.994022e+02 5.012760e+02 5.031509e+02 5.050258e+02
#> 15.3 5.155460e+02 5.172522e+02 5.191660e+02 5.210798e+02 5.229936e+02
#> 15.4 5.343453e+02 5.363212e+02 5.389537e+02 5.415862e+02 5.442187e+02
#> 15.5 5.575023e+02 5.594122e+02 5.608279e+02 5.622435e+02 5.636591e+02
#> 15.6 5.771339e+02 5.794331e+02 5.809025e+02 5.823718e+02 5.838412e+02
#> 15.7 5.936728e+02 5.955519e+02 5.977824e+02 6.000129e+02 6.022434e+02
#> 15.8 6.179301e+02 6.202766e+02 6.220016e+02 6.237267e+02 6.254517e+02
#> 15.9 6.387886e+02 6.407931e+02 6.423472e+02 6.439014e+02 6.454556e+02
#> 16   6.590372e+02 6.616278e+02 6.642970e+02 6.669663e+02 6.696355e+02
#> 16.1 6.836353e+02 6.855038e+02 6.871362e+02 6.887685e+02 6.904009e+02
#> 16.2 7.086861e+02 7.108528e+02 7.123863e+02 7.139197e+02 7.154532e+02
#> 16.3 7.272502e+02 7.298474e+02 7.326243e+02 7.354013e+02 7.381783e+02
#> 16.4 7.539141e+02 7.559604e+02 7.578796e+02 7.597987e+02 7.617179e+02
#> 16.5 7.771165e+02 7.797498e+02 7.823707e+02 7.849916e+02 7.876125e+02
#> 16.6 8.007827e+02 8.032546e+02 8.057480e+02 8.082415e+02 8.107349e+02
#> 16.7 8.312495e+02 8.325857e+02 8.339219e+02 8.352582e+02 8.365944e+02
#> 16.8 8.532385e+02 8.555580e+02 8.578774e+02 8.601969e+02 8.625164e+02
#> 16.9 8.821554e+02 8.851643e+02 8.881731e+02 8.911819e+02 8.941908e+02
#> 17   9.087777e+02 9.121138e+02 9.154498e+02 9.187859e+02 9.221220e+02
#> 17.1 9.335530e+02 9.381264e+02 9.426998e+02 9.472732e+02 9.518466e+02
#> 17.2 9.621965e+02 9.646083e+02 9.670201e+02 9.694320e+02 9.718438e+02
#> 17.3 9.953169e+02 9.983940e+02 1.001471e+03 1.004548e+03 1.007625e+03
#> 17.4 1.020632e+03 1.022049e+03 1.023467e+03 1.024885e+03 1.026302e+03
#> 17.5 1.048885e+03 1.051872e+03 1.054858e+03 1.057845e+03 1.060832e+03
#> 17.6 1.075458e+03 1.079697e+03 1.083937e+03 1.088176e+03 1.092415e+03
#> 17.7 1.112166e+03 1.115653e+03 1.119139e+03 1.122626e+03 1.126112e+03
#> 17.8 1.141940e+03 1.144834e+03 1.147728e+03 1.150622e+03 1.153516e+03
#> 17.9 1.175386e+03 1.179452e+03 1.183517e+03 1.187583e+03 1.191648e+03
#> 18   1.208095e+03 1.213173e+03 1.218252e+03 1.223330e+03 1.228409e+03
#> 18.1 1.239147e+03 1.242555e+03 1.245963e+03 1.249370e+03 1.252778e+03
#> 18.2 1.273012e+03 1.276164e+03 1.279315e+03 1.282466e+03 1.285621e+03
#> 18.3 1.306249e+03 1.310730e+03 1.315211e+03 1.319691e+03 1.324254e+03
#> 18.4 1.349483e+03 1.350162e+03 1.350840e+03 1.351518e+03 1.352909e+03
#> 18.5 1.377441e+03 1.381526e+03 1.385610e+03 1.389695e+03 1.393668e+03
#> 18.6 1.415779e+03 1.419475e+03 1.423172e+03 1.426868e+03 1.430624e+03
#> 18.7 1.454053e+03 1.458082e+03 1.462112e+03 1.466142e+03 1.469627e+03
#> 18.8 1.487206e+03 1.490820e+03 1.494434e+03 1.498048e+03 1.501942e+03
#> 18.9 1.526382e+03 1.530092e+03 1.533802e+03 1.537513e+03 1.542554e+03
#> 19   1.573658e+03 1.576208e+03 1.578758e+03 1.581307e+03 1.584427e+03
#> 19.1 1.604462e+03 1.607454e+03 1.610447e+03 1.613439e+03 1.619660e+03
#> 19.2 1.654509e+03 1.655631e+03 1.656753e+03 1.657875e+03 1.661550e+03
#> 19.3 1.687273e+03 1.691053e+03 1.694833e+03 1.698613e+03 1.703619e+03
#> 19.4 1.731119e+03 1.734176e+03 1.737232e+03 1.740289e+03 1.743238e+03
#> 19.5 1.766013e+03 1.771627e+03 1.777241e+03 1.782855e+03 1.786241e+03
#> 19.6 1.810155e+03 1.815530e+03 1.820905e+03 1.826254e+03 1.830723e+03
#> 19.7 1.855221e+03 1.858567e+03 1.861913e+03 1.865356e+03 1.869713e+03
#> 19.8 1.895158e+03 1.899189e+03 1.903220e+03 1.907434e+03 1.912570e+03
#> 19.9 1.942489e+03 1.947218e+03 1.951947e+03 1.956909e+03 1.962640e+03
#> 20   1.991080e+03 1.994188e+03 1.997297e+03 2.000482e+03 2.003845e+03
#> 20.1 2.032233e+03 2.039112e+03 2.045991e+03 2.051554e+03 2.054876e+03
#> 20.2 2.076850e+03 2.081022e+03 2.085193e+03 2.090121e+03 2.096019e+03
#> 20.3 2.128333e+03 2.132989e+03 2.137645e+03 2.140480e+03 2.141542e+03
#> 20.4 2.164060e+03 2.171476e+03 2.178893e+03 2.185726e+03 2.192128e+03
#> 20.5 2.227767e+03 2.233107e+03 2.238447e+03 2.243702e+03 2.248909e+03
#> 20.6 2.278035e+03 2.282452e+03 2.286868e+03 2.290846e+03 2.294646e+03
#> 20.7 2.323502e+03 2.329507e+03 2.335513e+03 2.338982e+03 2.341736e+03
#> 20.8 2.369730e+03 2.376561e+03 2.383391e+03 2.388319e+03 2.392906e+03
#> 20.9 2.423386e+03 2.428998e+03 2.434610e+03 2.438134e+03 2.441467e+03
#> 21   2.475849e+03 2.484056e+03 2.492264e+03 2.496239e+03 2.500149e+03
#> 21.1 2.526484e+03 2.531347e+03 2.536227e+03 2.541426e+03 2.546624e+03
#> 21.2 2.583488e+03 2.590523e+03 2.597011e+03 2.599538e+03 2.602064e+03
#> 21.3 2.642592e+03 2.653158e+03 2.661925e+03 2.663009e+03 2.664093e+03
#> 21.4 2.691319e+03 2.698830e+03 2.705607e+03 2.710275e+03 2.714943e+03
#> 21.5 2.750674e+03 2.757688e+03 2.763621e+03 2.767322e+03 2.771024e+03
#> 21.6 2.810316e+03 2.819100e+03 2.825613e+03 2.828642e+03 2.831671e+03
#> 21.7 2.865295e+03 2.872830e+03 2.879777e+03 2.886043e+03 2.892308e+03
#> 21.8 2.918161e+03 2.921070e+03 2.925053e+03 2.929986e+03 2.934918e+03
#> 21.9 2.966756e+03 2.972318e+03 2.978869e+03 2.986081e+03 2.993292e+03
#> 22   3.028583e+03 3.033600e+03 3.040842e+03 3.049190e+03 3.057539e+03
#> 22.1 3.097671e+03 3.103329e+03 3.107721e+03 3.111658e+03 3.115596e+03
#> 22.2 3.149139e+03 3.155707e+03 3.160803e+03 3.165542e+03 3.170282e+03
#> 22.3 3.200353e+03 3.205519e+03 3.214652e+03 3.224363e+03 3.234074e+03
#> 22.4 3.273375e+03 3.278231e+03 3.285734e+03 3.293402e+03 3.301070e+03
#> 22.5 3.341770e+03 3.348104e+03 3.354545e+03 3.360986e+03 3.367426e+03
#> 22.6 3.414113e+03 3.422084e+03 3.424621e+03 3.427159e+03 3.429696e+03
#> 22.7 3.480254e+03 3.489766e+03 3.489822e+03 3.489879e+03 3.489936e+03
#> 22.8 3.519119e+03 3.527114e+03 3.538911e+03 3.550708e+03 3.562506e+03
#> 22.9 3.608594e+03 3.614359e+03 3.619565e+03 3.624771e+03 3.629977e+03
#> 23   3.664115e+03 3.670172e+03 3.676556e+03 3.682941e+03 3.689325e+03
#> 23.1 3.722989e+03 3.729866e+03 3.738889e+03 3.747913e+03 3.756937e+03
#> 23.2 3.800107e+03 3.806965e+03 3.814136e+03 3.821307e+03 3.828478e+03
#> 23.3 3.866898e+03 3.873975e+03 3.881790e+03 3.889605e+03 3.897420e+03
#> 23.4 3.943288e+03 3.949067e+03 3.953755e+03 3.958442e+03 3.963129e+03
#> 23.5 3.994537e+03 4.005073e+03 4.017892e+03 4.030711e+03 4.043530e+03
#> 23.6 4.087729e+03 4.096635e+03 4.106486e+03 4.116337e+03 4.126188e+03
#> 23.7 4.159848e+03 4.165626e+03 4.171658e+03 4.177689e+03 4.183721e+03
#> 23.8 4.232058e+03 4.236261e+03 4.239971e+03 4.243680e+03 4.247389e+03
#> 23.9 4.288677e+03 4.301891e+03 4.315305e+03 4.328720e+03 4.342134e+03
#> 24             NA           NA           NA           NA           NA

# }