Plot bdrc model objects

Visualize results from model ojbects in bdrc, plm0, plm, gplm0,gplm

plot_fun(x, type = "rating_curve", param = NULL, transformed = F, title = NULL)

Arguments

x
an object of class "plm0","plm","gplm0" or "gplm".

type
a character denoting what type of plot should be drawn. Defaults to "rating_curve". Possible types are

"rating_curve" to plot the rating curve.

"rating_curve_mean" to plot the posterior mean of the rating curve.

"f" to plot the power-law exponent.

"beta" to plot the random effect in the power-law exponent.

"sigma_eps" to plot the standard deviation on the data level.

"residuals" to plot the log residuals.

"trace" to plot trace plots of parameters given in param.

"histogram" to plot histograms of parameters given in param.

param
a character vector with the parameters to plot. Defaults to NULL and is only used if type is "trace" or "histogram". Allowed values are the parameters given in the model summary of x as well as "hyperparameters" or "latent_parameters" for specific groups of parameters.

transformed
a logical value indicating whether the quantity should be plotted on a transformed scale used during the Bayesian inference. Defaults to FALSE.

title
a character denoting the title of the plot. Defaults to NULL, i.e. no title.

Value

returns an object of class ggplot2 or Grob object.