Predict method for discharge rating curves

Predict the discharge for given stage values based on a discharge rating curve model object.

# S3 method for plm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for plm
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm
predict(object, newdata = NULL, wide = FALSE, ...)

Arguments

object
an object of class "plm0", "plm", "gplm0" or "gplm".

newdata
a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.

wide
a logical value denoting whether to produce a wide prediction output.If TRUE, then the output is a table with median prediction values for an equally spaced grid of stages with 1 cm increments, each row containing predictions in a decimeter range of stages.

...
not used in this function

Value

an object of class "data.frame" with four columns, h (stage), lower (2.5% posterior predictive quantile), median (50% posterior predictive quantile), upper (97.5% posterior predictive quantile). If wide=TRUE, a matrix as described above (see wide parameter) is returned.

Functions

predict.plm0: Predict method for plm0

predict.plm: Predict method for plm

predict.gplm0: Predict method for gplm0

predict.gplm: Predict method for gplm

See also

Examples

# \donttest{
data(krokfors)
plm0.fit <- plm0(Q~W,krokfors,h_max=24)
#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(plm0.fit,newdata=seq(23,24,by=0.01))

#>         h    lower   median     upper
#> 1   23.00 1534.143 3618.840  8475.553
#> 2   23.01 1533.575 3625.891  8481.645
#> 3   23.02 1536.154 3636.481  8493.386
#> 4   23.03 1538.733 3647.071  8505.127
#> 5   23.04 1541.313 3657.661  8516.868
#> 6   23.05 1543.892 3668.252  8528.609
#> 7   23.06 1547.417 3675.944  8554.253
#> 8   23.07 1552.690 3678.281  8605.597
#> 9   23.08 1557.963 3680.617  8656.940
#> 10  23.09 1563.237 3682.954  8708.284
#> 11  23.10 1568.510 3685.290  8759.628
#> 12  23.11 1571.541 3690.462  8787.726
#> 13  23.12 1570.995 3700.157  8778.725
#> 14  23.13 1570.450 3709.853  8769.725
#> 15  23.14 1569.904 3719.548  8760.724
#> 16  23.15 1569.358 3729.244  8751.723
#> 17  23.16 1570.140 3737.047  8755.120
#> 18  23.17 1572.760 3742.230  8775.682
#> 19  23.18 1575.381 3747.413  8796.244
#> 20  23.19 1578.001 3752.596  8816.806
#> 21  23.20 1580.622 3757.778  8837.368
#> 22  23.21 1584.007 3762.965  8862.970
#> 23  23.22 1588.313 3768.156  8894.644
#> 24  23.23 1592.620 3773.348  8926.318
#> 25  23.24 1596.927 3778.539  8957.992
#> 26  23.25 1601.233 3783.730  8989.665
#> 27  23.26 1604.826 3789.213  8999.080
#> 28  23.27 1607.669 3795.002  8985.109
#> 29  23.28 1610.512 3800.791  8971.139
#> 30  23.29 1613.355 3806.580  8957.169
#> 31  23.30 1616.198 3812.369  8943.199
#> 32  23.31 1619.062 3818.479  8943.842
#> 33  23.32 1621.944 3824.884  8957.877
#> 34  23.33 1624.826 3831.289  8971.911
#> 35  23.34 1627.708 3837.693  8985.945
#> 36  23.35 1630.590 3844.098  8999.980
#> 37  23.36 1632.035 3852.512  9024.540
#> 38  23.37 1632.333 3862.531  9057.505
#> 39  23.38 1632.631 3872.549  9090.471
#> 40  23.39 1632.929 3882.568  9123.437
#> 41  23.40 1633.227 3892.587  9156.402
#> 42  23.41 1632.385 3899.839  9175.795
#> 43  23.42 1630.753 3905.169  9185.760
#> 44  23.43 1629.120 3910.499  9195.726
#> 45  23.44 1627.488 3915.829  9205.691
#> 46  23.45 1625.856 3921.159  9215.657
#> 47  23.46 1631.793 3927.121  9239.543
#> 48  23.47 1642.285 3933.464  9271.807
#> 49  23.48 1652.778 3939.807  9304.070
#> 50  23.49 1663.271 3946.149  9336.334
#> 51  23.50 1673.763 3952.492  9368.597
#> 52  23.51 1677.057 3963.693  9399.969
#> 53  23.52 1676.617 3977.414  9430.878
#> 54  23.53 1676.177 3991.134  9461.786
#> 55  23.54 1675.737 4004.855  9492.695
#> 56  23.55 1675.297 4018.576  9523.604
#> 57  23.56 1679.117 4026.010  9548.268
#> 58  23.57 1684.828 4030.653  9570.160
#> 59  23.58 1690.539 4035.296  9592.052
#> 60  23.59 1696.251 4039.939  9613.944
#> 61  23.60 1701.962 4044.582  9635.837
#> 62  23.61 1705.087 4050.996  9644.294
#> 63  23.62 1707.240 4058.075  9647.700
#> 64  23.63 1709.392 4065.155  9651.107
#> 65  23.64 1711.545 4072.234  9654.513
#> 66  23.65 1713.698 4079.313  9657.919
#> 67  23.66 1718.027 4088.675  9675.419
#> 68  23.67 1723.040 4098.754  9697.346
#> 69  23.68 1728.052 4108.833  9719.273
#> 70  23.69 1733.065 4118.912  9741.201
#> 71  23.70 1738.078 4128.991  9763.128
#> 72  23.71 1739.132 4133.414  9782.633
#> 73  23.72 1739.166 4136.378  9801.513
#> 74  23.73 1739.200 4139.342  9820.393
#> 75  23.74 1739.233 4142.306  9839.273
#> 76  23.75 1739.267 4145.270  9858.153
#> 77  23.76 1737.021 4152.597  9867.804
#> 78  23.77 1734.306 4160.821  9875.554
#> 79  23.78 1731.591 4169.046  9883.304
#> 80  23.79 1728.875 4177.271  9891.054
#> 81  23.80 1726.160 4185.496  9898.804
#> 82  23.81 1732.623 4187.845  9933.318
#> 83  23.82 1740.538 4189.264  9972.068
#> 84  23.83 1748.452 4190.684 10010.818
#> 85  23.84 1756.367 4192.104 10049.567
#> 86  23.85 1764.282 4193.523 10088.317
#> 87  23.86 1766.766 4203.394 10098.137
#> 88  23.87 1768.630 4214.229 10104.655
#> 89  23.88 1770.493 4225.064 10111.173
#> 90  23.89 1772.357 4235.899 10117.690
#> 91  23.90 1774.221 4246.735 10124.208
#> 92  23.91 1777.424 4255.282 10133.341
#> 93  23.92 1780.725 4263.662 10142.665
#> 94  23.93 1784.026 4272.041 10151.990
#> 95  23.94 1787.328 4280.421 10161.315
#> 96  23.95 1790.629 4288.800 10170.639
#> 97  23.96 1797.777 4299.913 10176.603
#> 98  23.97 1805.061 4311.122 10182.448
#> 99  23.98 1812.345 4322.331 10188.294
#> 100 23.99 1819.629 4333.540 10194.139
#> 101 24.00 1826.913 4344.749 10199.984

predict(plm0.fit,wide=TRUE)

#>                 0         0.01         0.02         0.03         0.04
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> 7.7  1.607674e-04 2.221544e-04 2.835414e-04 5.565447e-04 9.598460e-04
#> 7.8  4.990334e-03 6.008653e-03 7.026972e-03 8.605142e-03 1.052803e-02
#> 7.9  2.433573e-02 2.765550e-02 3.097528e-02 3.469272e-02 3.871056e-02
#> 8    6.905577e-02 7.499670e-02 8.170187e-02 8.840704e-02 9.532107e-02
#> 8.1  1.467512e-01 1.572185e-01 1.676859e-01 1.781533e-01 1.896342e-01
#> 8.2  2.665228e-01 2.805018e-01 2.957334e-01 3.116229e-01 3.275124e-01
#> 8.3  4.365842e-01 4.568385e-01 4.770113e-01 4.970619e-01 5.182981e-01
#> 8.4  6.613599e-01 6.860606e-01 7.117756e-01 7.401644e-01 7.685532e-01
#> 8.5  9.517035e-01 9.854048e-01 1.019593e+00 1.053257e+00 1.086488e+00
#> 8.6  1.312552e+00 1.355379e+00 1.399133e+00 1.442888e+00 1.486004e+00
#> 8.7  1.757134e+00 1.807754e+00 1.858374e+00 1.908994e+00 1.959499e+00
#> 8.8  2.278119e+00 2.341108e+00 2.404097e+00 2.466711e+00 2.526779e+00
#> 8.9  2.901741e+00 2.971044e+00 3.041709e+00 3.112373e+00 3.182706e+00
#> 9    3.624085e+00 3.711867e+00 3.785750e+00 3.859633e+00 3.934240e+00
#> 9.1  4.457922e+00 4.538400e+00 4.624826e+00 4.718437e+00 4.812003e+00
#> 9.2  5.398804e+00 5.501610e+00 5.605122e+00 5.714979e+00 5.816331e+00
#> 9.3  6.436549e+00 6.550817e+00 6.670333e+00 6.789849e+00 6.908355e+00
#> 9.4  7.654567e+00 7.768733e+00 7.882900e+00 7.997066e+00 8.134096e+00
#> 9.5  8.923039e+00 9.086087e+00 9.237866e+00 9.376768e+00 9.515957e+00
#> 9.6  1.040260e+01 1.056159e+01 1.071590e+01 1.086610e+01 1.101631e+01
#> 9.7  1.201980e+01 1.218154e+01 1.234328e+01 1.250502e+01 1.267385e+01
#> 9.8  1.373860e+01 1.391507e+01 1.410020e+01 1.430028e+01 1.450037e+01
#> 9.9  1.569156e+01 1.588372e+01 1.607588e+01 1.626805e+01 1.646021e+01
#> 10   1.771733e+01 1.792869e+01 1.814004e+01 1.835140e+01 1.856276e+01
#> 10.1 1.997683e+01 2.021012e+01 2.044342e+01 2.067671e+01 2.091001e+01
#> 10.2 2.235467e+01 2.263958e+01 2.292449e+01 2.320939e+01 2.349430e+01
#> 10.3 2.494002e+01 2.522274e+01 2.550546e+01 2.578819e+01 2.607091e+01
#> 10.4 2.777351e+01 2.804042e+01 2.830733e+01 2.857424e+01 2.884114e+01
#> 10.5 3.067714e+01 3.097088e+01 3.126461e+01 3.155834e+01 3.185208e+01
#> 10.6 3.393018e+01 3.423866e+01 3.454713e+01 3.485561e+01 3.516409e+01
#> 10.7 3.721185e+01 3.756934e+01 3.792683e+01 3.828432e+01 3.864181e+01
#> 10.8 4.094017e+01 4.127945e+01 4.161873e+01 4.195801e+01 4.229729e+01
#> 10.9 4.459687e+01 4.498965e+01 4.538243e+01 4.577521e+01 4.616829e+01
#> 11   4.862758e+01 4.903523e+01 4.944288e+01 4.985053e+01 5.026403e+01
#> 11.1 5.304993e+01 5.346339e+01 5.387684e+01 5.429030e+01 5.471081e+01
#> 11.2 5.744503e+01 5.788606e+01 5.832709e+01 5.876812e+01 5.921487e+01
#> 11.3 6.200940e+01 6.247131e+01 6.293321e+01 6.339511e+01 6.387005e+01
#> 11.4 6.691268e+01 6.742095e+01 6.792921e+01 6.843747e+01 6.898256e+01
#> 11.5 7.245602e+01 7.292915e+01 7.340227e+01 7.387540e+01 7.438026e+01
#> 11.6 7.771105e+01 7.829015e+01 7.886924e+01 7.944833e+01 8.001383e+01
#> 11.7 8.337923e+01 8.396716e+01 8.455508e+01 8.514301e+01 8.575906e+01
#> 11.8 8.959507e+01 9.023858e+01 9.088209e+01 9.152559e+01 9.208646e+01
#> 11.9 9.552019e+01 9.624731e+01 9.697444e+01 9.770157e+01 9.837414e+01
#> 12   1.024179e+02 1.031527e+02 1.038875e+02 1.046224e+02 1.052871e+02
#> 12.1 1.092327e+02 1.099225e+02 1.106124e+02 1.113023e+02 1.119563e+02
#> 12.2 1.158760e+02 1.165430e+02 1.172101e+02 1.178771e+02 1.187608e+02
#> 12.3 1.236700e+02 1.243036e+02 1.249373e+02 1.255710e+02 1.263786e+02
#> 12.4 1.311935e+02 1.319763e+02 1.327592e+02 1.335406e+02 1.342892e+02
#> 12.5 1.390415e+02 1.399156e+02 1.407897e+02 1.416543e+02 1.424427e+02
#> 12.6 1.475536e+02 1.485196e+02 1.494856e+02 1.503874e+02 1.509936e+02
#> 12.7 1.555911e+02 1.566313e+02 1.576715e+02 1.586653e+02 1.595175e+02
#> 12.8 1.647611e+02 1.656705e+02 1.665799e+02 1.675443e+02 1.686284e+02
#> 12.9 1.743108e+02 1.750450e+02 1.757792e+02 1.766817e+02 1.778546e+02
#> 13   1.839744e+02 1.847678e+02 1.855611e+02 1.865322e+02 1.877190e+02
#> 13.1 1.936138e+02 1.943073e+02 1.950009e+02 1.959908e+02 1.972543e+02
#> 13.2 2.035118e+02 2.042572e+02 2.050026e+02 2.060530e+02 2.073168e+02
#> 13.3 2.147400e+02 2.159430e+02 2.171459e+02 2.181808e+02 2.191278e+02
#> 13.4 2.249359e+02 2.259297e+02 2.269235e+02 2.280900e+02 2.293220e+02
#> 13.5 2.367340e+02 2.379733e+02 2.392126e+02 2.401601e+02 2.410316e+02
#> 13.6 2.476805e+02 2.490542e+02 2.504278e+02 2.516314e+02 2.528077e+02
#> 13.7 2.597137e+02 2.608376e+02 2.619615e+02 2.633751e+02 2.648105e+02
#> 13.8 2.730147e+02 2.743124e+02 2.756101e+02 2.767410e+02 2.778717e+02
#> 13.9 2.852757e+02 2.866108e+02 2.879556e+02 2.894367e+02 2.909179e+02
#> 14   2.987731e+02 2.999215e+02 3.011310e+02 3.027329e+02 3.043348e+02
#> 14.1 3.122649e+02 3.133362e+02 3.144248e+02 3.155811e+02 3.167375e+02
#> 14.2 3.259704e+02 3.278356e+02 3.295252e+02 3.307433e+02 3.319615e+02
#> 14.3 3.406324e+02 3.422626e+02 3.438054e+02 3.451786e+02 3.465518e+02
#> 14.4 3.551398e+02 3.566164e+02 3.579933e+02 3.592255e+02 3.604577e+02
#> 14.5 3.698880e+02 3.717120e+02 3.732689e+02 3.745324e+02 3.757959e+02
#> 14.6 3.849627e+02 3.866780e+02 3.884421e+02 3.902471e+02 3.920521e+02
#> 14.7 4.018957e+02 4.034251e+02 4.049045e+02 4.063523e+02 4.078002e+02
#> 14.8 4.166221e+02 4.181070e+02 4.197358e+02 4.214321e+02 4.231284e+02
#> 14.9 4.341654e+02 4.360930e+02 4.381728e+02 4.403035e+02 4.424342e+02
#> 15   4.520665e+02 4.533642e+02 4.552033e+02 4.571630e+02 4.591226e+02
#> 15.1 4.704086e+02 4.722457e+02 4.736093e+02 4.749121e+02 4.762150e+02
#> 15.2 4.871933e+02 4.893025e+02 4.908399e+02 4.923502e+02 4.938604e+02
#> 15.3 5.073547e+02 5.099469e+02 5.112849e+02 5.126229e+02 5.139609e+02
#> 15.4 5.246082e+02 5.265969e+02 5.286357e+02 5.306744e+02 5.327132e+02
#> 15.5 5.424914e+02 5.440936e+02 5.465303e+02 5.489669e+02 5.514036e+02
#> 15.6 5.653294e+02 5.674589e+02 5.691084e+02 5.707579e+02 5.724075e+02
#> 15.7 5.864782e+02 5.889589e+02 5.910876e+02 5.932163e+02 5.953450e+02
#> 15.8 6.038049e+02 6.053048e+02 6.074447e+02 6.095845e+02 6.117244e+02
#> 15.9 6.281460e+02 6.304184e+02 6.317970e+02 6.331756e+02 6.345542e+02
#> 16   6.516322e+02 6.541257e+02 6.557663e+02 6.574068e+02 6.590474e+02
#> 16.1 6.711040e+02 6.731142e+02 6.750348e+02 6.769554e+02 6.788760e+02
#> 16.2 6.948709e+02 6.971424e+02 6.990614e+02 7.009804e+02 7.028994e+02
#> 16.3 7.167913e+02 7.192862e+02 7.218091e+02 7.243320e+02 7.268549e+02
#> 16.4 7.410583e+02 7.433204e+02 7.455639e+02 7.478073e+02 7.500508e+02
#> 16.5 7.606380e+02 7.628765e+02 7.652257e+02 7.675749e+02 7.699242e+02
#> 16.6 7.894574e+02 7.916252e+02 7.936663e+02 7.957074e+02 7.977484e+02
#> 16.7 8.131557e+02 8.153183e+02 8.174698e+02 8.196213e+02 8.217729e+02
#> 16.8 8.423172e+02 8.439051e+02 8.454931e+02 8.470811e+02 8.486691e+02
#> 16.9 8.637429e+02 8.668782e+02 8.700134e+02 8.731487e+02 8.762840e+02
#> 17   8.951756e+02 8.977993e+02 9.004231e+02 9.030468e+02 9.056705e+02
#> 17.1 9.210653e+02 9.234521e+02 9.258389e+02 9.282257e+02 9.306125e+02
#> 17.2 9.502683e+02 9.518087e+02 9.533490e+02 9.548893e+02 9.564297e+02
#> 17.3 9.785371e+02 9.807366e+02 9.829360e+02 9.851355e+02 9.873349e+02
#> 17.4 1.001216e+03 1.005539e+03 1.009862e+03 1.014185e+03 1.018507e+03
#> 17.5 1.030938e+03 1.035215e+03 1.039492e+03 1.043769e+03 1.048046e+03
#>              0.05         0.06         0.07         0.08         0.09
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 3.799342e-05 9.938042e-05
#> 7.7  1.363147e-03 1.766448e-03 2.169750e-03 2.953696e-03 3.972015e-03
#> 7.8  1.245091e-02 1.437380e-02 1.629668e-02 1.869699e-02 2.139125e-02
#> 7.9  4.272841e-02 4.744929e-02 5.241809e-02 5.738688e-02 6.318004e-02
#> 8    1.033968e-01 1.114724e-01 1.197227e-01 1.285702e-01 1.374177e-01
#> 8.1  2.017296e-01 2.138249e-01 2.265926e-01 2.393832e-01 2.525438e-01
#> 8.2  3.440182e-01 3.605240e-01 3.782343e-01 3.973210e-01 4.164078e-01
#> 8.3  5.403249e-01 5.636754e-01 5.873568e-01 6.120054e-01 6.366592e-01
#> 8.4  7.975763e-01 8.273748e-01 8.571734e-01 8.881944e-01 9.199490e-01
#> 8.5  1.121210e+00 1.159502e+00 1.197792e+00 1.236046e+00 1.274299e+00
#> 8.6  1.528729e+00 1.571454e+00 1.615867e+00 1.662704e+00 1.709541e+00
#> 8.7  2.009996e+00 2.060493e+00 2.113162e+00 2.166929e+00 2.220697e+00
#> 8.8  2.586847e+00 2.646915e+00 2.706983e+00 2.770824e+00 2.836282e+00
#> 8.9  3.252943e+00 3.323180e+00 3.397379e+00 3.472473e+00 3.547567e+00
#> 9    4.015357e+00 4.096475e+00 4.178803e+00 4.272026e+00 4.365249e+00
#> 9.1  4.905080e+00 4.998157e+00 5.094562e+00 5.195502e+00 5.296442e+00
#> 9.2  5.902992e+00 6.010006e+00 6.128804e+00 6.232504e+00 6.334526e+00
#> 9.3  7.025851e+00 7.143347e+00 7.264740e+00 7.395226e+00 7.525712e+00
#> 9.4  8.273665e+00 8.413235e+00 8.532835e+00 8.643706e+00 8.778697e+00
#> 9.5  9.655337e+00 9.794718e+00 9.934099e+00 1.008460e+01 1.024360e+01
#> 9.6  1.116651e+01 1.132872e+01 1.150295e+01 1.167717e+01 1.185140e+01
#> 9.7  1.285197e+01 1.303008e+01 1.320819e+01 1.338565e+01 1.356212e+01
#> 9.8  1.470045e+01 1.489996e+01 1.509833e+01 1.529669e+01 1.549506e+01
#> 9.9  1.665890e+01 1.687061e+01 1.708233e+01 1.729404e+01 1.750575e+01
#> 10   1.878517e+01 1.902398e+01 1.926280e+01 1.950161e+01 1.974042e+01
#> 10.1 2.114509e+01 2.138218e+01 2.161926e+01 2.185635e+01 2.209344e+01
#> 10.2 2.375072e+01 2.398277e+01 2.421482e+01 2.444687e+01 2.467892e+01
#> 10.3 2.635564e+01 2.664167e+01 2.692770e+01 2.721373e+01 2.749975e+01
#> 10.4 2.913737e+01 2.944768e+01 2.975799e+01 3.006829e+01 3.037860e+01
#> 10.5 3.219139e+01 3.254639e+01 3.290139e+01 3.325639e+01 3.361139e+01
#> 10.6 3.549798e+01 3.583776e+01 3.617753e+01 3.651730e+01 3.685708e+01
#> 10.7 3.902946e+01 3.942120e+01 3.981294e+01 4.020468e+01 4.059642e+01
#> 10.8 4.267695e+01 4.305878e+01 4.344061e+01 4.382244e+01 4.420427e+01
#> 10.9 4.657865e+01 4.698900e+01 4.739936e+01 4.780972e+01 4.821993e+01
#> 11   5.074002e+01 5.121600e+01 5.169199e+01 5.216798e+01 5.263648e+01
#> 11.1 5.517014e+01 5.562946e+01 5.608879e+01 5.654812e+01 5.700400e+01
#> 11.2 5.968164e+01 6.014842e+01 6.061519e+01 6.108197e+01 6.154750e+01
#> 11.3 6.437683e+01 6.488361e+01 6.539038e+01 6.589716e+01 6.640442e+01
#> 11.4 6.959345e+01 7.020435e+01 7.081524e+01 7.142613e+01 7.198289e+01
#> 11.5 7.492768e+01 7.547510e+01 7.602251e+01 7.656993e+01 7.713196e+01
#> 11.6 8.056549e+01 8.111714e+01 8.166879e+01 8.222045e+01 8.279131e+01
#> 11.7 8.639688e+01 8.703470e+01 8.767252e+01 8.831034e+01 8.895156e+01
#> 11.8 9.259922e+01 9.311198e+01 9.362473e+01 9.413749e+01 9.479306e+01
#> 11.9 9.902336e+01 9.967259e+01 1.003218e+02 1.009710e+02 1.016831e+02
#> 12   1.059309e+02 1.065746e+02 1.072183e+02 1.078621e+02 1.085428e+02
#> 12.1 1.126033e+02 1.132504e+02 1.138974e+02 1.145444e+02 1.152089e+02
#> 12.2 1.196671e+02 1.205735e+02 1.214798e+02 1.223861e+02 1.230363e+02
#> 12.3 1.271909e+02 1.280033e+02 1.288156e+02 1.296278e+02 1.304106e+02
#> 12.4 1.350379e+02 1.357865e+02 1.365352e+02 1.372934e+02 1.381675e+02
#> 12.5 1.432310e+02 1.440193e+02 1.448076e+02 1.456216e+02 1.465876e+02
#> 12.6 1.515998e+02 1.522060e+02 1.528123e+02 1.535108e+02 1.545510e+02
#> 12.7 1.603697e+02 1.612219e+02 1.620741e+02 1.629424e+02 1.638518e+02
#> 12.8 1.697124e+02 1.707964e+02 1.718805e+02 1.728423e+02 1.735765e+02
#> 12.9 1.790275e+02 1.802005e+02 1.813734e+02 1.823878e+02 1.831811e+02
#> 13   1.889059e+02 1.900927e+02 1.912796e+02 1.922267e+02 1.929203e+02
#> 13.1 1.985178e+02 1.997812e+02 2.010447e+02 2.020210e+02 2.027664e+02
#> 13.2 2.085806e+02 2.098444e+02 2.111082e+02 2.123341e+02 2.135371e+02
#> 13.3 2.200749e+02 2.210219e+02 2.219690e+02 2.229483e+02 2.239421e+02
#> 13.4 2.305540e+02 2.317859e+02 2.330179e+02 2.342554e+02 2.354947e+02
#> 13.5 2.419032e+02 2.427747e+02 2.436462e+02 2.449332e+02 2.463069e+02
#> 13.6 2.539840e+02 2.551602e+02 2.563365e+02 2.574659e+02 2.585898e+02
#> 13.7 2.662460e+02 2.676814e+02 2.691168e+02 2.704194e+02 2.717171e+02
#> 13.8 2.790024e+02 2.801331e+02 2.812704e+02 2.826055e+02 2.839406e+02
#> 13.9 2.923991e+02 2.938802e+02 2.953279e+02 2.964763e+02 2.976247e+02
#> 14   3.059368e+02 3.075387e+02 3.090509e+02 3.101222e+02 3.111936e+02
#> 14.1 3.178938e+02 3.190501e+02 3.203747e+02 3.222399e+02 3.241051e+02
#> 14.2 3.331796e+02 3.343978e+02 3.357419e+02 3.373721e+02 3.390022e+02
#> 14.3 3.479250e+02 3.492982e+02 3.507100e+02 3.521866e+02 3.536632e+02
#> 14.4 3.616899e+02 3.629221e+02 3.644160e+02 3.662400e+02 3.680640e+02
#> 14.5 3.770594e+02 3.783229e+02 3.798170e+02 3.815322e+02 3.832475e+02
#> 14.6 3.938570e+02 3.956620e+02 3.973075e+02 3.988369e+02 4.003663e+02
#> 14.7 4.092480e+02 4.106959e+02 4.121677e+02 4.136525e+02 4.151373e+02
#> 14.8 4.248246e+02 4.265209e+02 4.283826e+02 4.303102e+02 4.322378e+02
#> 14.9 4.445649e+02 4.466956e+02 4.481734e+02 4.494711e+02 4.507688e+02
#> 15   4.610822e+02 4.630419e+02 4.648971e+02 4.667343e+02 4.685714e+02
#> 15.1 4.775178e+02 4.788206e+02 4.808657e+02 4.829749e+02 4.850841e+02
#> 15.2 4.953706e+02 4.968808e+02 4.994899e+02 5.021115e+02 5.047331e+02
#> 15.3 5.152988e+02 5.166736e+02 5.186573e+02 5.206409e+02 5.226246e+02
#> 15.4 5.347519e+02 5.367161e+02 5.381600e+02 5.396038e+02 5.410476e+02
#> 15.5 5.538403e+02 5.562449e+02 5.585160e+02 5.607871e+02 5.630583e+02
#> 15.6 5.740570e+02 5.759630e+02 5.785918e+02 5.812206e+02 5.838494e+02
#> 15.7 5.974737e+02 5.992735e+02 6.004063e+02 6.015392e+02 6.026721e+02
#> 15.8 6.138643e+02 6.163288e+02 6.192831e+02 6.222374e+02 6.251917e+02
#> 15.9 6.359328e+02 6.382318e+02 6.415819e+02 6.449320e+02 6.482821e+02
#> 16   6.606879e+02 6.625897e+02 6.647183e+02 6.668468e+02 6.689754e+02
#> 16.1 6.807966e+02 6.833034e+02 6.861952e+02 6.890871e+02 6.919790e+02
#> 16.2 7.048184e+02 7.070794e+02 7.095074e+02 7.119353e+02 7.143633e+02
#> 16.3 7.293777e+02 7.317548e+02 7.340807e+02 7.364066e+02 7.387325e+02
#> 16.4 7.522943e+02 7.540546e+02 7.557004e+02 7.573463e+02 7.589921e+02
#> 16.5 7.722734e+02 7.756003e+02 7.790645e+02 7.825288e+02 7.859931e+02
#> 16.6 7.997895e+02 8.024347e+02 8.051149e+02 8.077952e+02 8.104755e+02
#> 16.7 8.239451e+02 8.276395e+02 8.313340e+02 8.350284e+02 8.387228e+02
#> 16.8 8.503454e+02 8.530141e+02 8.556828e+02 8.583515e+02 8.610202e+02
#> 16.9 8.794245e+02 8.825949e+02 8.857652e+02 8.889356e+02 8.921059e+02
#> 17   9.082816e+02 9.108474e+02 9.134132e+02 9.159790e+02 9.185448e+02
#> 17.1 9.333229e+02 9.368387e+02 9.403545e+02 9.438704e+02 9.473862e+02
#> 17.2 9.588712e+02 9.629507e+02 9.670302e+02 9.711097e+02 9.751892e+02
#> 17.3 9.895099e+02 9.916516e+02 9.937934e+02 9.959351e+02 9.980768e+02
#> 17.4 1.021496e+03 1.023103e+03 1.024711e+03 1.026319e+03 1.027927e+03
#> 17.5 1.051082e+03 1.053143e+03 1.055203e+03 1.057264e+03 1.059324e+03
#>  [ reached getOption("max.print") -- omitted 65 rows ]

# }