Predict method for discharge rating curves

Predict the discharge for given stage values based on a discharge rating curve result object.

# S3 method for plm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for plm
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm
predict(object, newdata = NULL, wide = FALSE, ...)

Arguments

object
an object of class "plm0", "plm", "gplm0" or "gplm".

newdata
a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.

wide
a logical value denoting whether to produce a wide prediction output.If TRUE, then the output is a table with median prediction values for an equally spaced grid of stages with 1 cm increments, each row containing predictions in a decimeter range of stages.

...
not used in this function

Value

an object of class "data.frame" with four columns, h (stage), lower (2.5% posterior predictive quantile), median (50% posterior predictive quantile), upper (97.5% posterior predictive quantile). If wide=T, a matrix as described above (see wide parameter) is returned.

Functions

predict.plm0: Predict method for plm0

predict.plm: Predict method for plm

predict.gplm0: Predict method for gplm0

predict.gplm: Predict method for gplm

See also

Examples

# \donttest{
data(krokfors)
plm0.fit <- plm0(Q~W,krokfors,h_max=24)
#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(plm0.fit,newdata=seq(23,24,by=0.01))

#>         h    lower   median     upper
#> 1   23.00 1533.284 3628.778  8656.417
#> 2   23.01 1537.515 3635.542  8677.002
#> 3   23.02 1537.047 3641.872  8686.906
#> 4   23.03 1536.579 3648.201  8696.810
#> 5   23.04 1536.111 3654.530  8706.714
#> 6   23.05 1535.642 3660.860  8716.618
#> 7   23.06 1537.509 3667.255  8719.806
#> 8   23.07 1543.693 3673.771  8710.576
#> 9   23.08 1549.876 3680.287  8701.347
#> 10  23.09 1556.060 3686.803  8692.117
#> 11  23.10 1562.244 3693.320  8682.888
#> 12  23.11 1566.590 3698.517  8692.571
#> 13  23.12 1568.005 3701.611  8732.438
#> 14  23.13 1569.419 3704.704  8772.304
#> 15  23.14 1570.833 3707.798  8812.171
#> 16  23.15 1572.248 3710.891  8852.038
#> 17  23.16 1573.323 3716.603  8887.243
#> 18  23.17 1573.928 3725.939  8915.992
#> 19  23.18 1574.534 3735.275  8944.742
#> 20  23.19 1575.139 3744.612  8973.491
#> 21  23.20 1575.745 3753.948  9002.241
#> 22  23.21 1578.058 3761.744  9014.405
#> 23  23.22 1582.430 3767.684  9006.586
#> 24  23.23 1586.802 3773.623  8998.767
#> 25  23.24 1591.174 3779.563  8990.948
#> 26  23.25 1595.546 3785.503  8983.129
#> 27  23.26 1598.767 3791.349  8993.164
#> 28  23.27 1600.780 3797.097  9021.957
#> 29  23.28 1602.793 3802.844  9050.749
#> 30  23.29 1604.805 3808.592  9079.542
#> 31  23.30 1606.818 3814.339  9108.334
#> 32  23.31 1607.574 3821.757  9137.206
#> 33  23.32 1607.177 3830.705  9166.150
#> 34  23.33 1606.781 3839.654  9195.095
#> 35  23.34 1606.385 3848.602  9224.039
#> 36  23.35 1605.989 3857.551  9252.983
#> 37  23.36 1607.597 3865.411  9282.937
#> 38  23.37 1610.806 3872.403  9313.697
#> 39  23.38 1614.015 3879.395  9344.458
#> 40  23.39 1617.225 3886.386  9375.218
#> 41  23.40 1620.434 3893.378  9405.978
#> 42  23.41 1625.265 3900.943  9425.274
#> 43  23.42 1631.223 3908.906  9436.607
#> 44  23.43 1637.181 3916.870  9447.941
#> 45  23.44 1643.139 3924.833  9459.274
#> 46  23.45 1649.097 3932.796  9470.608
#> 47  23.46 1651.257 3939.589  9464.689
#> 48  23.47 1651.132 3945.677  9448.389
#> 49  23.48 1651.007 3951.765  9432.088
#> 50  23.49 1650.882 3957.854  9415.788
#> 51  23.50 1650.757 3963.942  9399.487
#> 52  23.51 1650.484 3969.747  9442.113
#> 53  23.52 1650.133 3975.405  9515.305
#> 54  23.53 1649.782 3981.063  9588.498
#> 55  23.54 1649.431 3986.721  9661.690
#> 56  23.55 1649.080 3992.379  9734.882
#> 57  23.56 1654.978 4000.759  9732.042
#> 58  23.57 1663.650 4010.348  9695.457
#> 59  23.58 1672.322 4019.937  9658.873
#> 60  23.59 1680.993 4029.525  9622.288
#> 61  23.60 1689.665 4039.114  9585.703
#> 62  23.61 1692.236 4046.907  9603.912
#> 63  23.62 1692.514 4054.026  9642.720
#> 64  23.63 1692.791 4061.144  9681.529
#> 65  23.64 1693.069 4068.263  9720.338
#> 66  23.65 1693.346 4075.382  9759.147
#> 67  23.66 1697.456 4085.167  9767.665
#> 68  23.67 1702.769 4095.790  9766.667
#> 69  23.68 1708.083 4106.413  9765.668
#> 70  23.69 1713.396 4117.037  9764.670
#> 71  23.70 1718.710 4127.660  9763.671
#> 72  23.71 1721.576 4132.576  9807.910
#> 73  23.72 1723.812 4136.022  9863.808
#> 74  23.73 1726.048 4139.468  9919.706
#> 75  23.74 1728.284 4142.914  9975.604
#> 76  23.75 1730.520 4146.360 10031.503
#> 77  23.76 1733.448 4151.772 10043.548
#> 78  23.77 1736.518 4157.589 10046.561
#> 79  23.78 1739.589 4163.406 10049.575
#> 80  23.79 1742.660 4169.223 10052.588
#> 81  23.80 1745.731 4175.039 10055.602
#> 82  23.81 1747.602 4183.780 10091.380
#> 83  23.82 1749.283 4192.984 10132.342
#> 84  23.83 1750.965 4202.188 10173.305
#> 85  23.84 1752.646 4211.392 10214.267
#> 86  23.85 1754.327 4220.595 10255.229
#> 87  23.86 1762.332 4224.872 10260.028
#> 88  23.87 1771.059 4228.587 10260.698
#> 89  23.88 1779.785 4232.302 10261.368
#> 90  23.89 1788.512 4236.016 10262.038
#> 91  23.90 1797.238 4239.731 10262.708
#> 92  23.91 1793.533 4248.490 10288.469
#> 93  23.92 1788.916 4257.619 10316.069
#> 94  23.93 1784.299 4266.748 10343.670
#> 95  23.94 1779.682 4275.877 10371.271
#> 96  23.95 1775.065 4285.005 10398.872
#> 97  23.96 1778.024 4298.034 10415.556
#> 98  23.97 1781.252 4311.201 10431.855
#> 99  23.98 1784.479 4324.368 10448.153
#> 100 23.99 1787.707 4337.535 10464.452
#> 101 24.00 1790.935 4350.702 10480.750

predict(plm0.fit,wide=T)

#>                 0         0.01         0.02         0.03         0.04
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> 7.7  1.649676e-04 2.279584e-04 2.909492e-04 5.630567e-04 9.639232e-04
#> 7.8  4.995751e-03 6.017725e-03 7.039699e-03 8.605841e-03 1.050704e-02
#> 7.9  2.439662e-02 2.768124e-02 3.096586e-02 3.473193e-02 3.886172e-02
#> 8    6.898085e-02 7.489416e-02 8.185282e-02 8.881148e-02 9.591305e-02
#> 8.1  1.472219e-01 1.575197e-01 1.678175e-01 1.781154e-01 1.892163e-01
#> 8.2  2.671732e-01 2.819269e-01 2.971981e-01 3.127410e-01 3.282840e-01
#> 8.3  4.366637e-01 4.569690e-01 4.772669e-01 4.975537e-01 5.191287e-01
#> 8.4  6.609478e-01 6.871114e-01 7.138172e-01 7.419526e-01 7.700880e-01
#> 8.5  9.512578e-01 9.841295e-01 1.017477e+00 1.052324e+00 1.088412e+00
#> 8.6  1.313718e+00 1.354312e+00 1.395113e+00 1.435914e+00 1.479664e+00
#> 8.7  1.751575e+00 1.801692e+00 1.851809e+00 1.901926e+00 1.952701e+00
#> 8.8  2.281567e+00 2.339157e+00 2.396747e+00 2.455157e+00 2.519125e+00
#> 8.9  2.910458e+00 2.974328e+00 3.037177e+00 3.100027e+00 3.173954e+00
#> 9    3.615553e+00 3.695935e+00 3.773991e+00 3.852046e+00 3.930776e+00
#> 9.1  4.448165e+00 4.536407e+00 4.620878e+00 4.723041e+00 4.823630e+00
#> 9.2  5.376873e+00 5.486685e+00 5.595189e+00 5.691915e+00 5.789921e+00
#> 9.3  6.462087e+00 6.574017e+00 6.682245e+00 6.790474e+00 6.904755e+00
#> 9.4  7.634756e+00 7.763632e+00 7.892509e+00 8.021386e+00 8.148555e+00
#> 9.5  8.943329e+00 9.068615e+00 9.199272e+00 9.336069e+00 9.478898e+00
#> 9.6  1.037590e+01 1.052894e+01 1.068835e+01 1.085334e+01 1.101832e+01
#> 9.7  1.200135e+01 1.216205e+01 1.232275e+01 1.248344e+01 1.265067e+01
#> 9.8  1.374607e+01 1.393884e+01 1.412612e+01 1.430391e+01 1.448170e+01
#> 9.9  1.560028e+01 1.580904e+01 1.601781e+01 1.622658e+01 1.643534e+01
#> 10   1.767906e+01 1.791088e+01 1.814269e+01 1.837451e+01 1.860632e+01
#> 10.1 1.992448e+01 2.015863e+01 2.039279e+01 2.062694e+01 2.086110e+01
#> 10.2 2.236777e+01 2.263396e+01 2.290015e+01 2.316634e+01 2.343253e+01
#> 10.3 2.495281e+01 2.521972e+01 2.548663e+01 2.575354e+01 2.602045e+01
#> 10.4 2.777831e+01 2.804994e+01 2.832157e+01 2.859319e+01 2.886482e+01
#> 10.5 3.070830e+01 3.098773e+01 3.126716e+01 3.154659e+01 3.182603e+01
#> 10.6 3.390577e+01 3.421552e+01 3.452527e+01 3.483502e+01 3.514477e+01
#> 10.7 3.713724e+01 3.750952e+01 3.788181e+01 3.825409e+01 3.862638e+01
#> 10.8 4.082733e+01 4.126093e+01 4.169454e+01 4.212814e+01 4.256175e+01
#> 10.9 4.460546e+01 4.497747e+01 4.534948e+01 4.572149e+01 4.609507e+01
#> 11   4.875850e+01 4.911298e+01 4.946746e+01 4.982195e+01 5.018603e+01
#> 11.1 5.294952e+01 5.338342e+01 5.381732e+01 5.425123e+01 5.468723e+01
#> 11.2 5.737275e+01 5.782041e+01 5.826807e+01 5.871574e+01 5.915960e+01
#> 11.3 6.189215e+01 6.244140e+01 6.299065e+01 6.353991e+01 6.408383e+01
#> 11.4 6.722477e+01 6.772216e+01 6.821955e+01 6.871695e+01 6.921607e+01
#> 11.5 7.226138e+01 7.278658e+01 7.331178e+01 7.383698e+01 7.435125e+01
#> 11.6 7.753445e+01 7.816102e+01 7.878759e+01 7.941416e+01 8.001142e+01
#> 11.7 8.345266e+01 8.404413e+01 8.463559e+01 8.522706e+01 8.576304e+01
#> 11.8 8.912667e+01 8.987339e+01 9.062011e+01 9.136683e+01 9.199175e+01
#> 11.9 9.547264e+01 9.612077e+01 9.676890e+01 9.741702e+01 9.814459e+01
#> 12   1.023694e+02 1.029324e+02 1.034953e+02 1.040583e+02 1.047276e+02
#> 12.1 1.089442e+02 1.096504e+02 1.103566e+02 1.110629e+02 1.118189e+02
#> 12.2 1.162508e+02 1.169297e+02 1.176086e+02 1.182875e+02 1.190184e+02
#> 12.3 1.236794e+02 1.245410e+02 1.254027e+02 1.262644e+02 1.270279e+02
#> 12.4 1.313405e+02 1.319755e+02 1.326106e+02 1.332545e+02 1.341005e+02
#> 12.5 1.392661e+02 1.401552e+02 1.410442e+02 1.419317e+02 1.428071e+02
#> 12.6 1.477834e+02 1.485300e+02 1.492766e+02 1.500294e+02 1.508113e+02
#> 12.7 1.559286e+02 1.569030e+02 1.578775e+02 1.587963e+02 1.595453e+02
#> 12.8 1.649351e+02 1.660768e+02 1.672185e+02 1.682674e+02 1.691144e+02
#> 12.9 1.742412e+02 1.751074e+02 1.759735e+02 1.768470e+02 1.777325e+02
#> 13   1.835584e+02 1.846561e+02 1.857537e+02 1.868750e+02 1.880248e+02
#> 13.1 1.942047e+02 1.950653e+02 1.959259e+02 1.968514e+02 1.978368e+02
#> 13.2 2.041183e+02 2.052483e+02 2.063782e+02 2.074047e+02 2.083588e+02
#> 13.3 2.144156e+02 2.154964e+02 2.165772e+02 2.177986e+02 2.190935e+02
#> 13.4 2.258293e+02 2.267401e+02 2.276509e+02 2.286481e+02 2.296779e+02
#> 13.5 2.366636e+02 2.379857e+02 2.393078e+02 2.406188e+02 2.419268e+02
#> 13.6 2.478616e+02 2.484928e+02 2.491241e+02 2.503615e+02 2.516962e+02
#> 13.7 2.596182e+02 2.609232e+02 2.622281e+02 2.636196e+02 2.650176e+02
#> 13.8 2.730447e+02 2.743210e+02 2.755972e+02 2.768103e+02 2.780233e+02
#> 13.9 2.851075e+02 2.862566e+02 2.874242e+02 2.888514e+02 2.902786e+02
#> 14   2.974720e+02 2.984574e+02 2.995874e+02 3.016448e+02 3.037021e+02
#> 14.1 3.124490e+02 3.133712e+02 3.143706e+02 3.156722e+02 3.169739e+02
#> 14.2 3.258272e+02 3.274512e+02 3.290562e+02 3.306103e+02 3.321643e+02
#> 14.3 3.406094e+02 3.418975e+02 3.432389e+02 3.446840e+02 3.461292e+02
#> 14.4 3.552882e+02 3.568780e+02 3.584906e+02 3.601362e+02 3.617817e+02
#> 14.5 3.710423e+02 3.725098e+02 3.739652e+02 3.754071e+02 3.768490e+02
#> 14.6 3.857249e+02 3.872308e+02 3.888202e+02 3.904794e+02 3.921386e+02
#> 14.7 4.018782e+02 4.034771e+02 4.050120e+02 4.065064e+02 4.080008e+02
#> 14.8 4.174279e+02 4.190486e+02 4.208078e+02 4.226317e+02 4.244557e+02
#> 14.9 4.357842e+02 4.377117e+02 4.393773e+02 4.409554e+02 4.425335e+02
#> 15   4.519174e+02 4.534731e+02 4.551599e+02 4.568758e+02 4.585917e+02
#> 15.1 4.692868e+02 4.711064e+02 4.732407e+02 4.754154e+02 4.775901e+02
#> 15.2 4.878240e+02 4.892808e+02 4.912375e+02 4.932180e+02 4.951985e+02
#> 15.3 5.067046e+02 5.085825e+02 5.101123e+02 5.116420e+02 5.131717e+02
#> 15.4 5.246226e+02 5.267007e+02 5.286613e+02 5.306219e+02 5.325826e+02
#> 15.5 5.440904e+02 5.460770e+02 5.485277e+02 5.509783e+02 5.534290e+02
#> 15.6 5.650996e+02 5.669846e+02 5.694035e+02 5.718225e+02 5.742415e+02
#> 15.7 5.866807e+02 5.884293e+02 5.897416e+02 5.910538e+02 5.923661e+02
#> 15.8 6.037270e+02 6.060697e+02 6.088051e+02 6.115404e+02 6.142758e+02
#> 15.9 6.253853e+02 6.273040e+02 6.297324e+02 6.321608e+02 6.345893e+02
#> 16   6.471526e+02 6.492579e+02 6.514889e+02 6.537200e+02 6.559510e+02
#> 16.1 6.699440e+02 6.723164e+02 6.746946e+02 6.770728e+02 6.794510e+02
#> 16.2 6.924403e+02 6.952386e+02 6.984335e+02 7.016284e+02 7.048232e+02
#> 16.3 7.153728e+02 7.174352e+02 7.197944e+02 7.221536e+02 7.245127e+02
#> 16.4 7.379415e+02 7.406291e+02 7.434570e+02 7.462850e+02 7.491129e+02
#> 16.5 7.625656e+02 7.650613e+02 7.676316e+02 7.702018e+02 7.727721e+02
#> 16.6 7.873533e+02 7.896818e+02 7.920035e+02 7.943252e+02 7.966470e+02
#> 16.7 8.123362e+02 8.157613e+02 8.192022e+02 8.226431e+02 8.260840e+02
#> 16.8 8.373859e+02 8.405079e+02 8.436299e+02 8.467518e+02 8.498738e+02
#> 16.9 8.657438e+02 8.685099e+02 8.712760e+02 8.740421e+02 8.768082e+02
#> 17   8.927961e+02 8.963234e+02 8.998507e+02 9.033780e+02 9.069054e+02
#> 17.1 9.225110e+02 9.238707e+02 9.252305e+02 9.265903e+02 9.279501e+02
#> 17.2 9.532610e+02 9.548028e+02 9.563447e+02 9.578865e+02 9.594284e+02
#> 17.3 9.753852e+02 9.796565e+02 9.839279e+02 9.881992e+02 9.924706e+02
#> 17.4 1.007273e+03 1.009387e+03 1.011501e+03 1.013615e+03 1.015729e+03
#> 17.5 1.031960e+03 1.035756e+03 1.039552e+03 1.043348e+03 1.047144e+03
#> 17.6 1.063251e+03 1.065361e+03 1.067471e+03 1.069581e+03 1.071691e+03
#> 17.7 1.093356e+03 1.096630e+03 1.099905e+03 1.103180e+03 1.106455e+03
#> 17.8 1.126493e+03 1.129881e+03 1.133268e+03 1.136655e+03 1.140042e+03
#> 17.9 1.155643e+03 1.159568e+03 1.163494e+03 1.167419e+03 1.171344e+03
#> 18   1.191335e+03 1.193892e+03 1.196450e+03 1.199007e+03 1.201565e+03
#> 18.1 1.227522e+03 1.228296e+03 1.229070e+03 1.229844e+03 1.230618e+03
#> 18.2 1.253603e+03 1.257346e+03 1.261089e+03 1.264832e+03 1.268553e+03
#> 18.3 1.286879e+03 1.289400e+03 1.291921e+03 1.294443e+03 1.297191e+03
#> 18.4 1.326056e+03 1.331533e+03 1.337011e+03 1.342488e+03 1.347196e+03
#> 18.5 1.357121e+03 1.361073e+03 1.365025e+03 1.368976e+03 1.372966e+03
#> 18.6 1.396280e+03 1.399345e+03 1.402410e+03 1.405475e+03 1.408784e+03
#> 18.7 1.432958e+03 1.437612e+03 1.442265e+03 1.446918e+03 1.450744e+03
#> 18.8 1.466854e+03 1.470222e+03 1.473589e+03 1.476957e+03 1.480422e+03
#> 18.9 1.503859e+03 1.508746e+03 1.513632e+03 1.518519e+03 1.522623e+03
#> 19   1.546337e+03 1.551957e+03 1.557577e+03 1.563197e+03 1.567068e+03
#> 19.1 1.584587e+03 1.588299e+03 1.592011e+03 1.595722e+03 1.598902e+03
#> 19.2 1.625350e+03 1.633624e+03 1.641898e+03 1.650171e+03 1.652544e+03
#> 19.3 1.660693e+03 1.665116e+03 1.669540e+03 1.673963e+03 1.677658e+03
#> 19.4 1.704122e+03 1.710617e+03 1.717112e+03 1.723607e+03 1.726651e+03
#> 19.5 1.746800e+03 1.752084e+03 1.757368e+03 1.762652e+03 1.766197e+03
#> 19.6 1.787431e+03 1.791244e+03 1.795056e+03 1.798869e+03 1.804335e+03
#> 19.7 1.833967e+03 1.837855e+03 1.841742e+03 1.845565e+03 1.848414e+03
#> 19.8 1.870834e+03 1.876223e+03 1.881612e+03 1.886914e+03 1.891644e+03
#> 19.9 1.920735e+03 1.925794e+03 1.930854e+03 1.935603e+03 1.939103e+03
#> 20   1.963534e+03 1.968569e+03 1.973604e+03 1.978306e+03 1.982099e+03
#> 20.1 2.003704e+03 2.006997e+03 2.010290e+03 2.014555e+03 2.020741e+03
#> 20.2 2.050465e+03 2.053532e+03 2.056600e+03 2.060950e+03 2.067192e+03
#> 20.3 2.100921e+03 2.105636e+03 2.110351e+03 2.115848e+03 2.122217e+03
#> 20.4 2.145206e+03 2.145501e+03 2.145795e+03 2.149573e+03 2.156307e+03
#> 20.5 2.195217e+03 2.201370e+03 2.207523e+03 2.212913e+03 2.217814e+03
#> 20.6 2.250892e+03 2.257183e+03 2.263475e+03 2.267398e+03 2.270194e+03
#> 20.7 2.289287e+03 2.292937e+03 2.296587e+03 2.302694e+03 2.309637e+03
#> 20.8 2.344503e+03 2.349002e+03 2.353502e+03 2.357494e+03 2.361371e+03
#> 20.9 2.393108e+03 2.399961e+03 2.406814e+03 2.412857e+03 2.418791e+03
#> 21   2.442904e+03 2.444898e+03 2.446892e+03 2.453523e+03 2.460394e+03
#> 21.1 2.492674e+03 2.496548e+03 2.500501e+03 2.508541e+03 2.516580e+03
#> 21.2 2.549093e+03 2.551982e+03 2.555274e+03 2.562770e+03 2.570267e+03
#> 21.3 2.596603e+03 2.598127e+03 2.601020e+03 2.611342e+03 2.621665e+03
#> 21.4 2.654420e+03 2.655595e+03 2.657879e+03 2.664009e+03 2.670139e+03
#> 21.5 2.712024e+03 2.719721e+03 2.726871e+03 2.732698e+03 2.738525e+03
#> 21.6 2.767787e+03 2.771901e+03 2.778229e+03 2.788483e+03 2.798736e+03
#> 21.7 2.828973e+03 2.830011e+03 2.834673e+03 2.844160e+03 2.853647e+03
#> 21.8 2.891181e+03 2.895069e+03 2.899389e+03 2.904146e+03 2.908903e+03
#> 21.9 2.938792e+03 2.943931e+03 2.948950e+03 2.953876e+03 2.958803e+03
#> 22   2.989968e+03 2.995341e+03 3.000561e+03 3.005693e+03 3.010824e+03
#> 22.1 3.063341e+03 3.074396e+03 3.080835e+03 3.085315e+03 3.089795e+03
#> 22.2 3.111120e+03 3.114113e+03 3.121031e+03 3.129119e+03 3.137206e+03
#> 22.3 3.169042e+03 3.172743e+03 3.177874e+03 3.183281e+03 3.188688e+03
#> 22.4 3.232817e+03 3.241243e+03 3.248335e+03 3.255290e+03 3.262245e+03
#> 22.5 3.296332e+03 3.301347e+03 3.308721e+03 3.316154e+03 3.323587e+03
#> 22.6 3.365826e+03 3.372753e+03 3.381462e+03 3.390172e+03 3.398882e+03
#> 22.7 3.430511e+03 3.434813e+03 3.444289e+03 3.453765e+03 3.463241e+03
#> 22.8 3.491129e+03 3.494893e+03 3.504453e+03 3.514014e+03 3.523574e+03
#> 22.9 3.575956e+03 3.583088e+03 3.586449e+03 3.589809e+03 3.593170e+03
#> 23   3.628778e+03 3.635542e+03 3.641872e+03 3.648201e+03 3.654530e+03
#> 23.1 3.693320e+03 3.698517e+03 3.701611e+03 3.704704e+03 3.707798e+03
#> 23.2 3.753948e+03 3.761744e+03 3.767684e+03 3.773623e+03 3.779563e+03
#> 23.3 3.814339e+03 3.821757e+03 3.830705e+03 3.839654e+03 3.848602e+03
#> 23.4 3.893378e+03 3.900943e+03 3.908906e+03 3.916870e+03 3.924833e+03
#> 23.5 3.963942e+03 3.969747e+03 3.975405e+03 3.981063e+03 3.986721e+03
#> 23.6 4.039114e+03 4.046907e+03 4.054026e+03 4.061144e+03 4.068263e+03
#> 23.7 4.127660e+03 4.132576e+03 4.136022e+03 4.139468e+03 4.142914e+03
#> 23.8 4.175039e+03 4.183780e+03 4.192984e+03 4.202188e+03 4.211392e+03
#> 23.9 4.239731e+03 4.248490e+03 4.257619e+03 4.266748e+03 4.275877e+03
#> 24   4.350702e+03           NA           NA           NA           NA
#>              0.05         0.06         0.07         0.08         0.09
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 3.898603e-05 1.019768e-04
#> 7.7  1.364790e-03 1.765656e-03 2.166522e-03 2.951803e-03 3.973777e-03
#> 7.8  1.240824e-02 1.430945e-02 1.621065e-02 1.864990e-02 2.142044e-02
#> 7.9  4.299151e-02 4.767139e-02 5.254525e-02 5.741911e-02 6.315672e-02
#> 8    1.038094e-01 1.117058e-01 1.198503e-01 1.288437e-01 1.378371e-01
#> 8.1  2.008039e-01 2.123916e-01 2.255374e-01 2.387365e-01 2.524195e-01
#> 8.2  3.450276e-01 3.617711e-01 3.793610e-01 3.979180e-01 4.164749e-01
#> 8.3  5.415625e-01 5.634392e-01 5.851767e-01 6.099089e-01 6.347842e-01
#> 8.4  7.990072e-01 8.288843e-01 8.587613e-01 8.893206e-01 9.202892e-01
#> 8.5  1.124261e+00 1.159539e+00 1.194995e+00 1.234569e+00 1.274144e+00
#> 8.6  1.525215e+00 1.570765e+00 1.616021e+00 1.660853e+00 1.705685e+00
#> 8.7  2.003525e+00 2.054349e+00 2.109490e+00 2.166814e+00 2.224137e+00
#> 8.8  2.583094e+00 2.647063e+00 2.711031e+00 2.776950e+00 2.843704e+00
#> 8.9  3.251059e+00 3.328164e+00 3.400574e+00 3.471924e+00 3.543274e+00
#> 9    4.015582e+00 4.100387e+00 4.185315e+00 4.271342e+00 4.357370e+00
#> 9.1  4.907261e+00 4.990891e+00 5.079533e+00 5.175006e+00 5.270478e+00
#> 9.2  5.890137e+00 5.994682e+00 6.101734e+00 6.220950e+00 6.341518e+00
#> 9.3  7.025088e+00 7.145421e+00 7.266424e+00 7.388991e+00 7.511558e+00
#> 9.4  8.275534e+00 8.402513e+00 8.542108e+00 8.686992e+00 8.815738e+00
#> 9.5  9.625749e+00 9.772600e+00 9.919452e+00 1.006981e+01 1.022285e+01
#> 9.6  1.118331e+01 1.134769e+01 1.151146e+01 1.167524e+01 1.183901e+01
#> 9.7  1.282644e+01 1.300220e+01 1.317796e+01 1.336053e+01 1.355330e+01
#> 9.8  1.465948e+01 1.484079e+01 1.502913e+01 1.521747e+01 1.540581e+01
#> 9.9  1.664294e+01 1.684821e+01 1.705348e+01 1.725875e+01 1.746402e+01
#> 10   1.883216e+01 1.904912e+01 1.926608e+01 1.948305e+01 1.970001e+01
#> 10.1 2.110336e+01 2.135475e+01 2.160613e+01 2.185752e+01 2.210890e+01
#> 10.2 2.369034e+01 2.394097e+01 2.419159e+01 2.444222e+01 2.469285e+01
#> 10.3 2.630612e+01 2.660392e+01 2.690172e+01 2.719951e+01 2.749731e+01
#> 10.4 2.916477e+01 2.947832e+01 2.979187e+01 3.010542e+01 3.041896e+01
#> 10.5 3.216233e+01 3.251819e+01 3.287406e+01 3.322993e+01 3.358579e+01
#> 10.6 3.546777e+01 3.579384e+01 3.611991e+01 3.644598e+01 3.677205e+01
#> 10.7 3.898291e+01 3.933730e+01 3.969169e+01 4.004608e+01 4.040047e+01
#> 10.8 4.290031e+01 4.323375e+01 4.356720e+01 4.390065e+01 4.423410e+01
#> 10.9 4.655798e+01 4.702088e+01 4.748378e+01 4.794668e+01 4.840401e+01
#> 11   5.065273e+01 5.111943e+01 5.158614e+01 5.205284e+01 5.251561e+01
#> 11.1 5.513480e+01 5.558237e+01 5.602993e+01 5.647750e+01 5.692508e+01
#> 11.2 5.959018e+01 6.002075e+01 6.045132e+01 6.088190e+01 6.134289e+01
#> 11.3 6.461471e+01 6.514560e+01 6.567648e+01 6.620737e+01 6.672738e+01
#> 11.4 6.971828e+01 7.022050e+01 7.072272e+01 7.122493e+01 7.173618e+01
#> 11.5 7.485086e+01 7.535047e+01 7.585009e+01 7.634970e+01 7.690788e+01
#> 11.6 8.057882e+01 8.114623e+01 8.171364e+01 8.228105e+01 8.286120e+01
#> 11.7 8.625609e+01 8.674914e+01 8.724218e+01 8.773523e+01 8.837995e+01
#> 11.8 9.254576e+01 9.309977e+01 9.365379e+01 9.420780e+01 9.482451e+01
#> 11.9 9.890614e+01 9.966769e+01 1.004292e+02 1.011908e+02 1.018065e+02
#> 12   1.054289e+02 1.061301e+02 1.068314e+02 1.075327e+02 1.082380e+02
#> 12.1 1.125846e+02 1.133503e+02 1.141161e+02 1.148818e+02 1.155719e+02
#> 12.2 1.197547e+02 1.204910e+02 1.212273e+02 1.219636e+02 1.228177e+02
#> 12.3 1.277888e+02 1.285497e+02 1.293105e+02 1.300704e+02 1.307055e+02
#> 12.4 1.349466e+02 1.357927e+02 1.366387e+02 1.374881e+02 1.383771e+02
#> 12.5 1.436826e+02 1.445580e+02 1.454334e+02 1.462902e+02 1.470368e+02
#> 12.6 1.515932e+02 1.523750e+02 1.531569e+02 1.539797e+02 1.549542e+02
#> 12.7 1.602943e+02 1.610432e+02 1.617922e+02 1.626516e+02 1.637933e+02
#> 12.8 1.699613e+02 1.708083e+02 1.716553e+02 1.725090e+02 1.733751e+02
#> 12.9 1.786180e+02 1.795035e+02 1.803890e+02 1.813630e+02 1.824607e+02
#> 13   1.891746e+02 1.903244e+02 1.914742e+02 1.924834e+02 1.933441e+02
#> 13.1 1.988222e+02 1.998075e+02 2.007929e+02 2.018584e+02 2.029884e+02
#> 13.2 2.093129e+02 2.102670e+02 2.112211e+02 2.122541e+02 2.133348e+02
#> 13.3 2.203883e+02 2.216832e+02 2.229781e+02 2.240076e+02 2.249184e+02
#> 13.4 2.307078e+02 2.317377e+02 2.327676e+02 2.340193e+02 2.353414e+02
#> 13.5 2.432349e+02 2.445430e+02 2.458510e+02 2.465991e+02 2.472303e+02
#> 13.6 2.530309e+02 2.543656e+02 2.557003e+02 2.570084e+02 2.583133e+02
#> 13.7 2.664155e+02 2.678135e+02 2.692115e+02 2.704922e+02 2.717684e+02
#> 13.8 2.792363e+02 2.804494e+02 2.816603e+02 2.828094e+02 2.839585e+02
#> 13.9 2.917058e+02 2.931330e+02 2.945157e+02 2.955011e+02 2.964866e+02
#> 14   3.057595e+02 3.078168e+02 3.096823e+02 3.106046e+02 3.115268e+02
#> 14.1 3.182756e+02 3.195773e+02 3.209554e+02 3.225794e+02 3.242033e+02
#> 14.2 3.337184e+02 3.352725e+02 3.367453e+02 3.380333e+02 3.393214e+02
#> 14.3 3.475743e+02 3.490194e+02 3.505186e+02 3.521085e+02 3.536983e+02
#> 14.4 3.634273e+02 3.650728e+02 3.666396e+02 3.681072e+02 3.695747e+02
#> 14.5 3.782909e+02 3.797329e+02 3.812074e+02 3.827133e+02 3.842191e+02
#> 14.6 3.937978e+02 3.954570e+02 3.970813e+02 3.986803e+02 4.002792e+02
#> 14.7 4.094952e+02 4.109896e+02 4.125658e+02 4.141865e+02 4.158072e+02
#> 14.8 4.262796e+02 4.281036e+02 4.300016e+02 4.319291e+02 4.338566e+02
#> 14.9 4.441116e+02 4.456897e+02 4.472502e+02 4.488059e+02 4.503617e+02
#> 15   4.603077e+02 4.620236e+02 4.638279e+02 4.656475e+02 4.674671e+02
#> 15.1 4.797648e+02 4.819395e+02 4.834534e+02 4.849103e+02 4.863671e+02
#> 15.2 4.971789e+02 4.991594e+02 5.010465e+02 5.029326e+02 5.048186e+02
#> 15.3 5.147014e+02 5.162631e+02 5.183530e+02 5.204429e+02 5.225327e+02
#> 15.4 5.345432e+02 5.364960e+02 5.383946e+02 5.402932e+02 5.421918e+02
#> 15.5 5.558796e+02 5.581902e+02 5.599175e+02 5.616449e+02 5.633723e+02
#> 15.6 5.766605e+02 5.789520e+02 5.808841e+02 5.828163e+02 5.847485e+02
#> 15.7 5.936783e+02 5.952566e+02 5.973742e+02 5.994918e+02 6.016094e+02
#> 15.8 6.170111e+02 6.192660e+02 6.207958e+02 6.223256e+02 6.238554e+02
#> 15.9 6.370177e+02 6.392364e+02 6.412154e+02 6.431945e+02 6.451735e+02
#> 16   6.581820e+02 6.604846e+02 6.628495e+02 6.652143e+02 6.675791e+02
#> 16.1 6.818292e+02 6.840396e+02 6.861398e+02 6.882400e+02 6.903401e+02
#> 16.2 7.080181e+02 7.099735e+02 7.113233e+02 7.126731e+02 7.140229e+02
#> 16.3 7.268719e+02 7.291177e+02 7.313236e+02 7.335296e+02 7.357355e+02
#> 16.4 7.519409e+02 7.541778e+02 7.562748e+02 7.583717e+02 7.604686e+02
#> 16.5 7.753424e+02 7.777615e+02 7.801595e+02 7.825574e+02 7.849553e+02
#> 16.6 7.989687e+02 8.016266e+02 8.043040e+02 8.069814e+02 8.096588e+02
#> 16.7 8.294998e+02 8.310622e+02 8.326246e+02 8.341870e+02 8.357494e+02
#> 16.8 8.529494e+02 8.555034e+02 8.580574e+02 8.606113e+02 8.631653e+02
#> 16.9 8.795517e+02 8.821670e+02 8.847823e+02 8.873976e+02 8.900129e+02
#> 17   9.102124e+02 9.127306e+02 9.152488e+02 9.177670e+02 9.202852e+02
#> 17.1 9.302965e+02 9.350985e+02 9.399006e+02 9.447026e+02 9.495047e+02
#> 17.2 9.613736e+02 9.640519e+02 9.667303e+02 9.694086e+02 9.720870e+02
#> 17.3 9.959034e+02 9.981935e+02 1.000484e+03 1.002774e+03 1.005064e+03
#> 17.4 1.018107e+03 1.020757e+03 1.023407e+03 1.026057e+03 1.028707e+03
#> 17.5 1.050301e+03 1.052956e+03 1.055610e+03 1.058265e+03 1.060919e+03
#> 17.6 1.074843e+03 1.078610e+03 1.082378e+03 1.086146e+03 1.089914e+03
#> 17.7 1.109773e+03 1.113110e+03 1.116446e+03 1.119783e+03 1.123120e+03
#> 17.8 1.142636e+03 1.144985e+03 1.147335e+03 1.149685e+03 1.152034e+03
#> 17.9 1.174871e+03 1.178318e+03 1.181765e+03 1.185212e+03 1.188659e+03
#> 18   1.206352e+03 1.211382e+03 1.216413e+03 1.221444e+03 1.226474e+03
#> 18.1 1.234391e+03 1.238259e+03 1.242126e+03 1.245993e+03 1.249859e+03
#> 18.2 1.271723e+03 1.274893e+03 1.278064e+03 1.281234e+03 1.284357e+03
#> 18.3 1.301845e+03 1.306500e+03 1.311154e+03 1.315808e+03 1.320578e+03
#> 18.4 1.348271e+03 1.349345e+03 1.350420e+03 1.351494e+03 1.353170e+03
#> 18.5 1.377074e+03 1.381181e+03 1.385289e+03 1.389397e+03 1.393216e+03
#> 18.6 1.412633e+03 1.416481e+03 1.420330e+03 1.424178e+03 1.428305e+03
#> 18.7 1.453219e+03 1.455693e+03 1.458167e+03 1.460642e+03 1.463486e+03
#> 18.8 1.484007e+03 1.487591e+03 1.491176e+03 1.494760e+03 1.498972e+03
#> 18.9 1.525995e+03 1.529366e+03 1.532737e+03 1.536109e+03 1.540718e+03
#> 19   1.569695e+03 1.572323e+03 1.574950e+03 1.577578e+03 1.580876e+03
#> 19.1 1.601798e+03 1.604694e+03 1.607589e+03 1.610485e+03 1.617076e+03
#> 19.2 1.652635e+03 1.652725e+03 1.652816e+03 1.652906e+03 1.656270e+03
#> 19.3 1.681159e+03 1.684659e+03 1.688160e+03 1.691661e+03 1.697627e+03
#> 19.4 1.729122e+03 1.731593e+03 1.734065e+03 1.736536e+03 1.741516e+03
#> 19.5 1.769603e+03 1.773009e+03 1.776416e+03 1.779822e+03 1.783618e+03
#> 19.6 1.809811e+03 1.815286e+03 1.820762e+03 1.826192e+03 1.830080e+03
#> 19.7 1.851263e+03 1.854112e+03 1.856961e+03 1.860056e+03 1.865445e+03
#> 19.8 1.896373e+03 1.901102e+03 1.905832e+03 1.910616e+03 1.915675e+03
#> 19.9 1.942604e+03 1.946105e+03 1.949606e+03 1.953465e+03 1.958500e+03
#> 20   1.985891e+03 1.989683e+03 1.993476e+03 1.997118e+03 2.000411e+03
#> 20.1 2.026927e+03 2.033113e+03 2.039299e+03 2.044331e+03 2.047398e+03
#> 20.2 2.073434e+03 2.079676e+03 2.085918e+03 2.091491e+03 2.096206e+03
#> 20.3 2.128587e+03 2.134956e+03 2.141326e+03 2.144617e+03 2.144912e+03
#> 20.4 2.163042e+03 2.169776e+03 2.176511e+03 2.182911e+03 2.189064e+03
#> 20.5 2.222714e+03 2.227614e+03 2.232515e+03 2.238310e+03 2.244601e+03
#> 20.6 2.272991e+03 2.275787e+03 2.278583e+03 2.281987e+03 2.285637e+03
#> 20.7 2.316581e+03 2.323524e+03 2.330468e+03 2.335505e+03 2.340004e+03
#> 20.8 2.365247e+03 2.369124e+03 2.373001e+03 2.379402e+03 2.386255e+03
#> 20.9 2.424725e+03 2.430659e+03 2.436593e+03 2.438916e+03 2.440910e+03
#> 21   2.467265e+03 2.474136e+03 2.481007e+03 2.484927e+03 2.488801e+03
#> 21.1 2.524620e+03 2.532660e+03 2.540425e+03 2.543314e+03 2.546204e+03
#> 21.2 2.577763e+03 2.585260e+03 2.592031e+03 2.593555e+03 2.595079e+03
#> 21.3 2.631987e+03 2.642309e+03 2.650896e+03 2.652070e+03 2.653245e+03
#> 21.4 2.676269e+03 2.682399e+03 2.688933e+03 2.696630e+03 2.704327e+03
#> 21.5 2.744351e+03 2.750178e+03 2.755446e+03 2.759560e+03 2.763674e+03
#> 21.6 2.808990e+03 2.819244e+03 2.825860e+03 2.826898e+03 2.827935e+03
#> 21.7 2.863134e+03 2.872622e+03 2.879516e+03 2.883405e+03 2.887293e+03
#> 21.8 2.913659e+03 2.918416e+03 2.923375e+03 2.928514e+03 2.933653e+03
#> 21.9 2.963729e+03 2.968656e+03 2.973850e+03 2.979223e+03 2.984595e+03
#> 22   3.015956e+03 3.021088e+03 3.030176e+03 3.041231e+03 3.052286e+03
#> 22.1 3.094275e+03 3.098756e+03 3.102141e+03 3.105134e+03 3.108127e+03
#> 22.2 3.145294e+03 3.153382e+03 3.157940e+03 3.161640e+03 3.165341e+03
#> 22.3 3.194094e+03 3.199501e+03 3.207542e+03 3.215967e+03 3.224392e+03
#> 22.4 3.269201e+03 3.276156e+03 3.281286e+03 3.286301e+03 3.291317e+03
#> 22.5 3.331020e+03 3.338447e+03 3.345292e+03 3.352137e+03 3.358982e+03
#> 22.6 3.407592e+03 3.415908e+03 3.419559e+03 3.423209e+03 3.426860e+03
#> 22.7 3.472717e+03 3.481172e+03 3.483662e+03 3.486151e+03 3.488640e+03
#> 22.8 3.533135e+03 3.542442e+03 3.550820e+03 3.559199e+03 3.567578e+03
#> 22.9 3.596531e+03 3.600911e+03 3.607878e+03 3.614845e+03 3.621811e+03
#> 23   3.660860e+03 3.667255e+03 3.673771e+03 3.680287e+03 3.686803e+03
#> 23.1 3.710891e+03 3.716603e+03 3.725939e+03 3.735275e+03 3.744612e+03
#> 23.2 3.785503e+03 3.791349e+03 3.797097e+03 3.802844e+03 3.808592e+03
#> 23.3 3.857551e+03 3.865411e+03 3.872403e+03 3.879395e+03 3.886386e+03
#> 23.4 3.932796e+03 3.939589e+03 3.945677e+03 3.951765e+03 3.957854e+03
#> 23.5 3.992379e+03 4.000759e+03 4.010348e+03 4.019937e+03 4.029525e+03
#> 23.6 4.075382e+03 4.085167e+03 4.095790e+03 4.106413e+03 4.117037e+03
#> 23.7 4.146360e+03 4.151772e+03 4.157589e+03 4.163406e+03 4.169223e+03
#> 23.8 4.220595e+03 4.224872e+03 4.228587e+03 4.232302e+03 4.236016e+03
#> 23.9 4.285005e+03 4.298034e+03 4.311201e+03 4.324368e+03 4.337535e+03
#> 24             NA           NA           NA           NA           NA

# }