Predict method for discharge rating curves

Predict the discharge for given stage values based on a discharge rating curve result object.

# S3 method for plm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for plm
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm
predict(object, newdata = NULL, wide = FALSE, ...)

Arguments

object
an object of class "plm0", "plm", "gplm0" or "gplm".

newdata
a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.

wide
a logical value denoting whether to produce a wide prediction output.If TRUE, then the output is a table with median prediction values for an equally spaced grid of stages with 1 cm increments, each row containing predictions in a decimeter range of stages.

...
not used in this function

Value

an object of class "data.frame" with four columns, h (stage), lower (2.5% posterior predictive quantile), median (50% posterior predictive quantile), upper (97.5% posterior predictive quantile). If wide=T, a matrix as described above (see wide parameter) is returned.

Functions

predict.plm0: Predict method for plm0

predict.plm: Predict method for plm

predict.gplm0: Predict method for gplm0

predict.gplm: Predict method for gplm

See also

Examples

# \donttest{
data(krokfors)
plm0.fit <- plm0(Q~W,krokfors,h_max=9)

#> Error in plm0.inference(y = log(Q), h = h, c_param, h_max, parallel, forcepoint): maximum stage value must be larger than the maximum stage value in the data, which is 9.897 m

#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(plm0.fit,newdata=seq(8,9,by=0.01))

#> Error in predict(plm0.fit, newdata = seq(8, 9, by = 0.01)): object 'plm0.fit' not found

predict(plm0.fit,wide=T)

#> Error in predict(plm0.fit, wide = T): object 'plm0.fit' not found

# }