Predict method for discharge rating curves

Predict the discharge for given stage values based on a discharge rating curve result object.

# S3 method for plm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for plm
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm
predict(object, newdata = NULL, wide = FALSE, ...)

Arguments

object
an object of class "plm0", "plm", "gplm0" or "gplm".

newdata
a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.

wide
a logical value denoting whether to produce a wide prediction output.If TRUE, then the output is a table with median prediction values for an equally spaced grid of stages with 1 cm increments, each row containing predictions in a decimeter range of stages.

...
not used in this function

Value

an object of class "data.frame" with four columns, h (stage), lower (2.5% posterior predictive quantile), median (50% posterior predictive quantile), upper (97.5% posterior predictive quantile). If wide=TRUE, a matrix as described above (see wide parameter) is returned.

Functions

predict.plm0: Predict method for plm0

predict.plm: Predict method for plm

predict.gplm0: Predict method for gplm0

predict.gplm: Predict method for gplm

See also

Examples

# \donttest{
data(krokfors)
plm0.fit <- plm0(Q~W,krokfors,h_max=24)
#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(plm0.fit,newdata=seq(23,24,by=0.01))

#>         h    lower   median     upper
#> 1   23.00 1511.523 3620.138  8797.530
#> 2   23.01 1514.006 3627.023  8816.192
#> 3   23.02 1522.204 3634.812  8809.027
#> 4   23.03 1530.402 3642.602  8801.862
#> 5   23.04 1538.600 3650.391  8794.698
#> 6   23.05 1546.798 3658.180  8787.533
#> 7   23.06 1553.304 3665.657  8793.912
#> 8   23.07 1556.683 3672.557  8825.327
#> 9   23.08 1560.063 3679.456  8856.741
#> 10  23.09 1563.442 3686.356  8888.156
#> 11  23.10 1566.821 3693.255  8919.571
#> 12  23.11 1568.885 3699.494  8940.843
#> 13  23.12 1568.850 3704.678  8945.927
#> 14  23.13 1568.814 3709.862  8951.011
#> 15  23.14 1568.779 3715.045  8956.095
#> 16  23.15 1568.744 3720.229  8961.179
#> 17  23.16 1570.411 3725.185  8972.572
#> 18  23.17 1574.435 3729.825  8992.700
#> 19  23.18 1578.458 3734.465  9012.829
#> 20  23.19 1582.482 3739.105  9032.957
#> 21  23.20 1586.506 3743.745  9053.086
#> 22  23.21 1589.313 3750.065  9074.542
#> 23  23.22 1590.656 3758.409  9097.598
#> 24  23.23 1591.999 3766.754  9120.654
#> 25  23.24 1593.342 3775.099  9143.710
#> 26  23.25 1594.684 3783.443  9166.766
#> 27  23.26 1595.976 3789.953  9190.311
#> 28  23.27 1597.213 3794.537  9214.371
#> 29  23.28 1598.450 3799.120  9238.431
#> 30  23.29 1599.688 3803.703  9262.491
#> 31  23.30 1600.925 3808.286  9286.551
#> 32  23.31 1604.006 3815.832  9312.301
#> 33  23.32 1608.776 3826.093  9339.599
#> 34  23.33 1613.546 3836.355  9366.898
#> 35  23.34 1618.315 3846.616  9394.196
#> 36  23.35 1623.085 3856.877  9421.495
#> 37  23.36 1626.468 3865.566  9457.551
#> 38  23.37 1628.745 3873.001  9500.601
#> 39  23.38 1631.021 3880.435  9543.652
#> 40  23.39 1633.297 3887.869  9586.702
#> 41  23.40 1635.573 3895.303  9629.753
#> 42  23.41 1641.809 3898.251  9646.052
#> 43  23.42 1650.796 3898.084  9643.771
#> 44  23.43 1659.784 3897.916  9641.491
#> 45  23.44 1668.771 3897.749  9639.210
#> 46  23.45 1677.758 3897.582  9636.930
#> 47  23.46 1677.113 3906.736  9642.551
#> 48  23.47 1670.672 3921.502  9652.928
#> 49  23.48 1664.231 3936.267  9663.304
#> 50  23.49 1657.790 3951.032  9673.681
#> 51  23.50 1651.348 3965.797  9684.058
#> 52  23.51 1648.196 3974.651  9707.608
#> 53  23.52 1646.750 3980.439  9737.992
#> 54  23.53 1645.304 3986.227  9768.376
#> 55  23.54 1643.858 3992.015  9798.759
#> 56  23.55 1642.412 3997.803  9829.143
#> 57  23.56 1648.443 4003.983  9818.568
#> 58  23.57 1657.792 4010.337  9789.815
#> 59  23.58 1667.142 4016.691  9761.061
#> 60  23.59 1676.491 4023.045  9732.308
#> 61  23.60 1685.841 4029.399  9703.554
#> 62  23.61 1688.826 4035.197  9730.968
#> 63  23.62 1689.419 4040.786  9779.498
#> 64  23.63 1690.012 4046.375  9828.029
#> 65  23.64 1690.604 4051.964  9876.560
#> 66  23.65 1691.197 4057.553  9925.090
#> 67  23.66 1696.237 4063.744  9945.393
#> 68  23.67 1702.675 4070.124  9956.826
#> 69  23.68 1709.113 4076.504  9968.258
#> 70  23.69 1715.551 4082.885  9979.691
#> 71  23.70 1721.988 4089.265  9991.124
#> 72  23.71 1721.794 4098.406 10011.880
#> 73  23.72 1719.891 4108.259 10035.039
#> 74  23.73 1717.988 4118.112 10058.198
#> 75  23.74 1716.085 4127.965 10081.357
#> 76  23.75 1714.181 4137.818 10104.516
#> 77  23.76 1718.056 4147.389 10141.363
#> 78  23.77 1723.121 4156.901 10181.029
#> 79  23.78 1728.186 4166.414 10220.696
#> 80  23.79 1733.250 4175.927 10260.362
#> 81  23.80 1738.315 4185.439 10300.028
#> 82  23.81 1742.193 4188.138 10327.531
#> 83  23.82 1745.884 4189.759 10353.109
#> 84  23.83 1749.574 4191.379 10378.687
#> 85  23.84 1753.264 4193.000 10404.264
#> 86  23.85 1756.955 4194.620 10429.842
#> 87  23.86 1753.411 4202.399 10457.683
#> 88  23.87 1749.041 4210.880 10485.782
#> 89  23.88 1744.672 4219.361 10513.882
#> 90  23.89 1740.302 4227.843 10541.981
#> 91  23.90 1735.933 4236.324 10570.081
#> 92  23.91 1748.067 4246.419 10569.375
#> 93  23.92 1761.412 4256.633 10566.557
#> 94  23.93 1774.757 4266.847 10563.739
#> 95  23.94 1788.102 4277.061 10560.921
#> 96  23.95 1801.447 4287.275 10558.103
#> 97  23.96 1801.517 4295.416 10582.502
#> 98  23.97 1801.118 4303.483 10607.863
#> 99  23.98 1800.719 4311.551 10633.223
#> 100 23.99 1800.320 4319.619 10658.584
#> 101 24.00 1799.920 4327.686 10683.945

predict(plm0.fit,wide=TRUE)

#>                 0         0.01         0.02         0.03         0.04
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> 7.7  1.587822e-04 2.194112e-04 2.800402e-04 5.472108e-04 9.415547e-04
#> 7.8  4.961837e-03 5.987824e-03 7.013810e-03 8.589662e-03 1.050408e-02
#> 7.9  2.428749e-02 2.762342e-02 3.095936e-02 3.470428e-02 3.875818e-02
#> 8    6.903890e-02 7.489737e-02 8.165469e-02 8.841201e-02 9.535033e-02
#> 8.1  1.473080e-01 1.576305e-01 1.679530e-01 1.782755e-01 1.895193e-01
#> 8.2  2.665456e-01 2.808941e-01 2.956729e-01 3.106777e-01 3.256825e-01
#> 8.3  4.357261e-01 4.545353e-01 4.750670e-01 4.981825e-01 5.200521e-01
#> 8.4  6.614879e-01 6.877813e-01 7.145921e-01 7.427670e-01 7.709419e-01
#> 8.5  9.515370e-01 9.854823e-01 1.020018e+00 1.055717e+00 1.092377e+00
#> 8.6  1.313787e+00 1.355642e+00 1.398193e+00 1.440745e+00 1.484950e+00
#> 8.7  1.756216e+00 1.805782e+00 1.855349e+00 1.904916e+00 1.956803e+00
#> 8.8  2.283542e+00 2.341193e+00 2.398843e+00 2.457153e+00 2.519932e+00
#> 8.9  2.902249e+00 2.969348e+00 3.037236e+00 3.105124e+00 3.179628e+00
#> 9    3.628815e+00 3.698778e+00 3.781225e+00 3.863671e+00 3.945818e+00
#> 9.1  4.445494e+00 4.532165e+00 4.613978e+00 4.713930e+00 4.813024e+00
#> 9.2  5.384908e+00 5.488893e+00 5.593331e+00 5.701859e+00 5.807631e+00
#> 9.3  6.466258e+00 6.577995e+00 6.687639e+00 6.797284e+00 6.908865e+00
#> 9.4  7.638563e+00 7.769972e+00 7.901381e+00 8.032789e+00 8.160688e+00
#> 9.5  8.939207e+00 9.079909e+00 9.226573e+00 9.380049e+00 9.529222e+00
#> 9.6  1.042456e+01 1.057769e+01 1.073091e+01 1.088421e+01 1.103750e+01
#> 9.7  1.198833e+01 1.215833e+01 1.232834e+01 1.249834e+01 1.266688e+01
#> 9.8  1.374814e+01 1.394873e+01 1.414433e+01 1.433130e+01 1.451826e+01
#> 9.9  1.557105e+01 1.579451e+01 1.601797e+01 1.624142e+01 1.646488e+01
#> 10   1.772083e+01 1.794703e+01 1.817322e+01 1.839942e+01 1.862562e+01
#> 10.1 1.993650e+01 2.018252e+01 2.042854e+01 2.067456e+01 2.092058e+01
#> 10.2 2.239021e+01 2.264483e+01 2.289946e+01 2.315408e+01 2.340871e+01
#> 10.3 2.490676e+01 2.520133e+01 2.549590e+01 2.579047e+01 2.608504e+01
#> 10.4 2.774879e+01 2.804329e+01 2.833779e+01 2.863229e+01 2.892680e+01
#> 10.5 3.070401e+01 3.100746e+01 3.131091e+01 3.161435e+01 3.191780e+01
#> 10.6 3.394414e+01 3.423671e+01 3.452929e+01 3.482187e+01 3.511445e+01
#> 10.7 3.732922e+01 3.765166e+01 3.797411e+01 3.829655e+01 3.861900e+01
#> 10.8 4.077401e+01 4.115934e+01 4.154467e+01 4.193000e+01 4.231533e+01
#> 10.9 4.469384e+01 4.511837e+01 4.554289e+01 4.596742e+01 4.639072e+01
#> 11   4.856218e+01 4.896449e+01 4.936680e+01 4.976912e+01 5.017651e+01
#> 11.1 5.291987e+01 5.335762e+01 5.379537e+01 5.423312e+01 5.466832e+01
#> 11.2 5.727639e+01 5.776573e+01 5.825508e+01 5.874442e+01 5.922476e+01
#> 11.3 6.196858e+01 6.245801e+01 6.294744e+01 6.343686e+01 6.392448e+01
#> 11.4 6.690868e+01 6.745604e+01 6.800339e+01 6.855075e+01 6.910038e+01
#> 11.5 7.228996e+01 7.274847e+01 7.320697e+01 7.366547e+01 7.422381e+01
#> 11.6 7.806579e+01 7.854484e+01 7.902390e+01 7.950295e+01 8.004397e+01
#> 11.7 8.361889e+01 8.419040e+01 8.476191e+01 8.533342e+01 8.587347e+01
#> 11.8 8.920464e+01 8.986863e+01 9.053262e+01 9.119661e+01 9.188262e+01
#> 11.9 9.584258e+01 9.640152e+01 9.696047e+01 9.751942e+01 9.820775e+01
#> 12   1.024934e+02 1.031354e+02 1.037773e+02 1.044193e+02 1.050040e+02
#> 12.1 1.087477e+02 1.095029e+02 1.102582e+02 1.110134e+02 1.117421e+02
#> 12.2 1.160422e+02 1.167437e+02 1.174453e+02 1.181469e+02 1.189039e+02
#> 12.3 1.235545e+02 1.243554e+02 1.251563e+02 1.259573e+02 1.266294e+02
#> 12.4 1.310297e+02 1.318916e+02 1.327536e+02 1.336087e+02 1.343061e+02
#> 12.5 1.389269e+02 1.398345e+02 1.407421e+02 1.416431e+02 1.424916e+02
#> 12.6 1.472839e+02 1.479932e+02 1.487025e+02 1.494494e+02 1.503696e+02
#> 12.7 1.559978e+02 1.569664e+02 1.579350e+02 1.588655e+02 1.596798e+02
#> 12.8 1.650247e+02 1.660404e+02 1.670561e+02 1.679982e+02 1.687803e+02
#> 12.9 1.740564e+02 1.750868e+02 1.761173e+02 1.771086e+02 1.780372e+02
#> 13   1.837551e+02 1.847443e+02 1.857335e+02 1.867581e+02 1.878255e+02
#> 13.1 1.940240e+02 1.950085e+02 1.959931e+02 1.969158e+02 1.977816e+02
#> 13.2 2.033370e+02 2.043440e+02 2.053510e+02 2.065569e+02 2.079018e+02
#> 13.3 2.147458e+02 2.156233e+02 2.165009e+02 2.174846e+02 2.185237e+02
#> 13.4 2.248866e+02 2.259735e+02 2.270603e+02 2.283158e+02 2.296354e+02
#> 13.5 2.371065e+02 2.382644e+02 2.394222e+02 2.405978e+02 2.417781e+02
#> 13.6 2.485366e+02 2.496026e+02 2.506686e+02 2.516650e+02 2.526502e+02
#> 13.7 2.591736e+02 2.603702e+02 2.615668e+02 2.632749e+02 2.650214e+02
#> 13.8 2.724147e+02 2.731201e+02 2.738254e+02 2.754336e+02 2.770433e+02
#> 13.9 2.850374e+02 2.860983e+02 2.871827e+02 2.885966e+02 2.900105e+02
#> 14   2.993411e+02 3.010282e+02 3.026457e+02 3.038170e+02 3.049882e+02
#> 14.1 3.128571e+02 3.142938e+02 3.156371e+02 3.166134e+02 3.175897e+02
#> 14.2 3.258778e+02 3.276048e+02 3.292545e+02 3.306967e+02 3.321389e+02
#> 14.3 3.396778e+02 3.407828e+02 3.420941e+02 3.438059e+02 3.455178e+02
#> 14.4 3.556147e+02 3.572749e+02 3.587107e+02 3.598211e+02 3.609315e+02
#> 14.5 3.711607e+02 3.733073e+02 3.749163e+02 3.759345e+02 3.769526e+02
#> 14.6 3.850047e+02 3.865764e+02 3.883099e+02 3.901786e+02 3.920474e+02
#> 14.7 4.014958e+02 4.028717e+02 4.044757e+02 4.062237e+02 4.079718e+02
#> 14.8 4.174583e+02 4.189317e+02 4.203288e+02 4.216900e+02 4.230513e+02
#> 14.9 4.347597e+02 4.370739e+02 4.387586e+02 4.402331e+02 4.417076e+02
#> 15   4.520292e+02 4.538934e+02 4.553198e+02 4.566488e+02 4.579777e+02
#> 15.1 4.685200e+02 4.705158e+02 4.724113e+02 4.742939e+02 4.761766e+02
#> 15.2 4.863750e+02 4.879777e+02 4.900318e+02 4.921074e+02 4.941830e+02
#> 15.3 5.055445e+02 5.073423e+02 5.089676e+02 5.105929e+02 5.122182e+02
#> 15.4 5.233671e+02 5.253541e+02 5.275139e+02 5.296738e+02 5.318336e+02
#> 15.5 5.430965e+02 5.449374e+02 5.472639e+02 5.495905e+02 5.519170e+02
#> 15.6 5.675350e+02 5.700349e+02 5.717817e+02 5.735284e+02 5.752752e+02
#> 15.7 5.845714e+02 5.862370e+02 5.883701e+02 5.905033e+02 5.926364e+02
#> 15.8 6.056497e+02 6.078698e+02 6.101552e+02 6.124406e+02 6.147260e+02
#> 15.9 6.261127e+02 6.278821e+02 6.296682e+02 6.314544e+02 6.332405e+02
#> 16   6.465934e+02 6.488604e+02 6.510186e+02 6.531768e+02 6.553349e+02
#> 16.1 6.700023e+02 6.721461e+02 6.739927e+02 6.758392e+02 6.776857e+02
#> 16.2 6.930809e+02 6.952744e+02 6.971323e+02 6.989902e+02 7.008481e+02
#> 16.3 7.171708e+02 7.199628e+02 7.226826e+02 7.254024e+02 7.281222e+02
#> 16.4 7.390513e+02 7.417911e+02 7.448682e+02 7.479453e+02 7.510224e+02
#> 16.5 7.634148e+02 7.659106e+02 7.685337e+02 7.711567e+02 7.737798e+02
#> 16.6 7.860001e+02 7.886638e+02 7.914021e+02 7.941404e+02 7.968786e+02
#> 16.7 8.166891e+02 8.175741e+02 8.184054e+02 8.192367e+02 8.200679e+02
#> 16.8 8.436313e+02 8.453889e+02 8.471464e+02 8.489040e+02 8.506616e+02
#> 16.9 8.613641e+02 8.646080e+02 8.678519e+02 8.710957e+02 8.743396e+02
#> 17   8.891867e+02 8.925234e+02 8.958601e+02 8.991968e+02 9.025335e+02
#> 17.1 9.235264e+02 9.253523e+02 9.271783e+02 9.290042e+02 9.308301e+02
#> 17.2 9.448163e+02 9.478475e+02 9.508786e+02 9.539098e+02 9.569409e+02
#> 17.3 9.749873e+02 9.781891e+02 9.813909e+02 9.845928e+02 9.877946e+02
#> 17.4 1.005690e+03 1.007934e+03 1.010178e+03 1.012421e+03 1.014665e+03
#> 17.5 1.034916e+03 1.038180e+03 1.041444e+03 1.044708e+03 1.047972e+03
#> 17.6 1.065066e+03 1.067737e+03 1.070408e+03 1.073078e+03 1.075749e+03
#> 17.7 1.095246e+03 1.098421e+03 1.101597e+03 1.104772e+03 1.107948e+03
#> 17.8 1.122897e+03 1.127261e+03 1.131625e+03 1.135988e+03 1.140352e+03
#> 17.9 1.152486e+03 1.157121e+03 1.161757e+03 1.166392e+03 1.171027e+03
#> 18   1.188496e+03 1.191172e+03 1.193848e+03 1.196523e+03 1.199199e+03
#> 18.1 1.219577e+03 1.222308e+03 1.225040e+03 1.227771e+03 1.230503e+03
#> 18.2 1.248214e+03 1.253152e+03 1.258090e+03 1.263028e+03 1.267930e+03
#> 18.3 1.291692e+03 1.295504e+03 1.299316e+03 1.303129e+03 1.306878e+03
#> 18.4 1.325898e+03 1.328856e+03 1.331815e+03 1.334773e+03 1.337817e+03
#> 18.5 1.359342e+03 1.363481e+03 1.367621e+03 1.371761e+03 1.375782e+03
#> 18.6 1.396857e+03 1.399855e+03 1.402853e+03 1.405851e+03 1.408635e+03
#> 18.7 1.425878e+03 1.430699e+03 1.435520e+03 1.440342e+03 1.444373e+03
#> 18.8 1.462175e+03 1.465879e+03 1.469584e+03 1.473289e+03 1.477140e+03
#> 18.9 1.502034e+03 1.506535e+03 1.511036e+03 1.515537e+03 1.519392e+03
#> 19   1.540239e+03 1.544354e+03 1.548470e+03 1.552586e+03 1.556426e+03
#> 19.1 1.578281e+03 1.581922e+03 1.585564e+03 1.589205e+03 1.593199e+03
#> 19.2 1.620262e+03 1.625615e+03 1.630967e+03 1.636319e+03 1.640516e+03
#> 19.3 1.660207e+03 1.662357e+03 1.664508e+03 1.666659e+03 1.670524e+03
#> 19.4 1.694215e+03 1.697307e+03 1.700399e+03 1.703491e+03 1.709254e+03
#> 19.5 1.742517e+03 1.746624e+03 1.750730e+03 1.754837e+03 1.757747e+03
#> 19.6 1.780544e+03 1.786375e+03 1.792205e+03 1.798036e+03 1.802122e+03
#> 19.7 1.820963e+03 1.822272e+03 1.823581e+03 1.825215e+03 1.831722e+03
#> 19.8 1.865381e+03 1.869321e+03 1.873261e+03 1.877541e+03 1.884076e+03
#> 19.9 1.912505e+03 1.914062e+03 1.915619e+03 1.918208e+03 1.924942e+03
#> 20   1.961850e+03 1.967019e+03 1.972187e+03 1.976694e+03 1.979391e+03
#> 20.1 2.000534e+03 2.005387e+03 2.010240e+03 2.014780e+03 2.018700e+03
#> 20.2 2.047329e+03 2.053404e+03 2.059479e+03 2.065393e+03 2.071070e+03
#> 20.3 2.096546e+03 2.098702e+03 2.100857e+03 2.104930e+03 2.111143e+03
#> 20.4 2.138564e+03 2.140845e+03 2.143127e+03 2.147730e+03 2.154304e+03
#> 20.5 2.193896e+03 2.200527e+03 2.207158e+03 2.211934e+03 2.215521e+03
#> 20.6 2.243691e+03 2.249793e+03 2.255894e+03 2.259447e+03 2.261785e+03
#> 20.7 2.287669e+03 2.294378e+03 2.301087e+03 2.306619e+03 2.311748e+03
#> 20.8 2.331970e+03 2.333302e+03 2.334634e+03 2.339802e+03 2.345845e+03
#> 20.9 2.382177e+03 2.388246e+03 2.394314e+03 2.401653e+03 2.409161e+03
#> 21   2.441812e+03 2.445070e+03 2.448327e+03 2.456462e+03 2.464850e+03
#> 21.1 2.485731e+03 2.484253e+03 2.482985e+03 2.492476e+03 2.501967e+03
#> 21.2 2.547862e+03 2.553734e+03 2.559561e+03 2.564918e+03 2.570274e+03
#> 21.3 2.602234e+03 2.607532e+03 2.612878e+03 2.618477e+03 2.624076e+03
#> 21.4 2.647419e+03 2.649804e+03 2.653339e+03 2.660856e+03 2.668374e+03
#> 21.5 2.710317e+03 2.716864e+03 2.722323e+03 2.725146e+03 2.727970e+03
#> 21.6 2.765818e+03 2.774926e+03 2.781147e+03 2.782247e+03 2.783348e+03
#> 21.7 2.818609e+03 2.828151e+03 2.836499e+03 2.843256e+03 2.850013e+03
#> 21.8 2.871880e+03 2.873244e+03 2.877663e+03 2.885173e+03 2.892683e+03
#> 21.9 2.932035e+03 2.937929e+03 2.944524e+03 2.951658e+03 2.958792e+03
#> 22   2.987931e+03 2.991269e+03 2.996628e+03 3.003153e+03 3.009679e+03
#> 22.1 3.048601e+03 3.055064e+03 3.062500e+03 3.070349e+03 3.078199e+03
#> 22.2 3.109443e+03 3.113050e+03 3.119348e+03 3.126449e+03 3.133549e+03
#> 22.3 3.165044e+03 3.169225e+03 3.174554e+03 3.180103e+03 3.185652e+03
#> 22.4 3.220219e+03 3.226096e+03 3.232396e+03 3.238739e+03 3.245082e+03
#> 22.5 3.297649e+03 3.307673e+03 3.312716e+03 3.317634e+03 3.322551e+03
#> 22.6 3.359722e+03 3.366546e+03 3.373257e+03 3.379967e+03 3.386678e+03
#> 22.7 3.412332e+03 3.415978e+03 3.423726e+03 3.431475e+03 3.439224e+03
#> 22.8 3.486570e+03 3.494652e+03 3.503311e+03 3.511970e+03 3.520630e+03
#> 22.9 3.563445e+03 3.569530e+03 3.574389e+03 3.579247e+03 3.584106e+03
#> 23   3.620138e+03 3.627023e+03 3.634812e+03 3.642602e+03 3.650391e+03
#> 23.1 3.693255e+03 3.699494e+03 3.704678e+03 3.709862e+03 3.715045e+03
#> 23.2 3.743745e+03 3.750065e+03 3.758409e+03 3.766754e+03 3.775099e+03
#> 23.3 3.808286e+03 3.815832e+03 3.826093e+03 3.836355e+03 3.846616e+03
#> 23.4 3.895303e+03 3.898251e+03 3.898084e+03 3.897916e+03 3.897749e+03
#> 23.5 3.965797e+03 3.974651e+03 3.980439e+03 3.986227e+03 3.992015e+03
#> 23.6 4.029399e+03 4.035197e+03 4.040786e+03 4.046375e+03 4.051964e+03
#> 23.7 4.089265e+03 4.098406e+03 4.108259e+03 4.118112e+03 4.127965e+03
#> 23.8 4.185439e+03 4.188138e+03 4.189759e+03 4.191379e+03 4.193000e+03
#> 23.9 4.236324e+03 4.246419e+03 4.256633e+03 4.266847e+03 4.277061e+03
#> 24   4.327686e+03           NA           NA           NA           NA
#>              0.05         0.06         0.07         0.08         0.09
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 3.752427e-05 9.815326e-05
#> 7.7  1.335899e-03 1.730242e-03 2.124586e-03 2.909864e-03 3.935851e-03
#> 7.8  1.241850e-02 1.433292e-02 1.624734e-02 1.864382e-02 2.133713e-02
#> 7.9  4.281208e-02 4.756049e-02 5.255380e-02 5.754711e-02 6.325711e-02
#> 8    1.032953e-01 1.112403e-01 1.194616e-01 1.286281e-01 1.377946e-01
#> 8.1  2.013215e-01 2.131237e-01 2.259724e-01 2.388569e-01 2.521971e-01
#> 8.2  3.439615e-01 3.622404e-01 3.804964e-01 3.987260e-01 4.169556e-01
#> 8.3  5.410912e-01 5.628596e-01 5.848103e-01 6.099444e-01 6.351945e-01
#> 8.4  7.995307e-01 8.286253e-01 8.577199e-01 8.883701e-01 9.199535e-01
#> 8.5  1.128528e+00 1.163457e+00 1.198531e+00 1.236950e+00 1.275369e+00
#> 8.6  1.530164e+00 1.575379e+00 1.620479e+00 1.665416e+00 1.710353e+00
#> 8.7  2.008859e+00 2.060915e+00 2.115448e+00 2.171233e+00 2.227018e+00
#> 8.8  2.582711e+00 2.645490e+00 2.708269e+00 2.772511e+00 2.837380e+00
#> 8.9  3.256029e+00 3.332431e+00 3.406980e+00 3.481110e+00 3.555240e+00
#> 9    4.025272e+00 4.104725e+00 4.184760e+00 4.270021e+00 4.355282e+00
#> 9.1  4.902874e+00 4.992724e+00 5.086075e+00 5.184198e+00 5.282321e+00
#> 9.2  5.908644e+00 6.012321e+00 6.117541e+00 6.233020e+00 6.349639e+00
#> 9.3  7.022382e+00 7.135900e+00 7.253805e+00 7.381949e+00 7.510093e+00
#> 9.4  8.288197e+00 8.415706e+00 8.541218e+00 8.667432e+00 8.802357e+00
#> 9.5  9.675527e+00 9.821832e+00 9.968137e+00 1.011831e+01 1.027144e+01
#> 9.6  1.119080e+01 1.134698e+01 1.150604e+01 1.166510e+01 1.182416e+01
#> 9.7  1.283351e+01 1.300013e+01 1.316675e+01 1.334696e+01 1.354755e+01
#> 9.8  1.470523e+01 1.488559e+01 1.505273e+01 1.521987e+01 1.538701e+01
#> 9.9  1.668256e+01 1.688874e+01 1.709492e+01 1.730110e+01 1.750728e+01
#> 10   1.884736e+01 1.906249e+01 1.927762e+01 1.949275e+01 1.970788e+01
#> 10.1 2.116558e+01 2.140941e+01 2.165325e+01 2.189709e+01 2.214092e+01
#> 10.2 2.365762e+01 2.390166e+01 2.414569e+01 2.438973e+01 2.463376e+01
#> 10.3 2.636694e+01 2.664064e+01 2.691434e+01 2.718804e+01 2.746174e+01
#> 10.4 2.922182e+01 2.951710e+01 2.981238e+01 3.010766e+01 3.040293e+01
#> 10.5 3.225333e+01 3.259989e+01 3.294645e+01 3.329302e+01 3.363958e+01
#> 10.6 3.547802e+01 3.585800e+01 3.623798e+01 3.661796e+01 3.699794e+01
#> 10.7 3.897032e+01 3.932555e+01 3.968078e+01 4.003602e+01 4.039125e+01
#> 10.8 4.270601e+01 4.309698e+01 4.348795e+01 4.387891e+01 4.426988e+01
#> 10.9 4.674404e+01 4.709737e+01 4.745069e+01 4.780402e+01 4.815986e+01
#> 11   5.063821e+01 5.109990e+01 5.156160e+01 5.202329e+01 5.248212e+01
#> 11.1 5.508950e+01 5.551068e+01 5.593187e+01 5.635305e+01 5.678705e+01
#> 11.2 5.967355e+01 6.012235e+01 6.057115e+01 6.101994e+01 6.147915e+01
#> 11.3 6.440768e+01 6.489089e+01 6.537409e+01 6.585729e+01 6.636132e+01
#> 11.4 6.965408e+01 7.020778e+01 7.076147e+01 7.131517e+01 7.183146e+01
#> 11.5 7.491607e+01 7.560832e+01 7.630058e+01 7.699283e+01 7.758674e+01
#> 11.6 8.064811e+01 8.125224e+01 8.185638e+01 8.246052e+01 8.304738e+01
#> 11.7 8.638918e+01 8.690488e+01 8.742059e+01 8.793630e+01 8.854066e+01
#> 11.8 9.258146e+01 9.328030e+01 9.397914e+01 9.467798e+01 9.528363e+01
#> 11.9 9.895144e+01 9.969513e+01 1.004388e+02 1.011825e+02 1.018515e+02
#> 12   1.055716e+02 1.061391e+02 1.067066e+02 1.072742e+02 1.079924e+02
#> 12.1 1.124656e+02 1.131891e+02 1.139126e+02 1.146362e+02 1.153406e+02
#> 12.2 1.196666e+02 1.204294e+02 1.211921e+02 1.219549e+02 1.227535e+02
#> 12.3 1.272981e+02 1.279668e+02 1.286355e+02 1.293057e+02 1.301677e+02
#> 12.4 1.350035e+02 1.357009e+02 1.363984e+02 1.371118e+02 1.380193e+02
#> 12.5 1.433400e+02 1.441885e+02 1.450369e+02 1.458653e+02 1.465746e+02
#> 12.6 1.512898e+02 1.522100e+02 1.531301e+02 1.540606e+02 1.550292e+02
#> 12.7 1.604940e+02 1.613082e+02 1.621225e+02 1.629933e+02 1.640090e+02
#> 12.8 1.695625e+02 1.703446e+02 1.711267e+02 1.719956e+02 1.730260e+02
#> 12.9 1.789657e+02 1.798943e+02 1.808229e+02 1.817767e+02 1.827659e+02
#> 13   1.888929e+02 1.899603e+02 1.910278e+02 1.920549e+02 1.930394e+02
#> 13.1 1.986474e+02 1.995132e+02 2.003789e+02 2.013230e+02 2.023300e+02
#> 13.2 2.092467e+02 2.105917e+02 2.119366e+02 2.129906e+02 2.138682e+02
#> 13.3 2.195628e+02 2.206018e+02 2.216409e+02 2.227130e+02 2.237998e+02
#> 13.4 2.309550e+02 2.322745e+02 2.335941e+02 2.347909e+02 2.359487e+02
#> 13.5 2.429584e+02 2.441386e+02 2.453189e+02 2.464046e+02 2.474706e+02
#> 13.6 2.536353e+02 2.546205e+02 2.556057e+02 2.567803e+02 2.579769e+02
#> 13.7 2.667680e+02 2.685146e+02 2.702612e+02 2.710039e+02 2.717093e+02
#> 13.8 2.786530e+02 2.802627e+02 2.818547e+02 2.829156e+02 2.839765e+02
#> 13.9 2.914244e+02 2.928383e+02 2.942797e+02 2.959668e+02 2.976540e+02
#> 14   3.061595e+02 3.073307e+02 3.085468e+02 3.099836e+02 3.114203e+02
#> 14.1 3.185660e+02 3.195423e+02 3.206968e+02 3.224238e+02 3.241508e+02
#> 14.2 3.335812e+02 3.350234e+02 3.363626e+02 3.374676e+02 3.385727e+02
#> 14.3 3.472297e+02 3.489415e+02 3.506341e+02 3.522943e+02 3.539545e+02
#> 14.4 3.620419e+02 3.631523e+02 3.647209e+02 3.668675e+02 3.690141e+02
#> 14.5 3.779708e+02 3.789890e+02 3.802898e+02 3.818614e+02 3.834331e+02
#> 14.6 3.939162e+02 3.957850e+02 3.973684e+02 3.987442e+02 4.001200e+02
#> 14.7 4.097198e+02 4.114679e+02 4.130382e+02 4.145115e+02 4.159849e+02
#> 14.8 4.244126e+02 4.257739e+02 4.278170e+02 4.301312e+02 4.324455e+02
#> 14.9 4.431820e+02 4.446565e+02 4.464365e+02 4.483007e+02 4.501649e+02
#> 15   4.593067e+02 4.606357e+02 4.625328e+02 4.645286e+02 4.665243e+02
#> 15.1 4.780592e+02 4.799419e+02 4.815669e+02 4.831696e+02 4.847723e+02
#> 15.2 4.962586e+02 4.983342e+02 5.001391e+02 5.019409e+02 5.037427e+02
#> 15.3 5.138435e+02 5.154885e+02 5.174581e+02 5.194278e+02 5.213974e+02
#> 15.4 5.339934e+02 5.361017e+02 5.378504e+02 5.395991e+02 5.413478e+02
#> 15.5 5.542435e+02 5.566467e+02 5.593688e+02 5.620909e+02 5.648129e+02
#> 15.6 5.770220e+02 5.786959e+02 5.801648e+02 5.816337e+02 5.831025e+02
#> 15.7 5.947696e+02 5.969191e+02 5.991017e+02 6.012844e+02 6.034670e+02
#> 15.8 6.170114e+02 6.190860e+02 6.208427e+02 6.225993e+02 6.243560e+02
#> 15.9 6.350267e+02 6.370883e+02 6.394646e+02 6.418409e+02 6.442172e+02
#> 16   6.574931e+02 6.598541e+02 6.623911e+02 6.649282e+02 6.674652e+02
#> 16.1 6.795322e+02 6.819446e+02 6.847287e+02 6.875128e+02 6.902968e+02
#> 16.2 7.027060e+02 7.053082e+02 7.082738e+02 7.112395e+02 7.142051e+02
#> 16.3 7.308420e+02 7.327203e+02 7.343030e+02 7.358858e+02 7.374685e+02
#> 16.4 7.540995e+02 7.561560e+02 7.579707e+02 7.597854e+02 7.616001e+02
#> 16.5 7.764029e+02 7.783934e+02 7.802951e+02 7.821968e+02 7.840984e+02
#> 16.6 7.996169e+02 8.030013e+02 8.064232e+02 8.098452e+02 8.132671e+02
#> 16.7 8.209491e+02 8.255122e+02 8.300753e+02 8.346384e+02 8.392015e+02
#> 16.8 8.524189e+02 8.541734e+02 8.559280e+02 8.576825e+02 8.594371e+02
#> 16.9 8.774435e+02 8.797544e+02 8.820652e+02 8.843761e+02 8.866869e+02
#> 17   9.059298e+02 9.095392e+02 9.131485e+02 9.167579e+02 9.203673e+02
#> 17.1 9.328087e+02 9.351670e+02 9.375254e+02 9.398838e+02 9.422422e+02
#> 17.2 9.599573e+02 9.629468e+02 9.659362e+02 9.689257e+02 9.719152e+02
#> 17.3 9.909217e+02 9.939469e+02 9.969722e+02 9.999974e+02 1.003023e+03
#> 17.4 1.017528e+03 1.021030e+03 1.024533e+03 1.028036e+03 1.031538e+03
#> 17.5 1.050995e+03 1.053829e+03 1.056662e+03 1.059495e+03 1.062329e+03
#> 17.6 1.078817e+03 1.082120e+03 1.085422e+03 1.088725e+03 1.092028e+03
#> 17.7 1.110426e+03 1.112600e+03 1.114774e+03 1.116949e+03 1.119123e+03
#> 17.8 1.142519e+03 1.144010e+03 1.145501e+03 1.146992e+03 1.148483e+03
#> 17.9 1.174212e+03 1.177107e+03 1.180002e+03 1.182896e+03 1.185791e+03
#> 18   1.202650e+03 1.206186e+03 1.209722e+03 1.213258e+03 1.216793e+03
#> 18.1 1.233060e+03 1.235611e+03 1.238163e+03 1.240715e+03 1.243276e+03
#> 18.2 1.271923e+03 1.275915e+03 1.279908e+03 1.283900e+03 1.287880e+03
#> 18.3 1.310097e+03 1.313317e+03 1.316537e+03 1.319757e+03 1.322940e+03
#> 18.4 1.341265e+03 1.344713e+03 1.348161e+03 1.351609e+03 1.355202e+03
#> 18.5 1.379434e+03 1.383085e+03 1.386737e+03 1.390389e+03 1.393859e+03
#> 18.6 1.410946e+03 1.413257e+03 1.415568e+03 1.417879e+03 1.421057e+03
#> 18.7 1.447112e+03 1.449852e+03 1.452592e+03 1.455332e+03 1.458470e+03
#> 18.8 1.481174e+03 1.485207e+03 1.489241e+03 1.493274e+03 1.497533e+03
#> 18.9 1.522643e+03 1.525894e+03 1.529145e+03 1.532396e+03 1.536123e+03
#> 19   1.560069e+03 1.563712e+03 1.567355e+03 1.570998e+03 1.574640e+03
#> 19.1 1.597380e+03 1.601562e+03 1.605743e+03 1.609924e+03 1.614910e+03
#> 19.2 1.644265e+03 1.648015e+03 1.651764e+03 1.655514e+03 1.658056e+03
#> 19.3 1.674846e+03 1.679169e+03 1.683491e+03 1.687814e+03 1.691123e+03
#> 19.4 1.715459e+03 1.721665e+03 1.727871e+03 1.734077e+03 1.738410e+03
#> 19.5 1.760561e+03 1.763375e+03 1.766189e+03 1.769003e+03 1.774714e+03
#> 19.6 1.806197e+03 1.810273e+03 1.814349e+03 1.818346e+03 1.819654e+03
#> 19.7 1.838229e+03 1.844736e+03 1.851243e+03 1.857501e+03 1.861441e+03
#> 19.8 1.890610e+03 1.897145e+03 1.903679e+03 1.909391e+03 1.910948e+03
#> 19.9 1.931676e+03 1.938410e+03 1.945144e+03 1.951513e+03 1.956681e+03
#> 20   1.982088e+03 1.984785e+03 1.987481e+03 1.990829e+03 1.995681e+03
#> 20.1 2.022620e+03 2.026540e+03 2.030461e+03 2.035178e+03 2.041254e+03
#> 20.2 2.076748e+03 2.082425e+03 2.088102e+03 2.092235e+03 2.094391e+03
#> 20.3 2.117355e+03 2.123568e+03 2.129780e+03 2.134001e+03 2.136283e+03
#> 20.4 2.160879e+03 2.167453e+03 2.174027e+03 2.180634e+03 2.187265e+03
#> 20.5 2.219108e+03 2.222695e+03 2.226282e+03 2.231487e+03 2.237589e+03
#> 20.6 2.264124e+03 2.266462e+03 2.268801e+03 2.274250e+03 2.280959e+03
#> 20.7 2.316878e+03 2.322008e+03 2.327138e+03 2.329306e+03 2.330638e+03
#> 20.8 2.351889e+03 2.357932e+03 2.363976e+03 2.370041e+03 2.376109e+03
#> 20.9 2.416669e+03 2.424177e+03 2.431685e+03 2.435297e+03 2.438555e+03
#> 21   2.473238e+03 2.481625e+03 2.490013e+03 2.488685e+03 2.487208e+03
#> 21.1 2.511458e+03 2.520949e+03 2.530247e+03 2.536119e+03 2.541991e+03
#> 21.2 2.575631e+03 2.580988e+03 2.586338e+03 2.591636e+03 2.596935e+03
#> 21.3 2.629675e+03 2.635275e+03 2.640264e+03 2.642649e+03 2.645034e+03
#> 21.4 2.675892e+03 2.683409e+03 2.690676e+03 2.697223e+03 2.703770e+03
#> 21.5 2.730794e+03 2.733618e+03 2.738493e+03 2.747601e+03 2.756710e+03
#> 21.6 2.784448e+03 2.785549e+03 2.789982e+03 2.799524e+03 2.809066e+03
#> 21.7 2.856770e+03 2.863528e+03 2.867788e+03 2.869152e+03 2.870516e+03
#> 21.8 2.900193e+03 2.907702e+03 2.914354e+03 2.920247e+03 2.926141e+03
#> 21.9 2.965925e+03 2.973059e+03 2.977917e+03 2.981255e+03 2.984593e+03
#> 22   3.016205e+03 3.022731e+03 3.029214e+03 3.035677e+03 3.042139e+03
#> 22.1 3.086048e+03 3.093898e+03 3.098623e+03 3.102230e+03 3.105837e+03
#> 22.2 3.140650e+03 3.147750e+03 3.152502e+03 3.156683e+03 3.160863e+03
#> 22.3 3.191201e+03 3.196750e+03 3.202586e+03 3.208463e+03 3.214341e+03
#> 22.4 3.251425e+03 3.257768e+03 3.267576e+03 3.277600e+03 3.287624e+03
#> 22.5 3.327468e+03 3.332404e+03 3.339233e+03 3.346063e+03 3.352892e+03
#> 22.6 3.393389e+03 3.399821e+03 3.402949e+03 3.406076e+03 3.409204e+03
#> 22.7 3.446972e+03 3.454751e+03 3.462706e+03 3.470661e+03 3.478615e+03
#> 22.8 3.529289e+03 3.537483e+03 3.543974e+03 3.550464e+03 3.556955e+03
#> 22.9 3.588964e+03 3.594277e+03 3.600742e+03 3.607208e+03 3.613673e+03
#> 23   3.658180e+03 3.665657e+03 3.672557e+03 3.679456e+03 3.686356e+03
#> 23.1 3.720229e+03 3.725185e+03 3.729825e+03 3.734465e+03 3.739105e+03
#> 23.2 3.783443e+03 3.789953e+03 3.794537e+03 3.799120e+03 3.803703e+03
#> 23.3 3.856877e+03 3.865566e+03 3.873001e+03 3.880435e+03 3.887869e+03
#> 23.4 3.897582e+03 3.906736e+03 3.921502e+03 3.936267e+03 3.951032e+03
#> 23.5 3.997803e+03 4.003983e+03 4.010337e+03 4.016691e+03 4.023045e+03
#> 23.6 4.057553e+03 4.063744e+03 4.070124e+03 4.076504e+03 4.082885e+03
#> 23.7 4.137818e+03 4.147389e+03 4.156901e+03 4.166414e+03 4.175927e+03
#> 23.8 4.194620e+03 4.202399e+03 4.210880e+03 4.219361e+03 4.227843e+03
#> 23.9 4.287275e+03 4.295416e+03 4.303483e+03 4.311551e+03 4.319619e+03
#> 24             NA           NA           NA           NA           NA

# }