Predict method for discharge rating curves

Predict the discharge for given stage values based on a discharge rating curve result object.

# S3 method for plm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for plm
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm0
predict(object, newdata = NULL, wide = FALSE, ...)

# S3 method for gplm
predict(object, newdata = NULL, wide = FALSE, ...)

Arguments

object
an object of class "plm0", "plm", "gplm0" or "gplm".

newdata
a numeric vector of stage values for which to predict. If omitted, the stage values in the data are used.

wide
a logical value denoting whether to produce a wide prediction output.If TRUE, then the output is a table with median prediction values for an equally spaced grid of stages with 1 cm increments, each row containing predictions in a decimeter range of stages.

...
not used in this function

Value

an object of class "data.frame" with four columns, h (stage), lower (2.5% posterior predictive quantile), median (50% posterior predictive quantile), upper (97.5% posterior predictive quantile). If wide=TRUE, a matrix as described above (see wide parameter) is returned.

Functions

predict.plm0: Predict method for plm0

predict.plm: Predict method for plm

predict.gplm0: Predict method for gplm0

predict.gplm: Predict method for gplm

See also

Examples

# \donttest{
data(krokfors)
plm0.fit <- plm0(Q~W,krokfors,h_max=24)
#predict rating curve on a equally 1 cm spaced grid from 1 to 2 meters
predict(plm0.fit,newdata=seq(23,24,by=0.01))

#>         h    lower   median     upper
#> 1   23.00 1511.161 3634.841  8829.606
#> 2   23.01 1514.743 3641.635  8849.981
#> 3   23.02 1524.655 3647.708  8839.057
#> 4   23.03 1534.568 3653.781  8828.133
#> 5   23.04 1544.480 3659.853  8817.209
#> 6   23.05 1554.393 3665.926  8806.285
#> 7   23.06 1561.146 3672.345  8807.276
#> 8   23.07 1562.059 3679.404  8830.294
#> 9   23.08 1562.972 3686.463  8853.312
#> 10  23.09 1563.885 3693.521  8876.330
#> 11  23.10 1564.798 3700.580  8899.348
#> 12  23.11 1565.655 3706.970  8920.848
#> 13  23.12 1566.422 3712.293  8939.927
#> 14  23.13 1567.190 3717.615  8959.006
#> 15  23.14 1567.957 3722.937  8978.085
#> 16  23.15 1568.725 3728.259  8997.163
#> 17  23.16 1571.067 3732.597  9011.815
#> 18  23.17 1575.591 3735.572  9020.335
#> 19  23.18 1580.114 3738.547  9028.856
#> 20  23.19 1584.637 3741.521  9037.377
#> 21  23.20 1589.160 3744.496  9045.898
#> 22  23.21 1591.536 3750.271  9063.692
#> 23  23.22 1591.326 3759.420  9092.661
#> 24  23.23 1591.115 3768.568  9121.629
#> 25  23.24 1590.905 3777.717  9150.598
#> 26  23.25 1590.694 3786.865  9179.566
#> 27  23.26 1591.494 3794.006  9203.039
#> 28  23.27 1593.354 3799.038  9220.739
#> 29  23.28 1595.215 3804.070  9238.440
#> 30  23.29 1597.076 3809.102  9256.140
#> 31  23.30 1598.937 3814.134  9273.840
#> 32  23.31 1601.313 3822.872  9295.235
#> 33  23.32 1604.160 3835.005  9320.017
#> 34  23.33 1607.008 3847.138  9344.799
#> 35  23.34 1609.855 3859.271  9369.581
#> 36  23.35 1612.703 3871.404  9394.362
#> 37  23.36 1615.089 3883.978  9423.546
#> 38  23.37 1617.107 3896.905  9456.246
#> 39  23.38 1619.125 3909.831  9488.945
#> 40  23.39 1621.144 3922.758  9521.644
#> 41  23.40 1623.162 3935.685  9554.343
#> 42  23.41 1629.995 3939.948  9572.662
#> 43  23.42 1640.172 3938.194  9580.992
#> 44  23.43 1650.349 3936.440  9589.322
#> 45  23.44 1660.526 3934.685  9597.652
#> 46  23.45 1670.703 3932.931  9605.982
#> 47  23.46 1672.890 3939.580  9617.166
#> 48  23.47 1670.268 3951.285  9630.066
#> 49  23.48 1667.646 3962.990  9642.967
#> 50  23.49 1665.024 3974.695  9655.868
#> 51  23.50 1662.402 3986.400  9668.769
#> 52  23.51 1660.046 3992.145  9695.378
#> 53  23.52 1657.827 3994.798  9729.098
#> 54  23.53 1655.608 3997.452  9762.819
#> 55  23.54 1653.389 4000.105  9796.539
#> 56  23.55 1651.171 4002.758  9830.259
#> 57  23.56 1656.754 4009.528  9826.817
#> 58  23.57 1665.801 4018.124  9806.881
#> 59  23.58 1674.848 4026.721  9786.944
#> 60  23.59 1683.895 4035.317  9767.008
#> 61  23.60 1692.942 4043.913  9747.072
#> 62  23.61 1694.469 4050.140  9776.968
#> 63  23.62 1693.169 4055.475  9825.599
#> 64  23.63 1691.870 4060.810  9874.229
#> 65  23.64 1690.570 4066.145  9922.860
#> 66  23.65 1689.270 4071.480  9971.491
#> 67  23.66 1693.080 4079.993  9986.742
#> 68  23.67 1698.495 4089.505  9991.506
#> 69  23.68 1703.910 4099.016  9996.270
#> 70  23.69 1709.324 4108.528 10001.033
#> 71  23.70 1714.739 4118.039 10005.797
#> 72  23.71 1715.546 4127.645 10035.498
#> 73  23.72 1715.164 4137.274 10071.626
#> 74  23.73 1714.782 4146.904 10107.754
#> 75  23.74 1714.401 4156.533 10143.882
#> 76  23.75 1714.019 4166.163 10180.010
#> 77  23.76 1718.934 4171.811 10212.830
#> 78  23.77 1724.939 4176.638 10244.968
#> 79  23.78 1730.945 4181.466 10277.107
#> 80  23.79 1736.951 4186.293 10309.246
#> 81  23.80 1742.956 4191.121 10341.385
#> 82  23.81 1744.147 4198.169 10351.561
#> 83  23.82 1744.575 4205.569 10358.261
#> 84  23.83 1745.004 4212.968 10364.962
#> 85  23.84 1745.433 4220.368 10371.662
#> 86  23.85 1745.862 4227.768 10378.363
#> 87  23.86 1741.205 4232.119 10432.292
#> 88  23.87 1735.968 4236.122 10491.613
#> 89  23.88 1730.731 4240.125 10550.934
#> 90  23.89 1725.494 4244.128 10610.254
#> 91  23.90 1720.258 4248.132 10669.575
#> 92  23.91 1732.517 4261.798 10643.064
#> 93  23.92 1746.059 4276.172 10610.261
#> 94  23.93 1759.601 4290.547 10577.457
#> 95  23.94 1773.143 4304.921 10544.653
#> 96  23.95 1786.685 4319.295 10511.849
#> 97  23.96 1788.333 4323.565 10558.960
#> 98  23.97 1789.561 4327.477 10608.896
#> 99  23.98 1790.789 4331.390 10658.832
#> 100 23.99 1792.017 4335.302 10708.768
#> 101 24.00 1793.244 4339.214 10758.705

predict(plm0.fit,wide=TRUE)

#>                 0         0.01         0.02         0.03         0.04
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> 7.7  1.632132e-04 2.255341e-04 2.878550e-04 5.595533e-04 9.601708e-04
#> 7.8  4.984681e-03 6.004183e-03 7.023684e-03 8.596049e-03 1.050883e-02
#> 7.9  2.431175e-02 2.764660e-02 3.098145e-02 3.471303e-02 3.874431e-02
#> 8    6.909599e-02 7.499178e-02 8.171038e-02 8.842898e-02 9.533813e-02
#> 8.1  1.468629e-01 1.573550e-01 1.678472e-01 1.783393e-01 1.896543e-01
#> 8.2  2.666313e-01 2.808803e-01 2.958787e-01 3.112707e-01 3.266626e-01
#> 8.3  4.351753e-01 4.546757e-01 4.752732e-01 4.975166e-01 5.196052e-01
#> 8.4  6.615279e-01 6.873489e-01 7.137693e-01 7.417697e-01 7.697702e-01
#> 8.5  9.505088e-01 9.848945e-01 1.020026e+00 1.055343e+00 1.090814e+00
#> 8.6  1.313767e+00 1.354910e+00 1.396512e+00 1.438114e+00 1.482026e+00
#> 8.7  1.756441e+00 1.807908e+00 1.859375e+00 1.910843e+00 1.961644e+00
#> 8.8  2.279626e+00 2.338324e+00 2.397023e+00 2.456191e+00 2.518544e+00
#> 8.9  2.904809e+00 2.972590e+00 3.040701e+00 3.108811e+00 3.181684e+00
#> 9    3.624042e+00 3.696247e+00 3.779819e+00 3.863391e+00 3.946733e+00
#> 9.1  4.448638e+00 4.527472e+00 4.612773e+00 4.713154e+00 4.812482e+00
#> 9.2  5.386750e+00 5.486821e+00 5.587071e+00 5.688939e+00 5.792254e+00
#> 9.3  6.455023e+00 6.567765e+00 6.682614e+00 6.797463e+00 6.913149e+00
#> 9.4  7.617731e+00 7.756043e+00 7.894355e+00 8.032667e+00 8.156170e+00
#> 9.5  8.949125e+00 9.068056e+00 9.209302e+00 9.376048e+00 9.526689e+00
#> 9.6  1.039928e+01 1.055210e+01 1.070535e+01 1.085896e+01 1.101258e+01
#> 9.7  1.196851e+01 1.215350e+01 1.233850e+01 1.252349e+01 1.269587e+01
#> 9.8  1.372549e+01 1.392067e+01 1.411385e+01 1.430355e+01 1.449326e+01
#> 9.9  1.559336e+01 1.580886e+01 1.602436e+01 1.623986e+01 1.645535e+01
#> 10   1.773400e+01 1.795347e+01 1.817294e+01 1.839241e+01 1.861188e+01
#> 10.1 1.994648e+01 2.018173e+01 2.041698e+01 2.065223e+01 2.088748e+01
#> 10.2 2.243494e+01 2.267729e+01 2.291965e+01 2.316200e+01 2.340436e+01
#> 10.3 2.493016e+01 2.520981e+01 2.548945e+01 2.576910e+01 2.604875e+01
#> 10.4 2.772030e+01 2.802310e+01 2.832589e+01 2.862869e+01 2.893148e+01
#> 10.5 3.070696e+01 3.100861e+01 3.131027e+01 3.161192e+01 3.191357e+01
#> 10.6 3.393770e+01 3.424085e+01 3.454400e+01 3.484715e+01 3.515030e+01
#> 10.7 3.729949e+01 3.766045e+01 3.802141e+01 3.838237e+01 3.874333e+01
#> 10.8 4.082669e+01 4.119348e+01 4.156027e+01 4.192706e+01 4.229385e+01
#> 10.9 4.479745e+01 4.519678e+01 4.559612e+01 4.599546e+01 4.639384e+01
#> 11   4.853863e+01 4.896171e+01 4.938479e+01 4.980787e+01 5.023312e+01
#> 11.1 5.293238e+01 5.338876e+01 5.384514e+01 5.430152e+01 5.475231e+01
#> 11.2 5.731680e+01 5.777393e+01 5.823106e+01 5.868819e+01 5.914947e+01
#> 11.3 6.205573e+01 6.257252e+01 6.308930e+01 6.360608e+01 6.411193e+01
#> 11.4 6.707828e+01 6.762641e+01 6.817454e+01 6.872267e+01 6.927239e+01
#> 11.5 7.245620e+01 7.291431e+01 7.337242e+01 7.383053e+01 7.436783e+01
#> 11.6 7.799048e+01 7.847081e+01 7.895113e+01 7.943145e+01 7.998420e+01
#> 11.7 8.358663e+01 8.411074e+01 8.463486e+01 8.515898e+01 8.572760e+01
#> 11.8 8.936251e+01 8.997592e+01 9.058934e+01 9.120276e+01 9.190008e+01
#> 11.9 9.605331e+01 9.660513e+01 9.715695e+01 9.770878e+01 9.831809e+01
#> 12   1.022555e+02 1.029666e+02 1.036777e+02 1.043888e+02 1.050234e+02
#> 12.1 1.090071e+02 1.097930e+02 1.105789e+02 1.113647e+02 1.120799e+02
#> 12.2 1.162251e+02 1.168928e+02 1.175604e+02 1.182281e+02 1.189320e+02
#> 12.3 1.235218e+02 1.244064e+02 1.252911e+02 1.261758e+02 1.268663e+02
#> 12.4 1.311605e+02 1.319367e+02 1.327129e+02 1.334869e+02 1.342099e+02
#> 12.5 1.389638e+02 1.398871e+02 1.408104e+02 1.417292e+02 1.426113e+02
#> 12.6 1.474433e+02 1.481105e+02 1.487776e+02 1.494902e+02 1.504122e+02
#> 12.7 1.560227e+02 1.569802e+02 1.579376e+02 1.588744e+02 1.597480e+02
#> 12.8 1.651650e+02 1.661156e+02 1.670661e+02 1.679701e+02 1.687732e+02
#> 12.9 1.740928e+02 1.751093e+02 1.761258e+02 1.771214e+02 1.780835e+02
#> 13   1.840647e+02 1.851132e+02 1.861617e+02 1.872102e+02 1.882586e+02
#> 13.1 1.942108e+02 1.951232e+02 1.960355e+02 1.969295e+02 1.978065e+02
#> 13.2 2.036023e+02 2.046883e+02 2.057743e+02 2.069881e+02 2.082913e+02
#> 13.3 2.151540e+02 2.160925e+02 2.170310e+02 2.180792e+02 2.191846e+02
#> 13.4 2.254720e+02 2.264491e+02 2.274262e+02 2.286004e+02 2.298493e+02
#> 13.5 2.371023e+02 2.382641e+02 2.394259e+02 2.405514e+02 2.416674e+02
#> 13.6 2.486766e+02 2.499034e+02 2.511302e+02 2.520804e+02 2.529863e+02
#> 13.7 2.594762e+02 2.607463e+02 2.620164e+02 2.636031e+02 2.652136e+02
#> 13.8 2.730285e+02 2.740154e+02 2.750024e+02 2.765065e+02 2.780115e+02
#> 13.9 2.858037e+02 2.869005e+02 2.880077e+02 2.892616e+02 2.905155e+02
#> 14   2.997411e+02 3.015440e+02 3.032454e+02 3.042958e+02 3.053463e+02
#> 14.1 3.126928e+02 3.140727e+02 3.154230e+02 3.166577e+02 3.178924e+02
#> 14.2 3.262878e+02 3.278275e+02 3.293353e+02 3.307575e+02 3.321797e+02
#> 14.3 3.403439e+02 3.416544e+02 3.431026e+02 3.448181e+02 3.465337e+02
#> 14.4 3.556909e+02 3.570697e+02 3.583692e+02 3.595537e+02 3.607382e+02
#> 14.5 3.723083e+02 3.747893e+02 3.766225e+02 3.777437e+02 3.788648e+02
#> 14.6 3.846500e+02 3.855029e+02 3.872979e+02 3.898804e+02 3.924629e+02
#> 14.7 4.022167e+02 4.031950e+02 4.047651e+02 4.067088e+02 4.086525e+02
#> 14.8 4.190303e+02 4.206218e+02 4.221338e+02 4.236087e+02 4.250835e+02
#> 14.9 4.362197e+02 4.383101e+02 4.398139e+02 4.411218e+02 4.424297e+02
#> 15   4.534772e+02 4.556309e+02 4.568615e+02 4.578867e+02 4.589119e+02
#> 15.1 4.688551e+02 4.708646e+02 4.729002e+02 4.749391e+02 4.769780e+02
#> 15.2 4.871117e+02 4.886150e+02 4.905122e+02 4.924282e+02 4.943442e+02
#> 15.3 5.076830e+02 5.100370e+02 5.113673e+02 5.126976e+02 5.140278e+02
#> 15.4 5.255150e+02 5.276708e+02 5.294431e+02 5.312154e+02 5.329877e+02
#> 15.5 5.450484e+02 5.471471e+02 5.491436e+02 5.511401e+02 5.531366e+02
#> 15.6 5.668153e+02 5.691507e+02 5.712615e+02 5.733723e+02 5.754830e+02
#> 15.7 5.851774e+02 5.868374e+02 5.890826e+02 5.913277e+02 5.935729e+02
#> 15.8 6.067969e+02 6.089521e+02 6.110499e+02 6.131478e+02 6.152456e+02
#> 15.9 6.275322e+02 6.293268e+02 6.308132e+02 6.322996e+02 6.337860e+02
#> 16   6.476453e+02 6.502073e+02 6.527391e+02 6.552709e+02 6.578027e+02
#> 16.1 6.726602e+02 6.746462e+02 6.762751e+02 6.779040e+02 6.795329e+02
#> 16.2 6.950845e+02 6.971372e+02 6.987175e+02 7.002978e+02 7.018782e+02
#> 16.3 7.186334e+02 7.211621e+02 7.234375e+02 7.257129e+02 7.279883e+02
#> 16.4 7.400376e+02 7.430028e+02 7.462677e+02 7.495325e+02 7.527974e+02
#> 16.5 7.659340e+02 7.682238e+02 7.705821e+02 7.729403e+02 7.752986e+02
#> 16.6 7.883834e+02 7.911666e+02 7.940127e+02 7.968588e+02 7.997049e+02
#> 16.7 8.169325e+02 8.185079e+02 8.200558e+02 8.216036e+02 8.231515e+02
#> 16.8 8.437942e+02 8.460383e+02 8.482824e+02 8.505265e+02 8.527706e+02
#> 16.9 8.665816e+02 8.690020e+02 8.714224e+02 8.738428e+02 8.762632e+02
#> 17   8.901669e+02 8.939761e+02 8.977853e+02 9.015946e+02 9.054038e+02
#> 17.1 9.242259e+02 9.263208e+02 9.284156e+02 9.305105e+02 9.326054e+02
#> 17.2 9.478016e+02 9.506587e+02 9.535159e+02 9.563730e+02 9.592302e+02
#> 17.3 9.789049e+02 9.815195e+02 9.841340e+02 9.867486e+02 9.893631e+02
#> 17.4 1.007315e+03 1.009221e+03 1.011127e+03 1.013033e+03 1.014938e+03
#> 17.5 1.034714e+03 1.037882e+03 1.041049e+03 1.044216e+03 1.047383e+03
#> 17.6 1.067478e+03 1.070302e+03 1.073126e+03 1.075951e+03 1.078775e+03
#> 17.7 1.096354e+03 1.099609e+03 1.102865e+03 1.106120e+03 1.109376e+03
#> 17.8 1.127825e+03 1.130812e+03 1.133799e+03 1.136786e+03 1.139774e+03
#> 17.9 1.156095e+03 1.160980e+03 1.165865e+03 1.170750e+03 1.175635e+03
#> 18   1.193043e+03 1.196362e+03 1.199680e+03 1.202999e+03 1.206317e+03
#> 18.1 1.222049e+03 1.224744e+03 1.227439e+03 1.230134e+03 1.232829e+03
#> 18.2 1.253106e+03 1.257557e+03 1.262007e+03 1.266458e+03 1.270899e+03
#> 18.3 1.294851e+03 1.297916e+03 1.300980e+03 1.304045e+03 1.307148e+03
#> 18.4 1.328088e+03 1.331849e+03 1.335611e+03 1.339373e+03 1.343033e+03
#> 18.5 1.362775e+03 1.366501e+03 1.370227e+03 1.373953e+03 1.377667e+03
#> 18.6 1.399433e+03 1.402874e+03 1.406315e+03 1.409756e+03 1.412827e+03
#> 18.7 1.429562e+03 1.434218e+03 1.438875e+03 1.443531e+03 1.447518e+03
#> 18.8 1.467160e+03 1.471665e+03 1.476171e+03 1.480676e+03 1.484369e+03
#> 18.9 1.503547e+03 1.508284e+03 1.513022e+03 1.517760e+03 1.522163e+03
#> 19   1.547120e+03 1.551483e+03 1.555845e+03 1.560207e+03 1.564157e+03
#> 19.1 1.584828e+03 1.587702e+03 1.590575e+03 1.593449e+03 1.597477e+03
#> 19.2 1.627830e+03 1.633955e+03 1.640079e+03 1.646204e+03 1.650096e+03
#> 19.3 1.665549e+03 1.667029e+03 1.668510e+03 1.669990e+03 1.674689e+03
#> 19.4 1.703640e+03 1.706787e+03 1.709934e+03 1.713081e+03 1.718614e+03
#> 19.5 1.750555e+03 1.754563e+03 1.758571e+03 1.762579e+03 1.764205e+03
#> 19.6 1.782400e+03 1.788724e+03 1.795048e+03 1.801372e+03 1.806026e+03
#> 19.7 1.827924e+03 1.829627e+03 1.831329e+03 1.833332e+03 1.839824e+03
#> 19.8 1.873546e+03 1.877543e+03 1.881541e+03 1.885831e+03 1.892065e+03
#> 19.9 1.920321e+03 1.922330e+03 1.924338e+03 1.927110e+03 1.932946e+03
#> 20   1.968147e+03 1.974067e+03 1.979986e+03 1.985010e+03 1.987586e+03
#> 20.1 2.005020e+03 2.008456e+03 2.011893e+03 2.015755e+03 2.020458e+03
#> 20.2 2.050716e+03 2.056278e+03 2.061841e+03 2.067580e+03 2.073578e+03
#> 20.3 2.101573e+03 2.104293e+03 2.107013e+03 2.111426e+03 2.117730e+03
#> 20.4 2.147497e+03 2.150588e+03 2.153679e+03 2.158313e+03 2.164257e+03
#> 20.5 2.200457e+03 2.206609e+03 2.212761e+03 2.217541e+03 2.221441e+03
#> 20.6 2.253786e+03 2.261071e+03 2.268355e+03 2.271776e+03 2.273358e+03
#> 20.7 2.297002e+03 2.303803e+03 2.310605e+03 2.314762e+03 2.318017e+03
#> 20.8 2.334725e+03 2.336965e+03 2.339205e+03 2.346102e+03 2.354064e+03
#> 20.9 2.389875e+03 2.393639e+03 2.397403e+03 2.405179e+03 2.413490e+03
#> 21   2.453317e+03 2.458187e+03 2.463056e+03 2.469648e+03 2.476328e+03
#> 21.1 2.492052e+03 2.490571e+03 2.489316e+03 2.499670e+03 2.510025e+03
#> 21.2 2.560309e+03 2.566785e+03 2.573097e+03 2.577703e+03 2.582310e+03
#> 21.3 2.609250e+03 2.613633e+03 2.618310e+03 2.624577e+03 2.630844e+03
#> 21.4 2.653030e+03 2.654464e+03 2.657193e+03 2.664412e+03 2.671630e+03
#> 21.5 2.710732e+03 2.716658e+03 2.722037e+03 2.726088e+03 2.730139e+03
#> 21.6 2.768193e+03 2.776377e+03 2.783191e+03 2.787576e+03 2.791961e+03
#> 21.7 2.828327e+03 2.835675e+03 2.842307e+03 2.847987e+03 2.853667e+03
#> 21.8 2.877757e+03 2.880553e+03 2.885506e+03 2.892643e+03 2.899779e+03
#> 21.9 2.952441e+03 2.962364e+03 2.968641e+03 2.972115e+03 2.975589e+03
#> 22   3.001679e+03 3.006611e+03 3.011117e+03 3.015378e+03 3.019639e+03
#> 22.1 3.056604e+03 3.063973e+03 3.071744e+03 3.079687e+03 3.087630e+03
#> 22.2 3.119699e+03 3.123469e+03 3.129591e+03 3.136414e+03 3.143237e+03
#> 22.3 3.169847e+03 3.172904e+03 3.179907e+03 3.187669e+03 3.195431e+03
#> 22.4 3.231733e+03 3.236843e+03 3.244252e+03 3.251897e+03 3.259542e+03
#> 22.5 3.315563e+03 3.325783e+03 3.331466e+03 3.337034e+03 3.342602e+03
#> 22.6 3.375645e+03 3.381147e+03 3.387197e+03 3.393246e+03 3.399295e+03
#> 22.7 3.429624e+03 3.434678e+03 3.443439e+03 3.452200e+03 3.460962e+03
#> 22.8 3.510298e+03 3.518166e+03 3.525513e+03 3.532861e+03 3.540208e+03
#> 22.9 3.580672e+03 3.586486e+03 3.590264e+03 3.594041e+03 3.597819e+03
#> 23   3.634841e+03 3.641635e+03 3.647708e+03 3.653781e+03 3.659853e+03
#> 23.1 3.700580e+03 3.706970e+03 3.712293e+03 3.717615e+03 3.722937e+03
#> 23.2 3.744496e+03 3.750271e+03 3.759420e+03 3.768568e+03 3.777717e+03
#> 23.3 3.814134e+03 3.822872e+03 3.835005e+03 3.847138e+03 3.859271e+03
#> 23.4 3.935685e+03 3.939948e+03 3.938194e+03 3.936440e+03 3.934685e+03
#> 23.5 3.986400e+03 3.992145e+03 3.994798e+03 3.997452e+03 4.000105e+03
#> 23.6 4.043913e+03 4.050140e+03 4.055475e+03 4.060810e+03 4.066145e+03
#> 23.7 4.118039e+03 4.127645e+03 4.137274e+03 4.146904e+03 4.156533e+03
#> 23.8 4.191121e+03 4.198169e+03 4.205569e+03 4.212968e+03 4.220368e+03
#> 23.9 4.248132e+03 4.261798e+03 4.276172e+03 4.290547e+03 4.304921e+03
#> 24   4.339214e+03           NA           NA           NA           NA
#>              0.05         0.06         0.07         0.08         0.09
#> 7.6  0.000000e+00 0.000000e+00 0.000000e+00 3.857142e-05 1.008923e-04
#> 7.7  1.360788e-03 1.761406e-03 2.162023e-03 2.945678e-03 3.965180e-03
#> 7.8  1.242160e-02 1.433438e-02 1.624716e-02 1.865013e-02 2.135493e-02
#> 7.9  4.277560e-02 4.752225e-02 5.252116e-02 5.752007e-02 6.327040e-02
#> 8    1.033071e-01 1.112760e-01 1.194733e-01 1.284518e-01 1.374303e-01
#> 8.1  2.014681e-01 2.132819e-01 2.261519e-01 2.390580e-01 2.523822e-01
#> 8.2  3.444011e-01 3.621395e-01 3.799586e-01 3.978697e-01 4.157809e-01
#> 8.3  5.415906e-01 5.629081e-01 5.840584e-01 6.098828e-01 6.357069e-01
#> 8.4  7.983772e-01 8.277256e-01 8.570740e-01 8.877057e-01 9.191072e-01
#> 8.5  1.126250e+00 1.161602e+00 1.197102e+00 1.235990e+00 1.274878e+00
#> 8.6  1.527349e+00 1.572673e+00 1.618156e+00 1.663867e+00 1.709578e+00
#> 8.7  2.012396e+00 2.063148e+00 2.115998e+00 2.169908e+00 2.223818e+00
#> 8.8  2.580897e+00 2.643250e+00 2.705603e+00 2.771105e+00 2.837957e+00
#> 8.9  3.255923e+00 3.330162e+00 3.403798e+00 3.477298e+00 3.550798e+00
#> 9    4.027998e+00 4.109263e+00 4.190933e+00 4.276242e+00 4.361552e+00
#> 9.1  4.900449e+00 4.988415e+00 5.082305e+00 5.184267e+00 5.286229e+00
#> 9.2  5.898070e+00 6.011942e+00 6.130479e+00 6.239375e+00 6.347199e+00
#> 9.3  7.029672e+00 7.146194e+00 7.263023e+00 7.380567e+00 7.498110e+00
#> 9.4  8.278027e+00 8.399884e+00 8.528793e+00 8.665382e+00 8.811841e+00
#> 9.5  9.666593e+00 9.806496e+00 9.946399e+00 1.009363e+01 1.024645e+01
#> 9.6  1.116619e+01 1.132224e+01 1.148071e+01 1.163919e+01 1.179766e+01
#> 9.7  1.285175e+01 1.300764e+01 1.316352e+01 1.333512e+01 1.353031e+01
#> 9.8  1.468296e+01 1.486888e+01 1.504721e+01 1.522555e+01 1.540388e+01
#> 9.9  1.666979e+01 1.688210e+01 1.709442e+01 1.730673e+01 1.751905e+01
#> 10   1.883223e+01 1.905390e+01 1.927556e+01 1.949722e+01 1.971889e+01
#> 10.1 2.113528e+01 2.139718e+01 2.165909e+01 2.192100e+01 2.218290e+01
#> 10.2 2.365217e+01 2.390465e+01 2.415714e+01 2.440962e+01 2.466210e+01
#> 10.3 2.632581e+01 2.660119e+01 2.687657e+01 2.715195e+01 2.742733e+01
#> 10.4 2.922877e+01 2.952342e+01 2.981806e+01 3.011270e+01 3.040735e+01
#> 10.5 3.224714e+01 3.259169e+01 3.293625e+01 3.328081e+01 3.362536e+01
#> 10.6 3.549948e+01 3.585928e+01 3.621909e+01 3.657890e+01 3.693871e+01
#> 10.7 3.908873e+01 3.943202e+01 3.977532e+01 4.011861e+01 4.046191e+01
#> 10.8 4.271240e+01 4.313373e+01 4.355507e+01 4.397640e+01 4.439773e+01
#> 10.9 4.673737e+01 4.708089e+01 4.742441e+01 4.776794e+01 4.811555e+01
#> 11   5.068150e+01 5.112989e+01 5.157827e+01 5.202666e+01 5.247600e+01
#> 11.1 5.517239e+01 5.559247e+01 5.601255e+01 5.643262e+01 5.685967e+01
#> 11.2 5.962526e+01 6.010106e+01 6.057685e+01 6.105265e+01 6.153895e+01
#> 11.3 6.459110e+01 6.507027e+01 6.554943e+01 6.602860e+01 6.653015e+01
#> 11.4 6.982495e+01 7.037752e+01 7.093008e+01 7.148265e+01 7.199809e+01
#> 11.5 7.501135e+01 7.565487e+01 7.629839e+01 7.694192e+01 7.751016e+01
#> 11.6 8.061071e+01 8.123721e+01 8.186372e+01 8.249023e+01 8.306251e+01
#> 11.7 8.633066e+01 8.693372e+01 8.753678e+01 8.813984e+01 8.874909e+01
#> 11.8 9.264626e+01 9.339243e+01 9.413861e+01 9.488478e+01 9.550148e+01
#> 11.9 9.895201e+01 9.958592e+01 1.002198e+02 1.008538e+02 1.015444e+02
#> 12   1.056350e+02 1.062466e+02 1.068582e+02 1.074698e+02 1.082213e+02
#> 12.1 1.127813e+02 1.134827e+02 1.141841e+02 1.148855e+02 1.155575e+02
#> 12.2 1.196398e+02 1.203476e+02 1.210554e+02 1.217632e+02 1.226371e+02
#> 12.3 1.275515e+02 1.282368e+02 1.289221e+02 1.296080e+02 1.303842e+02
#> 12.4 1.349329e+02 1.356560e+02 1.363790e+02 1.371172e+02 1.380405e+02
#> 12.5 1.434935e+02 1.443757e+02 1.452579e+02 1.461090e+02 1.467762e+02
#> 12.6 1.513343e+02 1.522563e+02 1.531783e+02 1.541078e+02 1.550653e+02
#> 12.7 1.606216e+02 1.614952e+02 1.623688e+02 1.632640e+02 1.642145e+02
#> 12.8 1.695762e+02 1.703792e+02 1.711822e+02 1.720598e+02 1.730763e+02
#> 12.9 1.790455e+02 1.800076e+02 1.809696e+02 1.819677e+02 1.830162e+02
#> 13   1.893070e+02 1.903554e+02 1.914038e+02 1.923861e+02 1.932984e+02
#> 13.1 1.986835e+02 1.995605e+02 2.004374e+02 2.014303e+02 2.025163e+02
#> 13.2 2.095945e+02 2.108976e+02 2.122008e+02 2.132770e+02 2.142155e+02
#> 13.3 2.202901e+02 2.213956e+02 2.225010e+02 2.235178e+02 2.244949e+02
#> 13.4 2.310982e+02 2.323471e+02 2.335960e+02 2.347787e+02 2.359405e+02
#> 13.5 2.427834e+02 2.438994e+02 2.450154e+02 2.462231e+02 2.474498e+02
#> 13.6 2.538921e+02 2.547980e+02 2.557038e+02 2.569360e+02 2.582061e+02
#> 13.7 2.668241e+02 2.684346e+02 2.700451e+02 2.710545e+02 2.720415e+02
#> 13.8 2.795165e+02 2.810215e+02 2.825133e+02 2.836101e+02 2.847069e+02
#> 13.9 2.917694e+02 2.930233e+02 2.943324e+02 2.961353e+02 2.979382e+02
#> 14   3.063967e+02 3.074471e+02 3.085532e+02 3.099331e+02 3.113129e+02
#> 14.1 3.191271e+02 3.203618e+02 3.216688e+02 3.232085e+02 3.247482e+02
#> 14.2 3.336020e+02 3.350242e+02 3.364123e+02 3.377228e+02 3.390334e+02
#> 14.3 3.482492e+02 3.499648e+02 3.515544e+02 3.529332e+02 3.543120e+02
#> 14.4 3.619226e+02 3.631071e+02 3.648650e+02 3.673461e+02 3.698272e+02
#> 14.5 3.799859e+02 3.811071e+02 3.820912e+02 3.829441e+02 3.837970e+02
#> 14.6 3.950454e+02 3.976278e+02 3.992817e+02 4.002600e+02 4.012384e+02
#> 14.7 4.105962e+02 4.125399e+02 4.142557e+02 4.158472e+02 4.174387e+02
#> 14.8 4.265584e+02 4.280333e+02 4.299485e+02 4.320389e+02 4.341293e+02
#> 14.9 4.437376e+02 4.450455e+02 4.470163e+02 4.491700e+02 4.513236e+02
#> 15   4.599371e+02 4.609623e+02 4.628263e+02 4.648359e+02 4.668455e+02
#> 15.1 4.790169e+02 4.810558e+02 4.826018e+02 4.841051e+02 4.856084e+02
#> 15.2 4.962602e+02 4.981762e+02 5.005490e+02 5.029270e+02 5.053050e+02
#> 15.3 5.153581e+02 5.167377e+02 5.189320e+02 5.211263e+02 5.233207e+02
#> 15.4 5.347600e+02 5.365757e+02 5.386938e+02 5.408120e+02 5.429302e+02
#> 15.5 5.551331e+02 5.572081e+02 5.596099e+02 5.620117e+02 5.644135e+02
#> 15.6 5.775938e+02 5.795220e+02 5.809358e+02 5.823497e+02 5.837636e+02
#> 15.7 5.958180e+02 5.980443e+02 6.002325e+02 6.024206e+02 6.046088e+02
#> 15.8 6.173434e+02 6.194140e+02 6.214436e+02 6.234731e+02 6.255027e+02
#> 15.9 6.352724e+02 6.372753e+02 6.398678e+02 6.424603e+02 6.450528e+02
#> 16   6.603345e+02 6.628270e+02 6.652853e+02 6.677436e+02 6.702019e+02
#> 16.1 6.811618e+02 6.835482e+02 6.864323e+02 6.893163e+02 6.922004e+02
#> 16.2 7.034585e+02 7.060848e+02 7.092219e+02 7.123591e+02 7.154962e+02
#> 16.3 7.302636e+02 7.322887e+02 7.342259e+02 7.361631e+02 7.381003e+02
#> 16.4 7.560623e+02 7.582423e+02 7.601652e+02 7.620882e+02 7.640111e+02
#> 16.5 7.776569e+02 7.798237e+02 7.819636e+02 7.841036e+02 7.862435e+02
#> 16.6 8.025509e+02 8.054259e+02 8.083026e+02 8.111792e+02 8.140558e+02
#> 16.7 8.247299e+02 8.285578e+02 8.323857e+02 8.362136e+02 8.400415e+02
#> 16.8 8.550201e+02 8.573299e+02 8.596396e+02 8.619493e+02 8.642591e+02
#> 16.9 8.786572e+02 8.809015e+02 8.831458e+02 8.853901e+02 8.876344e+02
#> 17   9.090547e+02 9.121389e+02 9.152231e+02 9.183073e+02 9.213915e+02
#> 17.1 9.348378e+02 9.374124e+02 9.399870e+02 9.425617e+02 9.451363e+02
#> 17.2 9.622750e+02 9.656610e+02 9.690470e+02 9.724331e+02 9.758191e+02
#> 17.3 9.921983e+02 9.953342e+02 9.984701e+02 1.001606e+03 1.004742e+03
#> 17.4 1.017604e+03 1.021056e+03 1.024508e+03 1.027960e+03 1.031412e+03
#> 17.5 1.050696e+03 1.054124e+03 1.057552e+03 1.060980e+03 1.064408e+03
#> 17.6 1.081643e+03 1.084538e+03 1.087432e+03 1.090326e+03 1.093220e+03
#> 17.7 1.112507e+03 1.115584e+03 1.118660e+03 1.121737e+03 1.124813e+03
#> 17.8 1.142281e+03 1.144640e+03 1.146999e+03 1.149359e+03 1.151718e+03
#> 17.9 1.178751e+03 1.181513e+03 1.184275e+03 1.187037e+03 1.189798e+03
#> 18   1.208983e+03 1.211577e+03 1.214172e+03 1.216766e+03 1.219361e+03
#> 18.1 1.235982e+03 1.239149e+03 1.242316e+03 1.245483e+03 1.248655e+03
#> 18.2 1.275093e+03 1.279287e+03 1.283481e+03 1.287675e+03 1.291787e+03
#> 18.3 1.310574e+03 1.314000e+03 1.317426e+03 1.320852e+03 1.324326e+03
#> 18.4 1.346214e+03 1.349394e+03 1.352575e+03 1.355755e+03 1.359049e+03
#> 18.5 1.381345e+03 1.385023e+03 1.388701e+03 1.392380e+03 1.395992e+03
#> 18.6 1.415076e+03 1.417326e+03 1.419575e+03 1.421825e+03 1.424906e+03
#> 18.7 1.450412e+03 1.453306e+03 1.456199e+03 1.459093e+03 1.462654e+03
#> 18.8 1.487060e+03 1.489750e+03 1.492441e+03 1.495132e+03 1.498809e+03
#> 18.9 1.526252e+03 1.530341e+03 1.534430e+03 1.538519e+03 1.542758e+03
#> 19   1.567813e+03 1.571470e+03 1.575126e+03 1.578783e+03 1.581955e+03
#> 19.1 1.602119e+03 1.606761e+03 1.611403e+03 1.616045e+03 1.621706e+03
#> 19.2 1.653124e+03 1.656153e+03 1.659182e+03 1.662210e+03 1.664069e+03
#> 19.3 1.680247e+03 1.685805e+03 1.691363e+03 1.696921e+03 1.700493e+03
#> 19.4 1.724544e+03 1.730473e+03 1.736402e+03 1.742332e+03 1.746547e+03
#> 19.5 1.765640e+03 1.767076e+03 1.768512e+03 1.769947e+03 1.776076e+03
#> 19.6 1.810670e+03 1.815314e+03 1.819958e+03 1.824519e+03 1.826221e+03
#> 19.7 1.846316e+03 1.852808e+03 1.859300e+03 1.865551e+03 1.869548e+03
#> 19.8 1.898299e+03 1.904533e+03 1.910767e+03 1.916304e+03 1.918312e+03
#> 19.9 1.938781e+03 1.944617e+03 1.950453e+03 1.956308e+03 1.962228e+03
#> 20   1.990161e+03 1.992736e+03 1.995311e+03 1.998146e+03 2.001583e+03
#> 20.1 2.025162e+03 2.029865e+03 2.034569e+03 2.039590e+03 2.045153e+03
#> 20.2 2.079576e+03 2.085574e+03 2.091572e+03 2.096133e+03 2.098853e+03
#> 20.3 2.124033e+03 2.130337e+03 2.136640e+03 2.141316e+03 2.144407e+03
#> 20.4 2.170201e+03 2.176145e+03 2.182089e+03 2.188153e+03 2.194305e+03
#> 20.5 2.225341e+03 2.229240e+03 2.233140e+03 2.239217e+03 2.246501e+03
#> 20.6 2.274939e+03 2.276521e+03 2.278102e+03 2.283398e+03 2.290200e+03
#> 20.7 2.321272e+03 2.324527e+03 2.327782e+03 2.330245e+03 2.332485e+03
#> 20.8 2.362025e+03 2.369986e+03 2.377948e+03 2.382348e+03 2.386112e+03
#> 20.9 2.421800e+03 2.430111e+03 2.438422e+03 2.443579e+03 2.448448e+03
#> 21   2.483009e+03 2.489690e+03 2.496370e+03 2.495013e+03 2.493533e+03
#> 21.1 2.520379e+03 2.530734e+03 2.540882e+03 2.547358e+03 2.553833e+03
#> 21.2 2.586916e+03 2.591522e+03 2.596101e+03 2.600484e+03 2.604867e+03
#> 21.3 2.637112e+03 2.643379e+03 2.648729e+03 2.650163e+03 2.651597e+03
#> 21.4 2.678849e+03 2.686067e+03 2.692953e+03 2.698879e+03 2.704805e+03
#> 21.5 2.734190e+03 2.738241e+03 2.743641e+03 2.751825e+03 2.760009e+03
#> 21.6 2.796346e+03 2.800730e+03 2.806285e+03 2.813632e+03 2.820980e+03
#> 21.7 2.859347e+03 2.865027e+03 2.869371e+03 2.872167e+03 2.874962e+03
#> 21.8 2.906916e+03 2.914052e+03 2.922670e+03 2.932593e+03 2.942517e+03
#> 21.9 2.979063e+03 2.982537e+03 2.986885e+03 2.991816e+03 2.996748e+03
#> 22   3.023899e+03 3.028160e+03 3.034497e+03 3.041866e+03 3.049235e+03
#> 22.1 3.095572e+03 3.103515e+03 3.108386e+03 3.112157e+03 3.115928e+03
#> 22.2 3.150060e+03 3.156882e+03 3.160675e+03 3.163733e+03 3.166790e+03
#> 22.3 3.203193e+03 3.210954e+03 3.216402e+03 3.221512e+03 3.226622e+03
#> 22.4 3.267187e+03 3.274832e+03 3.284901e+03 3.295122e+03 3.305342e+03
#> 22.5 3.348170e+03 3.353738e+03 3.359214e+03 3.364691e+03 3.370168e+03
#> 22.6 3.405344e+03 3.411280e+03 3.415866e+03 3.420452e+03 3.425038e+03
#> 22.7 3.469723e+03 3.478370e+03 3.486352e+03 3.494334e+03 3.502316e+03
#> 22.8 3.547555e+03 3.554719e+03 3.561207e+03 3.567695e+03 3.574184e+03
#> 22.9 3.601596e+03 3.606321e+03 3.613451e+03 3.620581e+03 3.627711e+03
#> 23   3.665926e+03 3.672345e+03 3.679404e+03 3.686463e+03 3.693521e+03
#> 23.1 3.728259e+03 3.732597e+03 3.735572e+03 3.738547e+03 3.741521e+03
#> 23.2 3.786865e+03 3.794006e+03 3.799038e+03 3.804070e+03 3.809102e+03
#> 23.3 3.871404e+03 3.883978e+03 3.896905e+03 3.909831e+03 3.922758e+03
#> 23.4 3.932931e+03 3.939580e+03 3.951285e+03 3.962990e+03 3.974695e+03
#> 23.5 4.002758e+03 4.009528e+03 4.018124e+03 4.026721e+03 4.035317e+03
#> 23.6 4.071480e+03 4.079993e+03 4.089505e+03 4.099016e+03 4.108528e+03
#> 23.7 4.166163e+03 4.171811e+03 4.176638e+03 4.181466e+03 4.186293e+03
#> 23.8 4.227768e+03 4.232119e+03 4.236122e+03 4.240125e+03 4.244128e+03
#> 23.9 4.319295e+03 4.323565e+03 4.327477e+03 4.331390e+03 4.335302e+03
#> 24             NA           NA           NA           NA           NA

# }